Unify Catboost model

Convert your GBM model into a standarised data frame. The returned data frame is easy to be interpreted by user and ready to be used as an argument in treeshap() function

catboost.unify(catboost_model, pool, data, recalculate = FALSE)

Arguments

catboost_model
An object of catboost.Model class. At the moment, models built on data with categorical features are not supported - please encode them before training.

pool
An object of catboost.Pool class used for training the model

data
data.frame for which calculations should be performed.

recalculate
logical indicating if covers should be recalculated according to the dataset given in data. Keep it FALSE if training data are used.

Value

a unified model representation - a model_unified.object object

See also

Examples

# library(catboost)
# data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
# label <- fifa20$target
# dt.pool <- catboost::catboost.load_pool(data = as.data.frame(lapply(data, as.numeric)),
#                                        label = label)
# cat_model <- catboost::catboost.train(
#             dt.pool,
#             params = list(loss_function = 'RMSE',
#                           iterations = 100,
#                           metric_period = 10,
#                           logging_level = 'Silent'))
# catboost.unify(cat_model, dt.pool)