Unify LightGBM model

Convert your xgboost model into a standarised data frame. The returned data frame is easy to be interpreted by user and ready to be used as an argument in the treeshap() function.

lightgbm.unify(lgb_model)

Arguments

lgb_model
A lightgbm model - object of class lgb.Booster

Value

Each row of a returned data frame indicates a specific node. The object has a defined structure:

Tree

0-indexed ID of a tree

Node

0-indexed ID of a node in a tree

Feature

In case of an internal node - name of a feature to split on. Otherwise - NA

Split

Threshold used for splitting observations. All observations with lower or equal value than it are proceeded to the node marked as 'Yes'. Otherwise to the 'No' node

Yes

Index of a row containing a child Node. Thanks to explicit indicating the row it is much faster to move between nodes.

No

Index of a row containing a child Node

Missing

Index of a row containing a child Node where are proceeded all observations with no value of the dividing feature

Quality/Score

For internal nodes - Quality: Split gain of a node. For leaves - Score: Value of prediction in the leaf

Cover

Number of observations seen by the internal node or collected by the leaf

See also

Examples

library(lightgbm)

#> Loading required package: R6

library(Matrix)
param_lgbm <- list(objective = "regression", max_depth = 2,  force_row_wise = TRUE)
data_fifa <- fifa20$data[!colnames(fifa20$data) %in%
             c('work_rate', 'value_eur', 'gk_diving', 'gk_handling',
             'gk_kicking', 'gk_reflexes', 'gk_speed', 'gk_positioning')]
data <- as.matrix(na.omit(data.table::as.data.table(cbind(data_fifa, fifa20$target))))
sparse_data <- as(data[,ncol(data)], 'sparseMatrix')
x <- lightgbm::lgb.Dataset(sparse_data, label = as(data[,ncol(data)], 'sparseMatrix'))
lgb_data <- lightgbm::lgb.Dataset.construct(x)
lgb_model <- lightgbm::lightgbm(data = lgb_data, params = param_lgbm)

#> [LightGBM] [Info] Total Bins 181
#> [LightGBM] [Info] Number of data points in the train set: 16242, number of used features: 1
#> [LightGBM] [Info] Start training from score 2578808.028568
#> [1]:	train's l2:2.65565e+13 
#> [2]:	train's l2:2.21523e+13 
#> [3]:	train's l2:1.85071e+13 
#> [4]:	train's l2:1.54623e+13 
#> [5]:	train's l2:1.29495e+13 
#> [6]:	train's l2:1.08784e+13 
#> [7]:	train's l2:9.1081e+12 
#> [8]:	train's l2:7.66033e+12 
#> [9]:	train's l2:6.43047e+12 
#> [10]:	train's l2:5.42396e+12 

lightgbm.unify(lgb_model)

#>     Tree Node  Feature    Split Yes No Missing Quality/Score Cover
#>  1:    0    0 Column_0 14750000   2  3       2  3.063193e+17 16242
#>  2:    0    1 Column_0  4450000   4  6       4  8.584868e+16 15775
#>  3:    0    2 Column_0 37000000   5  7       5  6.688754e+16   467
#>  4:    0    3     <NA>       NA  NA NA      NA  2.415325e+06 13780
#>  5:    0    4     <NA>       NA  NA NA      NA  4.522118e+06   378
#>  6:    0    5     <NA>       NA  NA NA      NA  3.117193e+06  1995
#>  7:    0    6     <NA>       NA  NA NA      NA  7.569242e+06    89
#>  8:    1    0 Column_0 16750000   9 10       9  2.500761e+17 16242
#>  9:    1    1 Column_0  5250000  11 13      11  7.883312e+16 15855
#> 10:    1    2 Column_0 43000000  12 14      12  4.758462e+16   387
#> 11:    1    3     <NA>       NA  NA NA      NA -1.400731e+05 14096
#> 12:    1    4     <NA>       NA  NA NA      NA  2.027243e+06   325
#> 13:    1    5     <NA>       NA  NA NA      NA  5.699240e+05  1759
#> 14:    1    6     <NA>       NA  NA NA      NA  5.050334e+06    62
#> 15:    2    0 Column_0 11750000  17 16      17  2.061295e+17 16242
#> 16:    2    1 Column_0 30750000  19 20      19  6.102996e+16   695
#> 17:    2    2 Column_0  2950000  18 21      18  4.444571e+16 15547
#> 18:    2    3     <NA>       NA  NA NA      NA -1.535281e+05 12807
#> 19:    2    4     <NA>       NA  NA NA      NA  1.224828e+06   560
#> 20:    2    5     <NA>       NA  NA NA      NA  3.593489e+06   135
#> 21:    2    6     <NA>       NA  NA NA      NA  2.902228e+05  2740
#> 22:    3    0 Column_0 19250000  23 24      23  1.695852e+17 16242
#> 23:    3    1 Column_0  6250000  25 27      25  6.101988e+16 15935
#> 24:    3    2 Column_0 52500000  26 28      26  2.967276e+16   307
#> 25:    3    3     <NA>       NA  NA NA      NA -1.086944e+05 14402
#> 26:    3    4     <NA>       NA  NA NA      NA  2.004448e+06   277
#> 27:    3    5     <NA>       NA  NA NA      NA  5.549405e+05  1533
#> 28:    3    6     <NA>       NA  NA NA      NA  5.315360e+06    30
#> 29:    4    0 Column_0 10250000  31 30      31  1.398929e+17 16242
#> 30:    4    1 Column_0 28250000  33 34      33  4.915308e+16   891
#> 31:    4    2 Column_0  2250000  32 35      32  2.576126e+16 15351
#> 32:    4    3     <NA>       NA  NA NA      NA -1.370525e+05 12161
#> 33:    4    4     <NA>       NA  NA NA      NA  8.706825e+05   731
#> 34:    4    5     <NA>       NA  NA NA      NA  2.805747e+06   160
#> 35:    4    6     <NA>       NA  NA NA      NA  1.822277e+05  3190
#> 36:    5    0 Column_0 13250000  38 37      38  1.148661e+17 16242
#> 37:    5    1 Column_0 47500000  40 41      40  3.186584e+16   549
#> 38:    5    2 Column_0  3450000  39 42      39  3.031321e+16 15693
#> 39:    5    3     <NA>       NA  NA NA      NA -1.100972e+05 13203
#> 40:    5    4     <NA>       NA  NA NA      NA  1.199721e+06   506
#> 41:    5    5     <NA>       NA  NA NA      NA  4.035285e+06    43
#> 42:    5    6     <NA>       NA  NA NA      NA  2.702962e+05  2490
#> 43:    6    0 Column_0 21250000  44 45      44  9.665059e+16 16242
#> 44:    6    1 Column_0  6750000  46 48      46  3.693527e+16 15978
#> 45:    6    2 Column_0 59250000  47 49      47  1.774785e+16   264
#> 46:    6    3     <NA>       NA  NA NA      NA -7.821417e+04 14592
#> 47:    6    4     <NA>       NA  NA NA      NA  1.656728e+06   243
#> 48:    6    5     <NA>       NA  NA NA      NA  4.619712e+05  1386
#> 49:    6    6     <NA>       NA  NA NA      NA  4.686864e+06    21
#> 50:    7    0 Column_0 22250000  51 52      51  7.862969e+16 16242
#> 51:    7    1 Column_0  7250000  53 55      53  3.160016e+16 16001
#> 52:    7    2 Column_0 59250000  54 56      54  1.353201e+16   241
#> 53:    7    3     <NA>       NA  NA NA      NA -6.824776e+04 14732
#> 54:    7    4     <NA>       NA  NA NA      NA  1.561317e+06   220
#> 55:    7    5     <NA>       NA  NA NA      NA  4.518160e+05  1269
#> 56:    7    6     <NA>       NA  NA NA      NA  4.218177e+06    21
#> 57:    8    0 Column_0  9750000  59 58      59  6.503226e+16 16242
#> 58:    8    1 Column_0 33500000  61 62      61  2.618155e+16   966
#> 59:    8    2 Column_0  1650000  60 63      60  1.391990e+16 15276
#> 60:    8    3     <NA>       NA  NA NA      NA -1.046617e+05 11537
#> 61:    8    4     <NA>       NA  NA NA      NA  6.110168e+05   858
#> 62:    8    5     <NA>       NA  NA NA      NA  2.263095e+06   108
#> 63:    8    6     <NA>       NA  NA NA      NA  1.173617e+05  3739
#> 64:    9    0 Column_0 18250000  65 66      65  5.392167e+16 16242
#> 65:    9    1 Column_0  2550000  67 69      67  2.005111e+16 15909
#> 66:    9    2 Column_0 56750000  68 70      68  1.206735e+16   333
#> 67:    9    3     <NA>       NA  NA NA      NA -8.527366e+04 12474
#> 68:    9    4     <NA>       NA  NA NA      NA  1.091624e+06   309
#> 69:    9    5     <NA>       NA  NA NA      NA  1.875766e+05  3435
#> 70:    9    6     <NA>       NA  NA NA      NA  3.419411e+06    24
#>     Tree Node  Feature    Split Yes No Missing Quality/Score Cover