Unify randomForest model

Convert your randomForest model into a standarised data frame. The returned data frame is easy to be interpreted by user and ready to be used as an argument in treeshap() function

randomForest.unify(rf_model, data)

Arguments

rf_model
An object of randomForest class. At the moment, models built on data with categorical features are not supported - please encode them before training.

data
A training frame used to fit the model.

Value

Each row of a returned data frame indicates a specific node. The object has a defined structure:

Tree

0-indexed ID of a tree

Node

0-indexed ID of a node in a tree

Feature

In case of an internal node - name of a feature to split on. Otherwise - NA

Split

Threshold used for splitting observations. All observations with lower or equal value than it are proceeded to the node marked as 'Yes'. Othwerwise to the 'No' node

Yes

Index of a row containing a child Node. Thanks to explicit indicating the row it is much faster to move between nodes.

No

Index of a row containing a child Node

Missing

Index of a row containing a child Node where are proceeded all observations with no value of the dividing feature

Quality/Score

For internal nodes - Quality: the reduction in the loss function as a result of splitting this node. For leaves - Score: Value of prediction in the leaf

Cover

Number of observations seen by the internal node or collected by the leaf

See also

Examples

# \dontrun{
library(randomForest)

#> randomForest 4.6-14

#> Type rfNews() to see new features/changes/bug fixes.

data_fifa <- fifa20$data[!colnames(fifa20$data) %in%
                           c('work_rate', 'value_eur', 'gk_diving', 'gk_handling',
                             'gk_kicking', 'gk_reflexes', 'gk_speed', 'gk_positioning')]
data <- as.matrix(na.omit(data.table::as.data.table(cbind(data_fifa, target = fifa20$target))))

rf <- randomForest::randomForest(target~., data = data, maxnodes = 10)
randomForest.unify(rf, data)

#>       Tree Node                 Feature Split Yes No Missing Quality/Score
#>    1:    0    0      movement_reactions  76.5   3  2       3     2612999.3
#>    2:    0    1               dribbling  73.5   5  4       5     1767369.4
#>    3:    0    2      movement_reactions  85.5   7  6       7    18002263.6
#>    4:    0    3 mentality_interceptions  70.5   9  8       9     1251190.4
#>    5:    0    4                 overall  74.5  11 10      11     6713110.0
#>   ---                                                                     
#> 9496:  499   14                    <NA>   0.0  NA NA      NA    54302816.9
#> 9497:  499   15                    <NA>   0.0  NA NA      NA      410307.7
#> 9498:  499   16                    <NA>   0.0  NA NA      NA     1186543.6
#> 9499:  499   17                    <NA>   0.0  NA NA      NA     5557388.9
#> 9500:  499   18                    <NA>   0.0  NA NA      NA    10846902.7
#>       Cover
#>    1: 16242
#>    2:   823
#>    3: 15419
#>    4:   559
#>    5:   264
#>   ---      
#> 9496: 15554
#> 9497:   452
#> 9498:     0
#> 9499:     0
#> 9500:     3

# }