Compute Cover values of your model for reference dataset other than the original set used to train model.

Compute Cover values of your model for reference dataset other than the original set used to train model.

recalculate_covers(model, X)

Arguments

model
Unified dataframe representation of the model created with a (model).unify function.

X
Reference dataset. A dataframe with the same columns as in the training set of the model.

Value

Unified dataframe representation of the model as created with a (model).unify function, but with Cover column containing updated values.

See also

Examples

# \dontrun{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
             formula = value_eur ~ .,
             data = data,
             distribution = "laplace",
             n.trees = 1000,
             cv.folds = 2,
             interaction.depth = 2,
             n.cores = 1)

#> CV: 1 
#> CV: 2 

unified <- gbm.unify(gbm_model)
recalculate_covers(unified, data[200:700, ])

#>       Tree Node   Feature Split Yes No Missing Quality/Score Cover
#>    1:    0    0   overall  65.5   2  3       7    5448.11127   501
#>    2:    0    1      <NA>    NA  NA NA      NA  -37500.00000     0
#>    3:    0    2 potential  68.5   4  5       6     993.25693   501
#>    4:    0    3      <NA>    NA  NA NA      NA   -7500.00000     0
#>    5:    0    4      <NA>    NA  NA NA      NA  170000.00000   501
#>   ---                                                             
#> 6996:  999    2      <NA>    NA  NA NA      NA     -50.60841     0
#> 6997:  999    3      <NA>    NA  NA NA      NA     350.72964     0
#> 6998:  999    4      <NA>    NA  NA NA      NA     232.18627     0
#> 6999:  999    5      <NA>    NA  NA NA      NA     -55.00373   501
#> 7000:  999    6      <NA>    NA  NA NA      NA     -21.06509     0

# }