Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
if (requireNamespace("gbm", quietly = TRUE)) {
  library(gbm)
  data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
  data['value_eur'] <- fifa20$target
  gbm_model <- gbm::gbm(
  formula = value_eur ~ .,
    data = data,
    distribution = "laplace",
    n.trees = 20,
    interaction.depth = 4,
    n.cores = 1)
  unified <- gbm.unify(gbm_model, data)
  set_reference_dataset(unified, data[200:700, ])
}# }
#>     Tree Node   Feature Decision.type Split Yes  No Missing Prediction Cover
#> 1      0    0   overall            <=  65.5   2   9      13         NA   501
#> 2      0    1 potential            <=  74.5   3   4       8         NA     0
#> 3      0    2      <NA>          <NA>    NA  NA  NA      NA  -6000.000     0
#> 4      0    3   overall            <=  62.5   5   6       7         NA     0
#> 5      0    4      <NA>          <NA>    NA  NA  NA      NA   7500.000     0
#> 6      0    5      <NA>          <NA>    NA  NA  NA      NA  45000.000     0
#> 7      0    6      <NA>          <NA>    NA  NA  NA      NA  25108.696     0
#> 8      0    7      <NA>          <NA>    NA  NA  NA      NA  -1726.828     0
#> 9      0    8 potential            <=  68.5  10  11      12         NA   501
#> 10     0    9      <NA>          <NA>    NA  NA  NA      NA  27500.000     0
#> 11     0   10      <NA>          <NA>    NA  NA  NA      NA 215000.000   501
#> 12     0   11      <NA>          <NA>    NA  NA  NA      NA 186343.125     0
#> 13     0   12      <NA>          <NA>    NA  NA  NA      NA 100200.131     0
#> 14     1    0   overall            <=  66.5  15  22      26         NA   501
#> 15     1    1 potential            <=  73.5  16  17      21         NA     0
#> 16     1    2      <NA>          <NA>    NA  NA  NA      NA   -900.000     0
#> 17     1    3   overall            <=  62.5  18  19      20         NA     0
#> 18     1    4      <NA>          <NA>    NA  NA  NA      NA  10250.000     0
#> 19     1    5      <NA>          <NA>    NA  NA  NA      NA  44100.000     0
#> 20     1    6      <NA>          <NA>    NA  NA  NA      NA  29494.281     0
#> 21     1    7      <NA>          <NA>    NA  NA  NA      NA   4746.571     0
#> 22     1    8 potential            <=  69.5  23  24      25         NA   501
#> 23     1    9      <NA>          <NA>    NA  NA  NA      NA  34500.000     0
#> 24     1   10      <NA>          <NA>    NA  NA  NA      NA 237000.000   501
#> 25     1   11      <NA>          <NA>    NA  NA  NA      NA 202926.813     0
#> 26     1   12      <NA>          <NA>    NA  NA  NA      NA  99857.473     0
#> 27     2    0   overall            <=  67.5  28  35      39         NA   501
#> 28     2    1   overall            <=  63.5  29  30      34         NA     0
#> 29     2    2      <NA>          <NA>    NA  NA  NA      NA  -3310.000     0
#> 30     2    3       age            <=  24.5  31  32      33         NA     0
#> 31     2    4      <NA>          <NA>    NA  NA  NA      NA  40190.000     0
#> 32     2    5      <NA>          <NA>    NA  NA  NA      NA  23090.000     0
#> 33     2    6      <NA>          <NA>    NA  NA  NA      NA  30363.588     0
#> 34     2    7      <NA>          <NA>    NA  NA  NA      NA  10613.533     0
#> 35     2    8 potential            <=  71.5  36  37      38         NA   501
#> 36     2    9      <NA>          <NA>    NA  NA  NA      NA  43300.000     0
#> 37     2   10      <NA>          <NA>    NA  NA  NA      NA 336800.000   501
#> 38     2   11      <NA>          <NA>    NA  NA  NA      NA 257705.008     0
#> 39     2   12      <NA>          <NA>    NA  NA  NA      NA 115355.011     0
#> 40     3    0   overall            <=  69.5  41  51      52         NA   501
#> 41     3    1   overall            <=  63.5  42  43      50         NA     0
#> 42     3    2      <NA>          <NA>    NA  NA  NA      NA     56.000     0
#> 43     3    3       age            <=  22.5  44  45      49         NA     0
#> 44     3    4      <NA>          <NA>    NA  NA  NA      NA  45571.000     0
#> 45     3    5 defending            <=  52.5  46  47      48         NA     0
#> 46     3    6      <NA>          <NA>    NA  NA  NA      NA  35531.000     0
#> 47     3    7      <NA>          <NA>    NA  NA  NA      NA  25531.000     0
#> 48     3    8      <NA>          <NA>    NA  NA  NA      NA  11381.000     0
#> 49     3    9      <NA>          <NA>    NA  NA  NA      NA  31438.177     0
#> 50     3   10      <NA>          <NA>    NA  NA  NA      NA  16182.883     0
#> 51     3   11      <NA>          <NA>    NA  NA  NA      NA 306620.000   501
#> 52     3   12      <NA>          <NA>    NA  NA  NA      NA 106374.429     0
#> 53     4    0   overall            <=  69.5  54  64      65         NA   501
#> 54     4    1   overall            <=  64.5  55  59      63         NA     0
#> 55     4    2 potential            <=  75.5  56  57      58         NA     0
#> 56     4    3      <NA>          <NA>    NA  NA  NA      NA   5015.400     0
#> 57     4    4      <NA>          <NA>    NA  NA  NA      NA  35415.400     0
#> 58     4    5      <NA>          <NA>    NA  NA  NA      NA   8116.417     0
#> 59     4    6 defending            <=  59.5  60  61      62         NA     0
#> 60     4    7      <NA>          <NA>    NA  NA  NA      NA  41113.900     0
#> 61     4    8      <NA>          <NA>    NA  NA  NA      NA  29827.900     0
#> 62     4    9      <NA>          <NA>    NA  NA  NA      NA  15112.400     0
#> 63     4   10      <NA>          <NA>    NA  NA  NA      NA  18915.272     0
#> 64     4   11      <NA>          <NA>    NA  NA  NA      NA 289458.000   501
#> 65     4   12      <NA>          <NA>    NA  NA  NA      NA 101241.492     0
#> 66     5    0   overall            <=  69.5  67  77      78         NA   501
#> 67     5    1   overall            <=  62.5  68  69      76         NA     0
#> 68     5    2      <NA>          <NA>    NA  NA  NA      NA   3548.860     0
#> 69     5    3 potential            <=  66.5  70  71      75         NA     0
#> 70     5    4      <NA>          <NA>    NA  NA  NA      NA  19866.510     0
#> 71     5    5 defending            <=  59.5  72  73      74         NA     0
#> 72     5    6      <NA>          <NA>    NA  NA  NA      NA  41472.360     0
#> 73     5    7      <NA>          <NA>    NA  NA  NA      NA  33413.860     0
#> 74     5    8      <NA>          <NA>    NA  NA  NA      NA  20091.660     0
#> 75     5    9      <NA>          <NA>    NA  NA  NA      NA  32134.894     0
#> 76     5   10      <NA>          <NA>    NA  NA  NA      NA  20368.158     0
#> 77     5   11      <NA>          <NA>    NA  NA  NA      NA 244012.200   501
#> 78     5   12      <NA>          <NA>    NA  NA  NA      NA  89083.824     0
#> 79     6    0 potential            <=  72.5  80  87      91         NA   501
#> 80     6    1   overall            <=  64.5  81  82      86         NA     0
#> 81     6    2      <NA>          <NA>    NA  NA  NA      NA   9158.974     0
#> 82     6    3       age            <=  30.5  83  84      85         NA     0
#> 83     6    4      <NA>          <NA>    NA  NA  NA      NA  36113.724     0
#> 84     6    5      <NA>          <NA>    NA  NA  NA      NA   9908.274     0
#> 85     6    6      <NA>          <NA>    NA  NA  NA      NA  28441.526     0
#> 86     6    7      <NA>          <NA>    NA  NA  NA      NA  18153.645     0
#> 87     6    8   overall            <=  61.5  88  89      90         NA   501
#> 88     6    9      <NA>          <NA>    NA  NA  NA      NA  18153.974     0
#> 89     6   10      <NA>          <NA>    NA  NA  NA      NA 143110.980   501
#> 90     6   11      <NA>          <NA>    NA  NA  NA      NA 130421.318     0
#> 91     6   12      <NA>          <NA>    NA  NA  NA      NA  64846.875     0
#> 92     7    0   overall            <=  70.5  93 100     104         NA   501
#> 93     7    1   overall            <=  63.5  94  95      99         NA     0
#> 94     7    2      <NA>          <NA>    NA  NA  NA      NA  12743.077     0
#> 95     7    3 potential            <=  77.5  96  97      98         NA     0
#> 96     7    4      <NA>          <NA>    NA  NA  NA      NA  28054.026     0
#> 97     7    5      <NA>          <NA>    NA  NA  NA      NA  55427.626     0
#> 98     7    6      <NA>          <NA>    NA  NA  NA      NA  30887.564     0
#> 99     7    7      <NA>          <NA>    NA  NA  NA      NA  22693.366     0
#> 100    7    8       age            <=  32.5 101 102     103         NA   501
#> 101    7    9      <NA>          <NA>    NA  NA  NA      NA 362299.882   435
#> 102    7   10      <NA>          <NA>    NA  NA  NA      NA  -5029.847    66
#> 103    7   11      <NA>          <NA>    NA  NA  NA      NA 319608.551     0
#> 104    7   12      <NA>          <NA>    NA  NA  NA      NA 100406.586     0
#> 105    8    0   overall            <=  70.5 106 113     117         NA   501
#> 106    8    1 dribbling            <=  63.5 107 111     112         NA     0
#> 107    8    2   overall            <=  61.5 108 109     110         NA     0
#> 108    8    3      <NA>          <NA>    NA  NA  NA      NA   8968.769     0
#> 109    8    4      <NA>          <NA>    NA  NA  NA      NA  28748.623     0
#> 110    8    5      <NA>          <NA>    NA  NA  NA      NA  19326.692     0
#> 111    8    6      <NA>          <NA>    NA  NA  NA      NA  33781.218     0
#> 112    8    7      <NA>          <NA>    NA  NA  NA      NA   9968.769     0
#> 113    8    8 potential            <=  72.5 114 115     116         NA   501
#> 114    8    9      <NA>          <NA>    NA  NA  NA      NA  30269.619     0
#> 115    8   10      <NA>          <NA>    NA  NA  NA      NA 379569.894   501
#> 116    8   11      <NA>          <NA>    NA  NA  NA      NA 324409.498     0
#> 117    8   12      <NA>          <NA>    NA  NA  NA      NA 101808.970     0
#> 118    9    0   overall            <=  71.5 119 129     130         NA   501
#> 119    9    1 dribbling            <=  63.5 120 124     128         NA     0
#> 120    9    2   overall            <=  60.5 121 122     123         NA     0
#> 121    9    3      <NA>          <NA>    NA  NA  NA      NA  10571.892     0
#> 122    9    4      <NA>          <NA>    NA  NA  NA      NA  29093.907     0
#> 123    9    5      <NA>          <NA>    NA  NA  NA      NA  21660.183     0
#> 124    9    6 potential            <=  77.5 125 126     127         NA     0
#> 125    9    7      <NA>          <NA>    NA  NA  NA      NA  33116.358     0
#> 126    9    8      <NA>          <NA>    NA  NA  NA      NA  59457.742     0
#> 127    9    9      <NA>          <NA>    NA  NA  NA      NA  36553.884     0
#> 128    9   10      <NA>          <NA>    NA  NA  NA      NA  11471.892     0
#> 129    9   11      <NA>          <NA>    NA  NA  NA      NA 345112.904   501
#> 130    9   12      <NA>          <NA>    NA  NA  NA      NA  97238.973     0
#> 131   10    0   overall            <=  71.5 132 142     143         NA   501
#> 132   10    1   overall            <=  62.5 133 134     141         NA     0
#> 133   10    2      <NA>          <NA>    NA  NA  NA      NA  16014.703     0
#> 134   10    3 potential            <=  77.5 135 139     140         NA     0
#> 135   10    4 defending            <=  58.5 136 137     138         NA     0
#> 136   10    5      <NA>          <NA>    NA  NA  NA      NA  34986.661     0
#> 137   10    6      <NA>          <NA>    NA  NA  NA      NA  27137.491     0
#> 138   10    7      <NA>          <NA>    NA  NA  NA      NA  20892.198     0
#> 139   10    8      <NA>          <NA>    NA  NA  NA      NA  53551.610     0
#> 140   10    9      <NA>          <NA>    NA  NA  NA      NA  32188.755     0
#> 141   10   10      <NA>          <NA>    NA  NA  NA      NA  26388.763     0
#> 142   10   11      <NA>          <NA>    NA  NA  NA      NA 364101.614   501
#> 143   10   12      <NA>          <NA>    NA  NA  NA      NA  99370.805     0
#> 144   11    0 potential            <=  75.5 145 152     156         NA   501
#> 145   11    1  shooting            <=  57.5 146 150     151         NA     0
#> 146   11    2   overall            <=  60.5 147 148     149         NA     0
#> 147   11    3      <NA>          <NA>    NA  NA  NA      NA  13898.733     0
#> 148   11    4      <NA>          <NA>    NA  NA  NA      NA  28839.002     0
#> 149   11    5      <NA>          <NA>    NA  NA  NA      NA  23791.042     0
#> 150   11    6      <NA>          <NA>    NA  NA  NA      NA  34987.995     0
#> 151   11    7      <NA>          <NA>    NA  NA  NA      NA  13708.733     0
#> 152   11    8   overall            <=  68.5 153 154     155         NA   501
#> 153   11    9      <NA>          <NA>    NA  NA  NA      NA  36508.733     0
#> 154   11   10      <NA>          <NA>    NA  NA  NA      NA 431191.453   501
#> 155   11   11      <NA>          <NA>    NA  NA  NA      NA 306774.913     0
#> 156   11   12      <NA>          <NA>    NA  NA  NA      NA  95736.213     0
#> 157   12    0   overall            <=  72.5 158 168     169         NA   501
#> 158   12    1   overall            <=  60.5 159 160     167         NA     0
#> 159   12    2      <NA>          <NA>    NA  NA  NA      NA  15023.359     0
#> 160   12    3 potential            <=  65.5 161 162     166         NA     0
#> 161   12    4      <NA>          <NA>    NA  NA  NA      NA  23670.185     0
#> 162   12    5 defending            <=  64.5 163 164     165         NA     0
#> 163   12    6      <NA>          <NA>    NA  NA  NA      NA  36293.641     0
#> 164   12    7      <NA>          <NA>    NA  NA  NA      NA  21807.552     0
#> 165   12    8      <NA>          <NA>    NA  NA  NA      NA  27874.075     0
#> 166   12    9      <NA>          <NA>    NA  NA  NA      NA  30765.735     0
#> 167   12   10      <NA>          <NA>    NA  NA  NA      NA  26843.744     0
#> 168   12   11      <NA>          <NA>    NA  NA  NA      NA 391572.307   501
#> 169   12   12      <NA>          <NA>    NA  NA  NA      NA  91376.635     0
#> 170   13    0   overall            <=  73.5 171 181     182         NA   501
#> 171   13    1   overall            <=  59.5 172 173     180         NA     0
#> 172   13    2      <NA>          <NA>    NA  NA  NA      NA  15850.023     0
#> 173   13    3 potential            <=  66.5 174 175     179         NA     0
#> 174   13    4      <NA>          <NA>    NA  NA  NA      NA  26192.763     0
#> 175   13    5 defending            <=  64.5 176 177     178         NA     0
#> 176   13    6      <NA>          <NA>    NA  NA  NA      NA  36454.882     0
#> 177   13    7      <NA>          <NA>    NA  NA  NA      NA  22387.718     0
#> 178   13    8      <NA>          <NA>    NA  NA  NA      NA  26207.488     0
#> 179   13    9      <NA>          <NA>    NA  NA  NA      NA  30668.419     0
#> 180   13   10      <NA>          <NA>    NA  NA  NA      NA  27877.100     0
#> 181   13   11      <NA>          <NA>    NA  NA  NA      NA 455915.077   501
#> 182   13   12      <NA>          <NA>    NA  NA  NA      NA  91902.476     0
#> 183   14    0   overall            <=  73.5 184 194     195         NA   501
#> 184   14    1   overall            <=  59.5 185 186     193         NA     0
#> 185   14    2      <NA>          <NA>    NA  NA  NA      NA  17827.095     0
#> 186   14    3       age            <=  30.5 187 191     192         NA     0
#> 187   14    4 potential            <=  65.5 188 189     190         NA     0
#> 188   14    5      <NA>          <NA>    NA  NA  NA      NA  27639.566     0
#> 189   14    6      <NA>          <NA>    NA  NA  NA      NA  35527.220     0
#> 190   14    7      <NA>          <NA>    NA  NA  NA      NA  34577.685     0
#> 191   14    8      <NA>          <NA>    NA  NA  NA      NA   7968.807     0
#> 192   14    9      <NA>          <NA>    NA  NA  NA      NA  30376.062     0
#> 193   14   10      <NA>          <NA>    NA  NA  NA      NA  27997.636     0
#> 194   14   11      <NA>          <NA>    NA  NA  NA      NA 363823.569   501
#> 195   14   12      <NA>          <NA>    NA  NA  NA      NA  76686.700     0
#> 196   15    0 potential            <=  76.5 197 207     208         NA   501
#> 197   15    1   overall            <=  59.5 198 199     206         NA     0
#> 198   15    2      <NA>          <NA>    NA  NA  NA      NA  19544.385     0
#> 199   15    3 potential            <=  65.5 200 201     205         NA     0
#> 200   15    4      <NA>          <NA>    NA  NA  NA      NA  26543.390     0
#> 201   15    5       age            <=  31.5 202 203     204         NA     0
#> 202   15    6      <NA>          <NA>    NA  NA  NA      NA  35226.701     0
#> 203   15    7      <NA>          <NA>    NA  NA  NA      NA -11646.912     0
#> 204   15    8      <NA>          <NA>    NA  NA  NA      NA  29357.957     0
#> 205   15    9      <NA>          <NA>    NA  NA  NA      NA  28909.090     0
#> 206   15   10      <NA>          <NA>    NA  NA  NA      NA  27097.150     0
#> 207   15   11      <NA>          <NA>    NA  NA  NA      NA 141244.733   501
#> 208   15   12      <NA>          <NA>    NA  NA  NA      NA  50741.025     0
#> 209   16    0   overall            <=  72.5 210 217     221         NA   501
#> 210   16    1   overall            <=  60.5 211 212     216         NA     0
#> 211   16    2      <NA>          <NA>    NA  NA  NA      NA  22027.873     0
#> 212   16    3 potential            <=  65.5 213 214     215         NA     0
#> 213   16    4      <NA>          <NA>    NA  NA  NA      NA  27653.784     0
#> 214   16    5      <NA>          <NA>    NA  NA  NA      NA  33020.009     0
#> 215   16    6      <NA>          <NA>    NA  NA  NA      NA  32230.274     0
#> 216   16    7      <NA>          <NA>    NA  NA  NA      NA  29724.492     0
#> 217   16    8       age            <=  30.5 218 219     220         NA   501
#> 218   16    9      <NA>          <NA>    NA  NA  NA      NA 320316.739   368
#> 219   16   10      <NA>          <NA>    NA  NA  NA      NA   2666.814   133
#> 220   16   11      <NA>          <NA>    NA  NA  NA      NA 234998.351     0
#> 221   16   12      <NA>          <NA>    NA  NA  NA      NA  66269.034     0
#> 222   17    0   overall            <=  73.5 223 233     234         NA   501
#> 223   17    1   overall            <=  59.5 224 225     232         NA     0
#> 224   17    2      <NA>          <NA>    NA  NA  NA      NA  21496.086     0
#> 225   17    3 potential            <=  64.5 226 227     231         NA     0
#> 226   17    4      <NA>          <NA>    NA  NA  NA      NA  26780.664     0
#> 227   17    5 defending            <=  57.5 228 229     230         NA     0
#> 228   17    6      <NA>          <NA>    NA  NA  NA      NA  36432.001     0
#> 229   17    7      <NA>          <NA>    NA  NA  NA      NA  30147.496     0
#> 230   17    8      <NA>          <NA>    NA  NA  NA      NA  27951.095     0
#> 231   17    9      <NA>          <NA>    NA  NA  NA      NA  32057.770     0
#> 232   17   10      <NA>          <NA>    NA  NA  NA      NA  30031.502     0
#> 233   17   11      <NA>          <NA>    NA  NA  NA      NA 341785.065   501
#> 234   17   12      <NA>          <NA>    NA  NA  NA      NA  74718.795     0
#> 235   18    0   overall            <=  73.5 236 246     247         NA   501
#> 236   18    1   overall            <=  58.5 237 238     245         NA     0
#> 237   18    2      <NA>          <NA>    NA  NA  NA      NA  21846.477     0
#> 238   18    3 potential            <=  65.5 239 240     244         NA     0
#> 239   18    4      <NA>          <NA>    NA  NA  NA      NA  28829.096     0
#> 240   18    5       age            <=  32.5 241 242     243         NA     0
#> 241   18    6      <NA>          <NA>    NA  NA  NA      NA  34214.777     0
#> 242   18    7      <NA>          <NA>    NA  NA  NA      NA  -8403.964     0
#> 243   18    8      <NA>          <NA>    NA  NA  NA      NA  31514.395     0
#> 244   18    9      <NA>          <NA>    NA  NA  NA      NA  31107.074     0
#> 245   18   10      <NA>          <NA>    NA  NA  NA      NA  29671.765     0
#> 246   18   11      <NA>          <NA>    NA  NA  NA      NA 311106.558   501
#> 247   18   12      <NA>          <NA>    NA  NA  NA      NA  71090.935     0
#> 248   19    0   overall            <=  74.5 249 259     260         NA   501
#> 249   19    1   overall            <=  58.5 250 251     258         NA     0
#> 250   19    2      <NA>          <NA>    NA  NA  NA      NA  23161.829     0
#> 251   19    3 potential            <=  64.5 252 253     257         NA     0
#> 252   19    4      <NA>          <NA>    NA  NA  NA      NA  27468.588     0
#> 253   19    5       age            <=  32.5 254 255     256         NA     0
#> 254   19    6      <NA>          <NA>    NA  NA  NA      NA  34367.178     0
#> 255   19    7      <NA>          <NA>    NA  NA  NA      NA   1464.875     0
#> 256   19    8      <NA>          <NA>    NA  NA  NA      NA  32072.042     0
#> 257   19    9      <NA>          <NA>    NA  NA  NA      NA  31587.186     0
#> 258   19   10      <NA>          <NA>    NA  NA  NA      NA  30347.425     0
#> 259   19   11      <NA>          <NA>    NA  NA  NA      NA 383495.903   501
#> 260   19   12      <NA>          <NA>    NA  NA  NA      NA  72042.053     0