treeshap results

treeshap object produced by treeshap function.

Value

List consisting of four elements:

shaps

A data.frame with M columns, X rows (M - number of features, X - number of explained observations). Every row corresponds to SHAP values for a observation.

interactions

An array with dimensions (M, M, X) (M - number of features, X - number of explained observations). Every [, , i] slice is a symmetric matrix - SHAP Interaction values for a observation. [a, b, i] element is SHAP Interaction value of features a and b for observation i. Is NULL if interactions where not calculated (parameter interactions set FALSE.)

unified_model

An object of type model_unified.object. Unified representation of a model for which SHAP values were calculated. It is used by some of the plotting functions.

observations

Explained dataset. data.frame or matrix. It is used by some of the plotting functions.

See also