Unify xgboost model

Convert your xgboost model into a standarised data frame. The returned data frame is easy to be interpreted by user and ready to be used as an argument in the treeshap() function.

xgboost.unify(xgb_model)

Arguments

xgb_model
A xgboost model - object of class xgb.Booster

Value

Each row of a returned data frame indicates a specific node. The object has a defined structure:

Tree

0-indexed ID of a tree

Node

0-indexed ID of a node in a tree

Feature

In case of an internal node - name of a feature to split on. Otherwise - NA.

Split

Threshold used for splitting observations. All observations with lower or equal value than it are proceeded to the node marked as 'Yes'. Otherwise to the 'No' node

Yes

Index of a row containing a child Node. Thanks to explicit indicating the row it is much faster to move between nodes.

No

Index of a row containing a child Node

Missing

Index of a row containing a child Node where are proceeded all observations with no value of the dividing feature

Quality/Score

For internal nodes - Quality: either the split gain (change in loss) or the leaf value. For leaves - Score: Value of prediction in the leaf

Cover

Number of observations seen by the internal node or collected by the leaf

See also

Examples

library(xgboost)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
target <- fifa20$target
param <- list(objective = "reg:squarederror", max_depth = 3)
xgb_model <- xgboost::xgboost(as.matrix(data), params = param, label = target, nrounds = 200)

#> [1]	train-rmse:4486581.000000 
#> [2]	train-rmse:3361763.250000 
#> [3]	train-rmse:2565844.250000 
#> [4]	train-rmse:2022569.875000 
#> [5]	train-rmse:1619285.875000 
#> [6]	train-rmse:1357081.875000 
#> [7]	train-rmse:1170674.625000 
#> [8]	train-rmse:1039663.937500 
#> [9]	train-rmse:932790.312500 
#> [10]	train-rmse:860090.625000 
#> [11]	train-rmse:804609.625000 
#> [12]	train-rmse:767626.812500 
#> [13]	train-rmse:740147.812500 
#> [14]	train-rmse:719762.812500 
#> [15]	train-rmse:706933.812500 
#> [16]	train-rmse:693178.625000 
#> [17]	train-rmse:679898.187500 
#> [18]	train-rmse:674524.875000 
#> [19]	train-rmse:665713.437500 
#> [20]	train-rmse:652927.562500 
#> [21]	train-rmse:639453.000000 
#> [22]	train-rmse:631931.875000 
#> [23]	train-rmse:628959.312500 
#> [24]	train-rmse:622040.187500 
#> [25]	train-rmse:615448.562500 
#> [26]	train-rmse:613797.937500 
#> [27]	train-rmse:605600.562500 
#> [28]	train-rmse:603055.250000 
#> [29]	train-rmse:596921.687500 
#> [30]	train-rmse:595214.062500 
#> [31]	train-rmse:594137.875000 
#> [32]	train-rmse:592722.750000 
#> [33]	train-rmse:587882.000000 
#> [34]	train-rmse:584771.500000 
#> [35]	train-rmse:578607.125000 
#> [36]	train-rmse:577946.750000 
#> [37]	train-rmse:575824.250000 
#> [38]	train-rmse:573312.750000 
#> [39]	train-rmse:568914.562500 
#> [40]	train-rmse:564939.500000 
#> [41]	train-rmse:559615.812500 
#> [42]	train-rmse:555904.937500 
#> [43]	train-rmse:553897.562500 
#> [44]	train-rmse:552377.812500 
#> [45]	train-rmse:551200.812500 
#> [46]	train-rmse:543481.187500 
#> [47]	train-rmse:542949.812500 
#> [48]	train-rmse:541332.062500 
#> [49]	train-rmse:538625.125000 
#> [50]	train-rmse:532953.812500 
#> [51]	train-rmse:531559.625000 
#> [52]	train-rmse:530476.812500 
#> [53]	train-rmse:528406.062500 
#> [54]	train-rmse:521505.218750 
#> [55]	train-rmse:520598.687500 
#> [56]	train-rmse:516478.875000 
#> [57]	train-rmse:515682.031250 
#> [58]	train-rmse:514609.000000 
#> [59]	train-rmse:512419.906250 
#> [60]	train-rmse:510127.781250 
#> [61]	train-rmse:504156.250000 
#> [62]	train-rmse:501481.812500 
#> [63]	train-rmse:499434.312500 
#> [64]	train-rmse:496394.031250 
#> [65]	train-rmse:495844.781250 
#> [66]	train-rmse:494176.375000 
#> [67]	train-rmse:492409.531250 
#> [68]	train-rmse:489478.843750 
#> [69]	train-rmse:487688.187500 
#> [70]	train-rmse:486379.781250 
#> [71]	train-rmse:485383.125000 
#> [72]	train-rmse:483979.718750 
#> [73]	train-rmse:481904.312500 
#> [74]	train-rmse:480251.093750 
#> [75]	train-rmse:478812.593750 
#> [76]	train-rmse:477639.562500 
#> [77]	train-rmse:476773.718750 
#> [78]	train-rmse:474864.750000 
#> [79]	train-rmse:473888.250000 
#> [80]	train-rmse:471276.656250 
#> [81]	train-rmse:469682.312500 
#> [82]	train-rmse:468950.281250 
#> [83]	train-rmse:468394.031250 
#> [84]	train-rmse:467058.343750 
#> [85]	train-rmse:465381.312500 
#> [86]	train-rmse:463539.406250 
#> [87]	train-rmse:462440.656250 
#> [88]	train-rmse:461192.937500 
#> [89]	train-rmse:460384.562500 
#> [90]	train-rmse:458866.000000 
#> [91]	train-rmse:456654.468750 
#> [92]	train-rmse:455613.968750 
#> [93]	train-rmse:455122.937500 
#> [94]	train-rmse:453131.906250 
#> [95]	train-rmse:451065.750000 
#> [96]	train-rmse:450597.968750 
#> [97]	train-rmse:449825.531250 
#> [98]	train-rmse:449018.906250 
#> [99]	train-rmse:447143.937500 
#> [100]	train-rmse:444112.531250 
#> [101]	train-rmse:441351.468750 
#> [102]	train-rmse:440108.000000 
#> [103]	train-rmse:438838.656250 
#> [104]	train-rmse:437694.312500 
#> [105]	train-rmse:437166.000000 
#> [106]	train-rmse:436571.062500 
#> [107]	train-rmse:436040.500000 
#> [108]	train-rmse:433524.031250 
#> [109]	train-rmse:432169.500000 
#> [110]	train-rmse:429161.281250 
#> [111]	train-rmse:427771.656250 
#> [112]	train-rmse:426731.281250 
#> [113]	train-rmse:425817.437500 
#> [114]	train-rmse:424805.718750 
#> [115]	train-rmse:423696.718750 
#> [116]	train-rmse:422583.187500 
#> [117]	train-rmse:421460.187500 
#> [118]	train-rmse:421063.437500 
#> [119]	train-rmse:417863.718750 
#> [120]	train-rmse:417616.375000 
#> [121]	train-rmse:417145.218750 
#> [122]	train-rmse:416822.812500 
#> [123]	train-rmse:416027.218750 
#> [124]	train-rmse:415133.468750 
#> [125]	train-rmse:414246.593750 
#> [126]	train-rmse:411989.687500 
#> [127]	train-rmse:410993.593750 
#> [128]	train-rmse:410076.531250 
#> [129]	train-rmse:408748.281250 
#> [130]	train-rmse:407586.687500 
#> [131]	train-rmse:406917.531250 
#> [132]	train-rmse:406422.218750 
#> [133]	train-rmse:405127.718750 
#> [134]	train-rmse:403180.062500 
#> [135]	train-rmse:402424.031250 
#> [136]	train-rmse:401981.343750 
#> [137]	train-rmse:400429.468750 
#> [138]	train-rmse:399471.656250 
#> [139]	train-rmse:398170.687500 
#> [140]	train-rmse:396799.343750 
#> [141]	train-rmse:394820.062500 
#> [142]	train-rmse:393622.031250 
#> [143]	train-rmse:393429.875000 
#> [144]	train-rmse:392676.875000 
#> [145]	train-rmse:392112.000000 
#> [146]	train-rmse:391111.500000 
#> [147]	train-rmse:389163.781250 
#> [148]	train-rmse:388552.281250 
#> [149]	train-rmse:387892.250000 
#> [150]	train-rmse:386419.750000 
#> [151]	train-rmse:385907.187500 
#> [152]	train-rmse:383801.937500 
#> [153]	train-rmse:383378.937500 
#> [154]	train-rmse:382820.937500 
#> [155]	train-rmse:382133.093750 
#> [156]	train-rmse:380109.406250 
#> [157]	train-rmse:378921.968750 
#> [158]	train-rmse:377941.093750 
#> [159]	train-rmse:377350.781250 
#> [160]	train-rmse:376439.875000 
#> [161]	train-rmse:375588.718750 
#> [162]	train-rmse:374107.531250 
#> [163]	train-rmse:373750.031250 
#> [164]	train-rmse:373187.437500 
#> [165]	train-rmse:372845.906250 
#> [166]	train-rmse:372117.187500 
#> [167]	train-rmse:371603.343750 
#> [168]	train-rmse:370779.500000 
#> [169]	train-rmse:369511.343750 
#> [170]	train-rmse:368900.656250 
#> [171]	train-rmse:368450.187500 
#> [172]	train-rmse:367442.625000 
#> [173]	train-rmse:366723.875000 
#> [174]	train-rmse:366248.031250 
#> [175]	train-rmse:365661.375000 
#> [176]	train-rmse:364705.468750 
#> [177]	train-rmse:364192.187500 
#> [178]	train-rmse:363884.250000 
#> [179]	train-rmse:363003.906250 
#> [180]	train-rmse:362021.406250 
#> [181]	train-rmse:361301.093750 
#> [182]	train-rmse:360885.218750 
#> [183]	train-rmse:360510.718750 
#> [184]	train-rmse:359024.343750 
#> [185]	train-rmse:358766.781250 
#> [186]	train-rmse:358507.031250 
#> [187]	train-rmse:357207.093750 
#> [188]	train-rmse:356523.718750 
#> [189]	train-rmse:355505.343750 
#> [190]	train-rmse:353378.562500 
#> [191]	train-rmse:352559.281250 
#> [192]	train-rmse:351736.781250 
#> [193]	train-rmse:349713.500000 
#> [194]	train-rmse:349121.843750 
#> [195]	train-rmse:348947.812500 
#> [196]	train-rmse:348441.281250 
#> [197]	train-rmse:347524.187500 
#> [198]	train-rmse:346942.875000 
#> [199]	train-rmse:346422.718750 
#> [200]	train-rmse:346166.062500 

xgboost.unify(xgb_model)

#>       Tree Node   Feature Split Yes No Missing Quality/Score Cover
#>    1:    0    0   overall  81.5   2  3       2  3.060549e+17 18278
#>    2:    0    1   overall  73.5   4  5       4  1.130790e+17 17949
#>    3:    0    2   overall  84.5   6  7       6  4.021330e+16   329
#>    4:    0    3   overall  69.5   8  9       8  1.165151e+16 15628
#>    5:    0    4 potential  79.5  10 11      10  1.813107e+16  2321
#>   ---                                                             
#> 2866:  199   10      <NA>    NA  NA NA      NA -6.182059e+03   310
#> 2867:  199   11      <NA>    NA  NA NA      NA  1.757646e+03 12943
#> 2868:  199   12      <NA>    NA  NA NA      NA -2.336901e+04   149
#> 2869:  199   13      <NA>    NA  NA NA      NA -5.065065e+03  1334
#> 2870:  199   14      <NA>    NA  NA NA      NA -9.290863e+04    22