Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
if (requireNamespace("gbm", quietly = TRUE)) {
  library(gbm)
  data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
  data['value_eur'] <- fifa20$target
  gbm_model <- gbm::gbm(
  formula = value_eur ~ .,
    data = data,
    distribution = "laplace",
    n.trees = 20,
    interaction.depth = 4,
    n.cores = 1)
  unified <- gbm.unify(gbm_model, data)
  set_reference_dataset(unified, data[200:700, ])
}# }
#>     Tree Node     Feature Decision.type Split Yes  No Missing Prediction Cover
#> 1      0    0     overall            <=  65.5   2   9      13         NA   501
#> 2      0    1   potential            <=  74.5   3   4       8         NA     0
#> 3      0    2        <NA>          <NA>    NA  NA  NA      NA  -6000.000     0
#> 4      0    3     overall            <=  62.5   5   6       7         NA     0
#> 5      0    4        <NA>          <NA>    NA  NA  NA      NA  10000.000     0
#> 6      0    5        <NA>          <NA>    NA  NA  NA      NA  45000.000     0
#> 7      0    6        <NA>          <NA>    NA  NA  NA      NA  26537.801     0
#> 8      0    7        <NA>          <NA>    NA  NA  NA      NA  -1481.508     0
#> 9      0    8   potential            <=  68.5  10  11      12         NA   501
#> 10     0    9        <NA>          <NA>    NA  NA  NA      NA  27500.000     0
#> 11     0   10        <NA>          <NA>    NA  NA  NA      NA 205000.000   501
#> 12     0   11        <NA>          <NA>    NA  NA  NA      NA 178705.032     0
#> 13     0   12        <NA>          <NA>    NA  NA  NA      NA  96074.352     0
#> 14     1    0     overall            <=  66.5  15  22      26         NA   501
#> 15     1    1     overall            <=  63.5  16  17      21         NA     0
#> 16     1    2        <NA>          <NA>    NA  NA  NA      NA  -6900.000     0
#> 17     1    3   potential            <=  71.5  18  19      20         NA     0
#> 18     1    4        <NA>          <NA>    NA  NA  NA      NA  21600.000     0
#> 19     1    5        <NA>          <NA>    NA  NA  NA      NA  43000.000     0
#> 20     1    6        <NA>          <NA>    NA  NA  NA      NA  28421.171     0
#> 21     1    7        <NA>          <NA>    NA  NA  NA      NA   5706.184     0
#> 22     1    8   potential            <=  69.5  23  24      25         NA   501
#> 23     1    9        <NA>          <NA>    NA  NA  NA      NA  35500.000     0
#> 24     1   10        <NA>          <NA>    NA  NA  NA      NA 248000.000   501
#> 25     1   11        <NA>          <NA>    NA  NA  NA      NA 212172.915     0
#> 26     1   12        <NA>          <NA>    NA  NA  NA      NA 105132.826     0
#> 27     2    0     overall            <=  67.5  28  35      39         NA   501
#> 28     2    1   potential            <=  73.5  29  30      34         NA     0
#> 29     2    2        <NA>          <NA>    NA  NA  NA      NA   5790.000     0
#> 30     2    3     overall            <=  62.5  31  32      33         NA     0
#> 31     2    4        <NA>          <NA>    NA  NA  NA      NA  14190.000     0
#> 32     2    5        <NA>          <NA>    NA  NA  NA      NA  45690.000     0
#> 33     2    6        <NA>          <NA>    NA  NA  NA      NA  33588.827     0
#> 34     2    7        <NA>          <NA>    NA  NA  NA      NA  11171.945     0
#> 35     2    8         age            <=  32.5  36  37      38         NA   501
#> 36     2    9        <NA>          <NA>    NA  NA  NA      NA 226700.000   435
#> 37     2   10        <NA>          <NA>    NA  NA  NA      NA  24200.000    66
#> 38     2   11        <NA>          <NA>    NA  NA  NA      NA 203902.335     0
#> 39     2   12        <NA>          <NA>    NA  NA  NA      NA  92469.205     0
#> 40     3    0     overall            <=  69.5  41  51      52         NA   501
#> 41     3    1   potential            <=  75.5  42  46      50         NA     0
#> 42     3    2     overall            <=  64.5  43  44      45         NA     0
#> 43     3    3        <NA>          <NA>    NA  NA  NA      NA   2211.000     0
#> 44     3    4        <NA>          <NA>    NA  NA  NA      NA  27511.000     0
#> 45     3    5        <NA>          <NA>    NA  NA  NA      NA  12402.943     0
#> 46     3    6     overall            <=  61.5  47  48      49         NA     0
#> 47     3    7        <NA>          <NA>    NA  NA  NA      NA  16271.000     0
#> 48     3    8        <NA>          <NA>    NA  NA  NA      NA  47530.000     0
#> 49     3    9        <NA>          <NA>    NA  NA  NA      NA  41497.561     0
#> 50     3   10        <NA>          <NA>    NA  NA  NA      NA  16097.645     0
#> 51     3   11        <NA>          <NA>    NA  NA  NA      NA 327530.000   501
#> 52     3   12        <NA>          <NA>    NA  NA  NA      NA 113388.308     0
#> 53     4    0     overall            <=  69.5  54  64      65         NA   501
#> 54     4    1   potential            <=  74.5  55  59      63         NA     0
#> 55     4    2     overall            <=  63.5  56  57      58         NA     0
#> 56     4    3        <NA>          <NA>    NA  NA  NA      NA   1489.900     0
#> 57     4    4        <NA>          <NA>    NA  NA  NA      NA  27028.900     0
#> 58     4    5        <NA>          <NA>    NA  NA  NA      NA  14089.915     0
#> 59     4    6     overall            <=  61.5  60  61      62         NA     0
#> 60     4    7        <NA>          <NA>    NA  NA  NA      NA  13143.900     0
#> 61     4    8        <NA>          <NA>    NA  NA  NA      NA  44878.000     0
#> 62     4    9        <NA>          <NA>    NA  NA  NA      NA  37620.294     0
#> 63     4   10        <NA>          <NA>    NA  NA  NA      NA  18089.413     0
#> 64     4   11        <NA>          <NA>    NA  NA  NA      NA 298277.000   501
#> 65     4   12        <NA>          <NA>    NA  NA  NA      NA 103473.200     0
#> 66     5    0   potential            <=  71.5  67  71      78         NA   501
#> 67     5    1    shooting            <=  59.5  68  69      70         NA     0
#> 68     5    2        <NA>          <NA>    NA  NA  NA      NA  10840.910     0
#> 69     5    3        <NA>          <NA>    NA  NA  NA      NA  30326.010     0
#> 70     5    4        <NA>          <NA>    NA  NA  NA      NA  -1159.090     0
#> 71     5    5     overall            <=  62.5  72  73      77         NA   501
#> 72     5    6        <NA>          <NA>    NA  NA  NA      NA  20329.510     0
#> 73     5    7 gk_reflexes            <=  73.5  74  75      76         NA   501
#> 74     5    8        <NA>          <NA>    NA  NA  NA      NA  24890.200     0
#> 75     5    9        <NA>          <NA>    NA  NA  NA      NA 321949.300    56
#> 76     5   10        <NA>          <NA>    NA  NA  NA      NA 141949.300   445
#> 77     5   11        <NA>          <NA>    NA  NA  NA      NA 125974.004     0
#> 78     5   12        <NA>          <NA>    NA  NA  NA      NA  67623.065     0
#> 79     6    0     overall            <=  70.5  80  87      91         NA   501
#> 80     6    1   dribbling            <=  63.5  81  82      86         NA     0
#> 81     6    2        <NA>          <NA>    NA  NA  NA      NA  15256.819     0
#> 82     6    3   potential            <=  77.5  83  84      85         NA     0
#> 83     6    4        <NA>          <NA>    NA  NA  NA      NA  30374.409     0
#> 84     6    5        <NA>          <NA>    NA  NA  NA      NA  62695.270     0
#> 85     6    6        <NA>          <NA>    NA  NA  NA      NA  34190.252     0
#> 86     6    7        <NA>          <NA>    NA  NA  NA      NA   7456.819     0
#> 87     6    8         age            <=  33.5  88  89      90         NA   501
#> 88     6    9        <NA>          <NA>    NA  NA  NA      NA 391254.370   462
#> 89     6   10        <NA>          <NA>    NA  NA  NA      NA   1504.370    39
#> 90     6   11        <NA>          <NA>    NA  NA  NA      NA 361082.316     0
#> 91     6   12        <NA>          <NA>    NA  NA  NA      NA 108528.233     0
#> 92     7    0     overall            <=  70.5  93 103     104         NA   501
#> 93     7    1     overall            <=  63.5  94  95     102         NA     0
#> 94     7    2        <NA>          <NA>    NA  NA  NA      NA  11231.137     0
#> 95     7    3   potential            <=  76.5  96 100     101         NA     0
#> 96     7    4   defending            <=  57.5  97  98      99         NA     0
#> 97     7    5        <NA>          <NA>    NA  NA  NA      NA  34327.237     0
#> 98     7    6        <NA>          <NA>    NA  NA  NA      NA  24966.237     0
#> 99     7    7        <NA>          <NA>    NA  NA  NA      NA  16657.398     0
#> 100    7    8        <NA>          <NA>    NA  NA  NA      NA  51824.743     0
#> 101    7    9        <NA>          <NA>    NA  NA  NA      NA  30818.181     0
#> 102    7   10        <NA>          <NA>    NA  NA  NA      NA  21997.045     0
#> 103    7   11        <NA>          <NA>    NA  NA  NA      NA 305628.933   501
#> 104    7   12        <NA>          <NA>    NA  NA  NA      NA  96202.521     0
#> 105    8    0     overall            <=  71.5 106 116     117         NA   501
#> 106    8    1     overall            <=  61.5 107 108     115         NA     0
#> 107    8    2        <NA>          <NA>    NA  NA  NA      NA   7608.023     0
#> 108    8    3         age            <=  27.5 109 113     114         NA     0
#> 109    8    4   potential            <=  77.5 110 111     112         NA     0
#> 110    8    5        <NA>          <NA>    NA  NA  NA      NA  33374.355     0
#> 111    8    6        <NA>          <NA>    NA  NA  NA      NA  54743.269     0
#> 112    8    7        <NA>          <NA>    NA  NA  NA      NA  36555.978     0
#> 113    8    8        <NA>          <NA>    NA  NA  NA      NA  19728.369     0
#> 114    8    9        <NA>          <NA>    NA  NA  NA      NA  30765.321     0
#> 115    8   10        <NA>          <NA>    NA  NA  NA      NA  23723.964     0
#> 116    8   11        <NA>          <NA>    NA  NA  NA      NA 378566.040   501
#> 117    8   12        <NA>          <NA>    NA  NA  NA      NA 102620.900     0
#> 118    9    0   potential            <=  73.5 119 126     130         NA   501
#> 119    9    1     overall            <=  62.5 120 121     125         NA     0
#> 120    9    2        <NA>          <NA>    NA  NA  NA      NA  11327.221     0
#> 121    9    3         age            <=  29.5 122 123     124         NA     0
#> 122    9    4        <NA>          <NA>    NA  NA  NA      NA  33730.580     0
#> 123    9    5        <NA>          <NA>    NA  NA  NA      NA  12362.167     0
#> 124    9    6        <NA>          <NA>    NA  NA  NA      NA  26451.167     0
#> 125    9    7        <NA>          <NA>    NA  NA  NA      NA  20799.614     0
#> 126    9    8     overall            <=  69.5 127 128     129         NA   501
#> 127    9    9        <NA>          <NA>    NA  NA  NA      NA  33130.281     0
#> 128    9   10        <NA>          <NA>    NA  NA  NA      NA 344209.436   501
#> 129    9   11        <NA>          <NA>    NA  NA  NA      NA 218708.051     0
#> 130    9   12        <NA>          <NA>    NA  NA  NA      NA  91331.158     0
#> 131   10    0     overall            <=  71.5 132 142     143         NA   501
#> 132   10    1     overall            <=  60.5 133 134     141         NA     0
#> 133   10    2        <NA>          <NA>    NA  NA  NA      NA  10714.499     0
#> 134   10    3         age            <=  27.5 135 139     140         NA     0
#> 135   10    4    shooting            <=  52.5 136 137     138         NA     0
#> 136   10    5        <NA>          <NA>    NA  NA  NA      NA  32106.542     0
#> 137   10    6        <NA>          <NA>    NA  NA  NA      NA  37949.020     0
#> 138   10    7        <NA>          <NA>    NA  NA  NA      NA  26617.866     0
#> 139   10    8        <NA>          <NA>    NA  NA  NA      NA  21984.719     0
#> 140   10    9        <NA>          <NA>    NA  NA  NA      NA  30409.296     0
#> 141   10   10        <NA>          <NA>    NA  NA  NA      NA  25467.753     0
#> 142   10   11        <NA>          <NA>    NA  NA  NA      NA 363288.492   501
#> 143   10   12        <NA>          <NA>    NA  NA  NA      NA  98103.463     0
#> 144   11    0     overall            <=  72.5 145 155     156         NA   501
#> 145   11    1     overall            <=  60.5 146 147     154         NA     0
#> 146   11    2        <NA>          <NA>    NA  NA  NA      NA  13143.049     0
#> 147   11    3         age            <=  27.5 148 152     153         NA     0
#> 148   11    4   potential            <=  77.5 149 150     151         NA     0
#> 149   11    5        <NA>          <NA>    NA  NA  NA      NA  33415.453     0
#> 150   11    6        <NA>          <NA>    NA  NA  NA      NA  57404.857     0
#> 151   11    7        <NA>          <NA>    NA  NA  NA      NA  37173.136     0
#> 152   11    8        <NA>          <NA>    NA  NA  NA      NA  21744.059     0
#> 153   11    9        <NA>          <NA>    NA  NA  NA      NA  31948.074     0
#> 154   11   10        <NA>          <NA>    NA  NA  NA      NA  27502.930     0
#> 155   11   11        <NA>          <NA>    NA  NA  NA      NA 380459.643   501
#> 156   11   12        <NA>          <NA>    NA  NA  NA      NA  91922.647     0
#> 157   12    0   potential            <=  76.5 158 165     169         NA   501
#> 158   12    1    shooting            <=  58.5 159 163     164         NA     0
#> 159   12    2     overall            <=  59.5 160 161     162         NA     0
#> 160   12    3        <NA>          <NA>    NA  NA  NA      NA  13328.744     0
#> 161   12    4        <NA>          <NA>    NA  NA  NA      NA  29376.195     0
#> 162   12    5        <NA>          <NA>    NA  NA  NA      NA  24883.061     0
#> 163   12    6        <NA>          <NA>    NA  NA  NA      NA  35535.814     0
#> 164   12    7        <NA>          <NA>    NA  NA  NA      NA  15308.744     0
#> 165   12    8     overall            <=  63.5 166 167     168         NA   501
#> 166   12    9        <NA>          <NA>    NA  NA  NA      NA  31444.333     0
#> 167   12   10        <NA>          <NA>    NA  NA  NA      NA 345913.679   501
#> 168   12   11        <NA>          <NA>    NA  NA  NA      NA 309423.530     0
#> 169   12   12        <NA>          <NA>    NA  NA  NA      NA  83414.352     0
#> 170   13    0     overall            <=  72.5 171 181     182         NA   501
#> 171   13    1     overall            <=  59.5 172 173     180         NA     0
#> 172   13    2        <NA>          <NA>    NA  NA  NA      NA  15495.870     0
#> 173   13    3         age            <=  30.5 174 178     179         NA     0
#> 174   13    4   defending            <=  54.5 175 176     177         NA     0
#> 175   13    5        <NA>          <NA>    NA  NA  NA      NA  35886.513     0
#> 176   13    6        <NA>          <NA>    NA  NA  NA      NA  30153.546     0
#> 177   13    7        <NA>          <NA>    NA  NA  NA      NA  27277.870     0
#> 178   13    8        <NA>          <NA>    NA  NA  NA      NA  11217.242     0
#> 179   13    9        <NA>          <NA>    NA  NA  NA      NA  28769.039     0
#> 180   13   10        <NA>          <NA>    NA  NA  NA      NA  26157.457     0
#> 181   13   11        <NA>          <NA>    NA  NA  NA      NA 364822.311   501
#> 182   13   12        <NA>          <NA>    NA  NA  NA      NA  88339.274     0
#> 183   14    0     overall            <=  73.5 184 194     195         NA   501
#> 184   14    1     overall            <=  60.5 185 186     193         NA     0
#> 185   14    2        <NA>          <NA>    NA  NA  NA      NA  19290.714     0
#> 186   14    3         age            <=  30.5 187 191     192         NA     0
#> 187   14    4   defending            <=  53.5 188 189     190         NA     0
#> 188   14    5        <NA>          <NA>    NA  NA  NA      NA  37187.169     0
#> 189   14    6        <NA>          <NA>    NA  NA  NA      NA  31209.325     0
#> 190   14    7        <NA>          <NA>    NA  NA  NA      NA  26978.201     0
#> 191   14    8        <NA>          <NA>    NA  NA  NA      NA  12578.338     0
#> 192   14    9        <NA>          <NA>    NA  NA  NA      NA  29702.550     0
#> 193   14   10        <NA>          <NA>    NA  NA  NA      NA  27258.113     0
#> 194   14   11        <NA>          <NA>    NA  NA  NA      NA 431840.080   501
#> 195   14   12        <NA>          <NA>    NA  NA  NA      NA  85827.097     0
#> 196   15    0     overall            <=  73.5 197 207     208         NA   501
#> 197   15    1     overall            <=  59.5 198 199     206         NA     0
#> 198   15    2        <NA>          <NA>    NA  NA  NA      NA  19017.211     0
#> 199   15    3   potential            <=  65.5 200 201     205         NA     0
#> 200   15    4        <NA>          <NA>    NA  NA  NA      NA  25841.474     0
#> 201   15    5   defending            <=  63.5 202 203     204         NA     0
#> 202   15    6        <NA>          <NA>    NA  NA  NA      NA  35726.354     0
#> 203   15    7        <NA>          <NA>    NA  NA  NA      NA  23613.175     0
#> 204   15    8        <NA>          <NA>    NA  NA  NA      NA  26492.362     0
#> 205   15    9        <NA>          <NA>    NA  NA  NA      NA  30556.940     0
#> 206   15   10        <NA>          <NA>    NA  NA  NA      NA  28415.364     0
#> 207   15   11        <NA>          <NA>    NA  NA  NA      NA 392156.072   501
#> 208   15   12        <NA>          <NA>    NA  NA  NA      NA  80753.588     0
#> 209   16    0     overall            <=  73.5 210 220     221         NA   501
#> 210   16    1     overall            <=  59.5 211 212     219         NA     0
#> 211   16    2        <NA>          <NA>    NA  NA  NA      NA  20615.490     0
#> 212   16    3   potential            <=  65.5 213 214     218         NA     0
#> 213   16    4        <NA>          <NA>    NA  NA  NA      NA  26862.551     0
#> 214   16    5   defending            <=  65.5 215 216     217         NA     0
#> 215   16    6        <NA>          <NA>    NA  NA  NA      NA  35410.106     0
#> 216   16    7        <NA>          <NA>    NA  NA  NA      NA  23182.874     0
#> 217   16    8        <NA>          <NA>    NA  NA  NA      NA  27691.954     0
#> 218   16    9        <NA>          <NA>    NA  NA  NA      NA  31386.456     0
#> 219   16   10        <NA>          <NA>    NA  NA  NA      NA  29321.506     0
#> 220   16   11        <NA>          <NA>    NA  NA  NA      NA 356440.465   501
#> 221   16   12        <NA>          <NA>    NA  NA  NA      NA  75638.601     0
#> 222   17    0     overall            <=  74.5 223 233     234         NA   501
#> 223   17    1     overall            <=  59.5 224 225     232         NA     0
#> 224   17    2        <NA>          <NA>    NA  NA  NA      NA  22053.941     0
#> 225   17    3         age            <=  30.5 226 230     231         NA     0
#> 226   17    4   potential            <=  64.5 227 228     229         NA     0
#> 227   17    5        <NA>          <NA>    NA  NA  NA      NA  27200.222     0
#> 228   17    6        <NA>          <NA>    NA  NA  NA      NA  34704.236     0
#> 229   17    7        <NA>          <NA>    NA  NA  NA      NA  34162.287     0
#> 230   17    8        <NA>          <NA>    NA  NA  NA      NA  12481.900     0
#> 231   17    9        <NA>          <NA>    NA  NA  NA      NA  30599.426     0
#> 232   17   10        <NA>          <NA>    NA  NA  NA      NA  29010.897     0
#> 233   17   11        <NA>          <NA>    NA  NA  NA      NA 424296.418   501
#> 234   17   12        <NA>          <NA>    NA  NA  NA      NA  74815.401     0
#> 235   18    0     overall            <=  73.5 236 246     247         NA   501
#> 236   18    1     overall            <=  58.5 237 238     245         NA     0
#> 237   18    2        <NA>          <NA>    NA  NA  NA      NA  22130.547     0
#> 238   18    3   potential            <=  64.5 239 240     244         NA     0
#> 239   18    4        <NA>          <NA>    NA  NA  NA      NA  27097.968     0
#> 240   18    5   defending            <=  65.5 241 242     243         NA     0
#> 241   18    6        <NA>          <NA>    NA  NA  NA      NA  35050.000     0
#> 242   18    7        <NA>          <NA>    NA  NA  NA      NA  23079.412     0
#> 243   18    8        <NA>          <NA>    NA  NA  NA      NA  28452.335     0
#> 244   18    9        <NA>          <NA>    NA  NA  NA      NA  31508.002     0
#> 245   18   10        <NA>          <NA>    NA  NA  NA      NA  30046.236     0
#> 246   18   11        <NA>          <NA>    NA  NA  NA      NA 285366.776   501
#> 247   18   12        <NA>          <NA>    NA  NA  NA      NA  67789.758     0
#> 248   19    0     overall            <=  74.5 249 259     260         NA   501
#> 249   19    1     overall            <=  58.5 250 251     258         NA     0
#> 250   19    2        <NA>          <NA>    NA  NA  NA      NA  23635.492     0
#> 251   19    3   potential            <=  64.5 252 253     257         NA     0
#> 252   19    4        <NA>          <NA>    NA  NA  NA      NA  27470.765     0
#> 253   19    5         age            <=  31.5 254 255     256         NA     0
#> 254   19    6        <NA>          <NA>    NA  NA  NA      NA  34458.705     0
#> 255   19    7        <NA>          <NA>    NA  NA  NA      NA   6064.628     0
#> 256   19    8        <NA>          <NA>    NA  NA  NA      NA  31578.879     0
#> 257   19    9        <NA>          <NA>    NA  NA  NA      NA  31163.707     0
#> 258   19   10        <NA>          <NA>    NA  NA  NA      NA  30053.185     0
#> 259   19   11        <NA>          <NA>    NA  NA  NA      NA 310330.099   501
#> 260   19   12        <NA>          <NA>    NA  NA  NA      NA  62929.523     0