Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node     Feature Decision.type Split Yes  No Missing Prediction Cover
#> 1      0    0     overall            <=  65.5   2   9      13         NA   501
#> 2      0    1   potential            <=  73.5   3   4       8         NA     0
#> 3      0    2        <NA>          <NA>    NA  NA  NA      NA  -7000.000     0
#> 4      0    3     overall            <=  62.5   5   6       7         NA     0
#> 5      0    4        <NA>          <NA>    NA  NA  NA      NA   7500.000     0
#> 6      0    5        <NA>          <NA>    NA  NA  NA      NA  45000.000     0
#> 7      0    6        <NA>          <NA>    NA  NA  NA      NA  25352.257     0
#> 8      0    7        <NA>          <NA>    NA  NA  NA      NA  -1369.167     0
#> 9      0    8   potential            <=  68.5  10  11      12         NA   501
#> 10     0    9        <NA>          <NA>    NA  NA  NA      NA  30000.000     0
#> 11     0   10        <NA>          <NA>    NA  NA  NA      NA 205000.000   501
#> 12     0   11        <NA>          <NA>    NA  NA  NA      NA 178071.472     0
#> 13     0   12        <NA>          <NA>    NA  NA  NA      NA  95606.139     0
#> 14     1    0     overall            <=  66.5  15  22      26         NA   501
#> 15     1    1   potential            <=  71.5  16  17      21         NA     0
#> 16     1    2        <NA>          <NA>    NA  NA  NA      NA  -2800.000     0
#> 17     1    3     overall            <=  63.5  18  19      20         NA     0
#> 18     1    4        <NA>          <NA>    NA  NA  NA      NA  12750.000     0
#> 19     1    5        <NA>          <NA>    NA  NA  NA      NA  43000.000     0
#> 20     1    6        <NA>          <NA>    NA  NA  NA      NA  25191.532     0
#> 21     1    7        <NA>          <NA>    NA  NA  NA      NA   5265.156     0
#> 22     1    8   potential            <=  69.5  23  24      25         NA   501
#> 23     1    9        <NA>          <NA>    NA  NA  NA      NA  35500.000     0
#> 24     1   10        <NA>          <NA>    NA  NA  NA      NA 248000.000   501
#> 25     1   11        <NA>          <NA>    NA  NA  NA      NA 211964.245     0
#> 26     1   12        <NA>          <NA>    NA  NA  NA      NA 104894.162     0
#> 27     2    0     overall            <=  67.5  28  35      39         NA   501
#> 28     2    1     overall            <=  63.5  29  30      34         NA     0
#> 29     2    2        <NA>          <NA>    NA  NA  NA      NA  -4020.000     0
#> 30     2    3         age            <=  23.5  31  32      33         NA     0
#> 31     2    4        <NA>          <NA>    NA  NA  NA      NA  40700.000     0
#> 32     2    5        <NA>          <NA>    NA  NA  NA      NA  24280.000     0
#> 33     2    6        <NA>          <NA>    NA  NA  NA      NA  30067.199     0
#> 34     2    7        <NA>          <NA>    NA  NA  NA      NA  10491.865     0
#> 35     2    8         age            <=  32.5  36  37      38         NA   501
#> 36     2    9        <NA>          <NA>    NA  NA  NA      NA 246700.000   435
#> 37     2   10        <NA>          <NA>    NA  NA  NA      NA  26700.000    66
#> 38     2   11        <NA>          <NA>    NA  NA  NA      NA 221031.411     0
#> 39     2   12        <NA>          <NA>    NA  NA  NA      NA  98356.820     0
#> 40     3    0     overall            <=  69.5  41  51      52         NA   501
#> 41     3    1   potential            <=  74.5  42  46      50         NA     0
#> 42     3    2     overall            <=  64.5  43  44      45         NA     0
#> 43     3    3        <NA>          <NA>    NA  NA  NA      NA   1882.000     0
#> 44     3    4        <NA>          <NA>    NA  NA  NA      NA  25530.000     0
#> 45     3    5        <NA>          <NA>    NA  NA  NA      NA  11383.998     0
#> 46     3    6     overall            <=  61.5  47  48      49         NA     0
#> 47     3    7        <NA>          <NA>    NA  NA  NA      NA  11377.000     0
#> 48     3    8        <NA>          <NA>    NA  NA  NA      NA  46130.000     0
#> 49     3    9        <NA>          <NA>    NA  NA  NA      NA  38772.383     0
#> 50     3   10        <NA>          <NA>    NA  NA  NA      NA  16206.508     0
#> 51     3   11        <NA>          <NA>    NA  NA  NA      NA 315530.000   501
#> 52     3   12        <NA>          <NA>    NA  NA  NA      NA 109059.348     0
#> 53     4    0     overall            <=  69.5  54  64      65         NA   501
#> 54     4    1   potential            <=  75.5  55  62      63         NA     0
#> 55     4    2     overall            <=  63.5  56  57      61         NA     0
#> 56     4    3        <NA>          <NA>    NA  NA  NA      NA   2193.800     0
#> 57     4    4   defending            <=  55.5  58  59      60         NA     0
#> 58     4    5        <NA>          <NA>    NA  NA  NA      NA  36299.000     0
#> 59     4    6        <NA>          <NA>    NA  NA  NA      NA  23799.000     0
#> 60     4    7        <NA>          <NA>    NA  NA  NA      NA   7499.000     0
#> 61     4    8        <NA>          <NA>    NA  NA  NA      NA  14512.045     0
#> 62     4    9        <NA>          <NA>    NA  NA  NA      NA  44014.000     0
#> 63     4   10        <NA>          <NA>    NA  NA  NA      NA  18219.634     0
#> 64     4   11        <NA>          <NA>    NA  NA  NA      NA 287477.000   501
#> 65     4   12        <NA>          <NA>    NA  NA  NA      NA 101333.217     0
#> 66     5    0     overall            <=  70.5  67  77      78         NA   501
#> 67     5    1     overall            <=  62.5  68  69      76         NA     0
#> 68     5    2        <NA>          <NA>    NA  NA  NA      NA   3474.420     0
#> 69     5    3         age            <=  24.5  70  71      75         NA     0
#> 70     5    4        <NA>          <NA>    NA  NA  NA      NA  39615.600     0
#> 71     5    5         age            <=  28.5  72  73      74         NA     0
#> 72     5    6        <NA>          <NA>    NA  NA  NA      NA  28847.100     0
#> 73     5    7        <NA>          <NA>    NA  NA  NA      NA  13847.100     0
#> 74     5    8        <NA>          <NA>    NA  NA  NA      NA  21789.331     0
#> 75     5    9        <NA>          <NA>    NA  NA  NA      NA  28913.870     0
#> 76     5   10        <NA>          <NA>    NA  NA  NA      NA  19203.840     0
#> 77     5   11        <NA>          <NA>    NA  NA  NA      NA 352229.300   501
#> 78     5   12        <NA>          <NA>    NA  NA  NA      NA 105821.798     0
#> 79     6    0   potential            <=  72.5  80  84      91         NA   501
#> 80     6    1    shooting            <=  59.5  81  82      83         NA     0
#> 81     6    2        <NA>          <NA>    NA  NA  NA      NA  15626.978     0
#> 82     6    3        <NA>          <NA>    NA  NA  NA      NA  31984.390     0
#> 83     6    4        <NA>          <NA>    NA  NA  NA      NA   2213.390     0
#> 84     6    5     overall            <=  61.5  85  86      90         NA   501
#> 85     6    6        <NA>          <NA>    NA  NA  NA      NA  17672.478     0
#> 86     6    7 gk_reflexes            <=  73.5  87  88      89         NA   501
#> 87     6    8        <NA>          <NA>    NA  NA  NA      NA  25730.340     0
#> 88     6    9        <NA>          <NA>    NA  NA  NA      NA 320506.370    56
#> 89     6   10        <NA>          <NA>    NA  NA  NA      NA 160506.370   445
#> 90     6   11        <NA>          <NA>    NA  NA  NA      NA 146635.805     0
#> 91     6   12        <NA>          <NA>    NA  NA  NA      NA  72042.303     0
#> 92     7    0     overall            <=  70.5  93 103     104         NA   501
#> 93     7    1     overall            <=  63.5  94  95     102         NA     0
#> 94     7    2        <NA>          <NA>    NA  NA  NA      NA  11928.539     0
#> 95     7    3   potential            <=  77.5  96 100     101         NA     0
#> 96     7    4   defending            <=  58.5  97  98      99         NA     0
#> 97     7    5        <NA>          <NA>    NA  NA  NA      NA  33334.903     0
#> 98     7    6        <NA>          <NA>    NA  NA  NA      NA  23971.692     0
#> 99     7    7        <NA>          <NA>    NA  NA  NA      NA  17943.051     0
#> 100    7    8        <NA>          <NA>    NA  NA  NA      NA  57603.403     0
#> 101    7    9        <NA>          <NA>    NA  NA  NA      NA  30107.581     0
#> 102    7   10        <NA>          <NA>    NA  NA  NA      NA  21751.391     0
#> 103    7   11        <NA>          <NA>    NA  NA  NA      NA 307955.733   501
#> 104    7   12        <NA>          <NA>    NA  NA  NA      NA  97569.392     0
#> 105    8    0     overall            <=  71.5 106 116     117         NA   501
#> 106    8    1     overall            <=  62.5 107 108     115         NA     0
#> 107    8    2        <NA>          <NA>    NA  NA  NA      NA  10871.426     0
#> 108    8    3   potential            <=  77.5 109 113     114         NA     0
#> 109    8    4         age            <=  29.5 110 111     112         NA     0
#> 110    8    5        <NA>          <NA>    NA  NA  NA      NA  31606.234     0
#> 111    8    6        <NA>          <NA>    NA  NA  NA      NA  13632.409     0
#> 112    8    7        <NA>          <NA>    NA  NA  NA      NA  26908.629     0
#> 113    8    8        <NA>          <NA>    NA  NA  NA      NA  58243.063     0
#> 114    8    9        <NA>          <NA>    NA  NA  NA      NA  30042.762     0
#> 115    8   10        <NA>          <NA>    NA  NA  NA      NA  23094.328     0
#> 116    8   11        <NA>          <NA>    NA  NA  NA      NA 380660.160   501
#> 117    8   12        <NA>          <NA>    NA  NA  NA      NA 101501.367     0
#> 118    9    0     overall            <=  71.5 119 129     130         NA   501
#> 119    9    1   dribbling            <=  63.5 120 124     128         NA     0
#> 120    9    2     overall            <=  60.5 121 122     123         NA     0
#> 121    9    3        <NA>          <NA>    NA  NA  NA      NA   9284.284     0
#> 122    9    4        <NA>          <NA>    NA  NA  NA      NA  28127.110     0
#> 123    9    5        <NA>          <NA>    NA  NA  NA      NA  20428.939     0
#> 124    9    6         age            <=  29.5 125 126     127         NA     0
#> 125    9    7        <NA>          <NA>    NA  NA  NA      NA  35519.837     0
#> 126    9    8        <NA>          <NA>    NA  NA  NA      NA  16949.420     0
#> 127    9    9        <NA>          <NA>    NA  NA  NA      NA  32482.586     0
#> 128    9   10        <NA>          <NA>    NA  NA  NA      NA  11416.123     0
#> 129    9   11        <NA>          <NA>    NA  NA  NA      NA 346094.144   501
#> 130    9   12        <NA>          <NA>    NA  NA  NA      NA  94876.746     0
#> 131   10    0     overall            <=  72.5 132 142     143         NA   501
#> 132   10    1     overall            <=  62.5 133 134     141         NA     0
#> 133   10    2        <NA>          <NA>    NA  NA  NA      NA  15855.855     0
#> 134   10    3         age            <=  29.5 135 139     140         NA     0
#> 135   10    4   potential            <=  77.5 136 137     138         NA     0
#> 136   10    5        <NA>          <NA>    NA  NA  NA      NA  33064.715     0
#> 137   10    6        <NA>          <NA>    NA  NA  NA      NA  62966.773     0
#> 138   10    7        <NA>          <NA>    NA  NA  NA      NA  37248.314     0
#> 139   10    8        <NA>          <NA>    NA  NA  NA      NA  14068.571     0
#> 140   10    9        <NA>          <NA>    NA  NA  NA      NA  31809.990     0
#> 141   10   10        <NA>          <NA>    NA  NA  NA      NA  26382.558     0
#> 142   10   11        <NA>          <NA>    NA  NA  NA      NA 414984.729   501
#> 143   10   12        <NA>          <NA>    NA  NA  NA      NA  96627.748     0
#> 144   11    0     overall            <=  69.5 145 152     156         NA   501
#> 145   11    1     overall            <=  60.5 146 147     151         NA     0
#> 146   11    2        <NA>          <NA>    NA  NA  NA      NA  13611.220     0
#> 147   11    3   potential            <=  65.5 148 149     150         NA     0
#> 148   11    4        <NA>          <NA>    NA  NA  NA      NA  24679.338     0
#> 149   11    5        <NA>          <NA>    NA  NA  NA      NA  33022.334     0
#> 150   11    6        <NA>          <NA>    NA  NA  NA      NA  31479.284     0
#> 151   11    7        <NA>          <NA>    NA  NA  NA      NA  26219.126     0
#> 152   11    8   potential            <=  73.5 153 154     155         NA   501
#> 153   11    9        <NA>          <NA>    NA  NA  NA      NA   8604.917     0
#> 154   11   10        <NA>          <NA>    NA  NA  NA      NA 312188.052   501
#> 155   11   11        <NA>          <NA>    NA  NA  NA      NA 218917.250     0
#> 156   11   12        <NA>          <NA>    NA  NA  NA      NA  86269.926     0
#> 157   12    0     overall            <=  72.5 158 168     169         NA   501
#> 158   12    1    shooting            <=  55.5 159 160     167         NA     0
#> 159   12    2        <NA>          <NA>    NA  NA  NA      NA  24306.422     0
#> 160   12    3         age            <=  30.5 161 165     166         NA     0
#> 161   12    4   potential            <=  65.5 162 163     164         NA     0
#> 162   12    5        <NA>          <NA>    NA  NA  NA      NA  27950.677     0
#> 163   12    6        <NA>          <NA>    NA  NA  NA      NA  38109.149     0
#> 164   12    7        <NA>          <NA>    NA  NA  NA      NA  36960.534     0
#> 165   12    8        <NA>          <NA>    NA  NA  NA      NA  12341.549     0
#> 166   12    9        <NA>          <NA>    NA  NA  NA      NA  33233.106     0
#> 167   12   10        <NA>          <NA>    NA  NA  NA      NA  16037.323     0
#> 168   12   11        <NA>          <NA>    NA  NA  NA      NA 349267.451   501
#> 169   12   12        <NA>          <NA>    NA  NA  NA      NA  84709.005     0
#> 170   13    0   potential            <=  75.5 171 178     182         NA   501
#> 171   13    1   dribbling            <=  63.5 172 173     177         NA     0
#> 172   13    2        <NA>          <NA>    NA  NA  NA      NA  23978.506     0
#> 173   13    3         age            <=  30.5 174 175     176         NA     0
#> 174   13    4        <NA>          <NA>    NA  NA  NA      NA  36201.931     0
#> 175   13    5        <NA>          <NA>    NA  NA  NA      NA  10539.582     0
#> 176   13    6        <NA>          <NA>    NA  NA  NA      NA  31348.209     0
#> 177   13    7        <NA>          <NA>    NA  NA  NA      NA  16259.070     0
#> 178   13    8     overall            <=  68.5 179 180     181         NA   501
#> 179   13    9        <NA>          <NA>    NA  NA  NA      NA  34056.947     0
#> 180   13   10        <NA>          <NA>    NA  NA  NA      NA 367840.706   501
#> 181   13   11        <NA>          <NA>    NA  NA  NA      NA 269547.652     0
#> 182   13   12        <NA>          <NA>    NA  NA  NA      NA  86265.659     0
#> 183   14    0     overall            <=  73.5 184 194     195         NA   501
#> 184   14    1     overall            <=  59.5 185 186     193         NA     0
#> 185   14    2        <NA>          <NA>    NA  NA  NA      NA  16080.655     0
#> 186   14    3         age            <=  28.5 187 191     192         NA     0
#> 187   14    4   potential            <=  65.5 188 189     190         NA     0
#> 188   14    5        <NA>          <NA>    NA  NA  NA      NA  29046.499     0
#> 189   14    6        <NA>          <NA>    NA  NA  NA      NA  35755.261     0
#> 190   14    7        <NA>          <NA>    NA  NA  NA      NA  35168.189     0
#> 191   14    8        <NA>          <NA>    NA  NA  NA      NA  20958.966     0
#> 192   14    9        <NA>          <NA>    NA  NA  NA      NA  31298.087     0
#> 193   14   10        <NA>          <NA>    NA  NA  NA      NA  28371.207     0
#> 194   14   11        <NA>          <NA>    NA  NA  NA      NA 434556.635   501
#> 195   14   12        <NA>          <NA>    NA  NA  NA      NA  88861.235     0
#> 196   15    0     overall            <=  73.5 197 207     208         NA   501
#> 197   15    1     overall            <=  59.5 198 199     206         NA     0
#> 198   15    2        <NA>          <NA>    NA  NA  NA      NA  17972.590     0
#> 199   15    3         age            <=  30.5 200 204     205         NA     0
#> 200   15    4   potential            <=  64.5 201 202     203         NA     0
#> 201   15    5        <NA>          <NA>    NA  NA  NA      NA  27646.079     0
#> 202   15    6        <NA>          <NA>    NA  NA  NA      NA  35132.640     0
#> 203   15    7        <NA>          <NA>    NA  NA  NA      NA  34570.408     0
#> 204   15    8        <NA>          <NA>    NA  NA  NA      NA  13359.026     0
#> 205   15    9        <NA>          <NA>    NA  NA  NA      NA  31140.041     0
#> 206   15   10        <NA>          <NA>    NA  NA  NA      NA  28633.717     0
#> 207   15   11        <NA>          <NA>    NA  NA  NA      NA 394600.972   501
#> 208   15   12        <NA>          <NA>    NA  NA  NA      NA  81132.138     0
#> 209   16    0   potential            <=  78.5 210 220     221         NA   501
#> 210   16    1     overall            <=  60.5 211 212     219         NA     0
#> 211   16    2        <NA>          <NA>    NA  NA  NA      NA  21310.905     0
#> 212   16    3         age            <=  31.5 213 217     218         NA     0
#> 213   16    4    shooting            <=  60.5 214 215     216         NA     0
#> 214   16    5        <NA>          <NA>    NA  NA  NA      NA  32661.916     0
#> 215   16    6        <NA>          <NA>    NA  NA  NA      NA  39692.838     0
#> 216   16    7        <NA>          <NA>    NA  NA  NA      NA  29120.526     0
#> 217   16    8        <NA>          <NA>    NA  NA  NA      NA   7599.189     0
#> 218   16    9        <NA>          <NA>    NA  NA  NA      NA  30653.363     0
#> 219   16   10        <NA>          <NA>    NA  NA  NA      NA  28531.365     0
#> 220   16   11        <NA>          <NA>    NA  NA  NA      NA 308640.875   501
#> 221   16   12        <NA>          <NA>    NA  NA  NA      NA  66016.202     0
#> 222   17    0     overall            <=  73.5 223 233     234         NA   501
#> 223   17    1     overall            <=  59.5 224 225     232         NA     0
#> 224   17    2        <NA>          <NA>    NA  NA  NA      NA  21044.240     0
#> 225   17    3   potential            <=  64.5 226 227     231         NA     0
#> 226   17    4        <NA>          <NA>    NA  NA  NA      NA  27654.111     0
#> 227   17    5   defending            <=  63.5 228 229     230         NA     0
#> 228   17    6        <NA>          <NA>    NA  NA  NA      NA  35860.774     0
#> 229   17    7        <NA>          <NA>    NA  NA  NA      NA  24397.400     0
#> 230   17    8        <NA>          <NA>    NA  NA  NA      NA  28087.803     0
#> 231   17    9        <NA>          <NA>    NA  NA  NA      NA  31421.707     0
#> 232   17   10        <NA>          <NA>    NA  NA  NA      NA  29446.122     0
#> 233   17   11        <NA>          <NA>    NA  NA  NA      NA 331276.787   501
#> 234   17   12        <NA>          <NA>    NA  NA  NA      NA  73305.529     0
#> 235   18    0     overall            <=  59.5 236 237     247         NA   501
#> 236   18    1        <NA>          <NA>    NA  NA  NA      NA  22439.816     0
#> 237   18    2     overall            <=  74.5 238 245     246         NA   501
#> 238   18    3         age            <=  31.5 239 243     244         NA     0
#> 239   18    4   defending            <=  54.5 240 241     242         NA     0
#> 240   18    5        <NA>          <NA>    NA  NA  NA      NA  36855.889     0
#> 241   18    6        <NA>          <NA>    NA  NA  NA      NA  31388.700     0
#> 242   18    7        <NA>          <NA>    NA  NA  NA      NA  30039.613     0
#> 243   18    8        <NA>          <NA>    NA  NA  NA      NA  14802.958     0
#> 244   18    9        <NA>          <NA>    NA  NA  NA      NA  31132.524     0
#> 245   18   10        <NA>          <NA>    NA  NA  NA      NA 401649.109   501
#> 246   18   11        <NA>          <NA>    NA  NA  NA      NA  82174.562     0
#> 247   18   12        <NA>          <NA>    NA  NA  NA      NA  72259.078     0
#> 248   19    0     overall            <=  74.5 249 259     260         NA   501
#> 249   19    1     overall            <=  58.5 250 251     258         NA     0
#> 250   19    2        <NA>          <NA>    NA  NA  NA      NA  22695.834     0
#> 251   19    3   potential            <=  64.5 252 253     257         NA     0
#> 252   19    4        <NA>          <NA>    NA  NA  NA      NA  28756.638     0
#> 253   19    5   defending            <=  64.5 254 255     256         NA     0
#> 254   19    6        <NA>          <NA>    NA  NA  NA      NA  35920.501     0
#> 255   19    7        <NA>          <NA>    NA  NA  NA      NA  23923.885     0
#> 256   19    8        <NA>          <NA>    NA  NA  NA      NA  28544.037     0
#> 257   19    9        <NA>          <NA>    NA  NA  NA      NA  31882.472     0
#> 258   19   10        <NA>          <NA>    NA  NA  NA      NA  30497.918     0
#> 259   19   11        <NA>          <NA>    NA  NA  NA      NA 364984.198   501
#> 260   19   12        <NA>          <NA>    NA  NA  NA      NA  70172.185     0
# }