Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node            Feature Decision.type Split Yes  No Missing Prediction
#> 1      0    0            overall            <=  65.5   2   9      13         NA
#> 2      0    1          potential            <=  73.5   3   4       8         NA
#> 3      0    2               <NA>          <NA>    NA  NA  NA      NA  -7000.000
#> 4      0    3            overall            <=  63.5   5   6       7         NA
#> 5      0    4               <NA>          <NA>    NA  NA  NA      NA  10000.000
#> 6      0    5               <NA>          <NA>    NA  NA  NA      NA  50000.000
#> 7      0    6               <NA>          <NA>    NA  NA  NA      NA  23145.058
#> 8      0    7               <NA>          <NA>    NA  NA  NA      NA  -1386.087
#> 9      0    8          potential            <=  68.5  10  11      12         NA
#> 10     0    9               <NA>          <NA>    NA  NA  NA      NA  27500.000
#> 11     0   10               <NA>          <NA>    NA  NA  NA      NA 205000.000
#> 12     0   11               <NA>          <NA>    NA  NA  NA      NA 178425.141
#> 13     0   12               <NA>          <NA>    NA  NA  NA      NA  96123.974
#> 14     1    0            overall            <=  66.5  15  22      26         NA
#> 15     1    1            overall            <=  63.5  16  17      21         NA
#> 16     1    2               <NA>          <NA>    NA  NA  NA      NA  -6800.000
#> 17     1    3          potential            <=  71.5  18  19      20         NA
#> 18     1    4               <NA>          <NA>    NA  NA  NA      NA  21700.000
#> 19     1    5               <NA>          <NA>    NA  NA  NA      NA  43500.000
#> 20     1    6               <NA>          <NA>    NA  NA  NA      NA  28903.697
#> 21     1    7               <NA>          <NA>    NA  NA  NA      NA   6097.141
#> 22     1    8          potential            <=  69.5  23  24      25         NA
#> 23     1    9               <NA>          <NA>    NA  NA  NA      NA  33250.000
#> 24     1   10               <NA>          <NA>    NA  NA  NA      NA 258000.000
#> 25     1   11               <NA>          <NA>    NA  NA  NA      NA 219947.090
#> 26     1   12               <NA>          <NA>    NA  NA  NA      NA 107815.680
#> 27     2    0            overall            <=  67.5  28  38      39         NA
#> 28     2    1          potential            <=  74.5  29  33      37         NA
#> 29     2    2            overall            <=  64.5  30  31      32         NA
#> 30     2    3               <NA>          <NA>    NA  NA  NA      NA  -1620.000
#> 31     2    4               <NA>          <NA>    NA  NA  NA      NA  25925.000
#> 32     2    5               <NA>          <NA>    NA  NA  NA      NA   6536.230
#> 33     2    6            overall            <=  62.5  34  35      36         NA
#> 34     2    7               <NA>          <NA>    NA  NA  NA      NA  16680.000
#> 35     2    8               <NA>          <NA>    NA  NA  NA      NA  47150.000
#> 36     2    9               <NA>          <NA>    NA  NA  NA      NA  35873.113
#> 37     2   10               <NA>          <NA>    NA  NA  NA      NA  11066.680
#> 38     2   11               <NA>          <NA>    NA  NA  NA      NA 205700.000
#> 39     2   12               <NA>          <NA>    NA  NA  NA      NA  93166.630
#> 40     3    0            overall            <=  69.5  41  51      52         NA
#> 41     3    1          potential            <=  75.5  42  46      50         NA
#> 42     3    2 skill_ball_control            <=  63.5  43  44      45         NA
#> 43     3    3               <NA>          <NA>    NA  NA  NA      NA   6042.000
#> 44     3    4               <NA>          <NA>    NA  NA  NA      NA  28630.000
#> 45     3    5               <NA>          <NA>    NA  NA  NA      NA  13114.386
#> 46     3    6            overall            <=  61.5  47  48      49         NA
#> 47     3    7               <NA>          <NA>    NA  NA  NA      NA  13512.000
#> 48     3    8               <NA>          <NA>    NA  NA  NA      NA  48630.000
#> 49     3    9               <NA>          <NA>    NA  NA  NA      NA  41193.753
#> 50     3   10               <NA>          <NA>    NA  NA  NA      NA  16753.479
#> 51     3   11               <NA>          <NA>    NA  NA  NA      NA 328630.000
#> 52     3   12               <NA>          <NA>    NA  NA  NA      NA 113500.381
#> 53     4    0            overall            <=  69.5  54  64      65         NA
#> 54     4    1          potential            <=  73.5  55  59      63         NA
#> 55     4    2           shooting            <=  59.5  56  57      58         NA
#> 56     4    3               <NA>          <NA>    NA  NA  NA      NA  11525.800
#> 57     4    4               <NA>          <NA>    NA  NA  NA      NA  29937.800
#> 58     4    5               <NA>          <NA>    NA  NA  NA      NA  -1062.200
#> 59     4    6            overall            <=  62.5  60  61      62         NA
#> 60     4    7               <NA>          <NA>    NA  NA  NA      NA  18160.800
#> 61     4    8               <NA>          <NA>    NA  NA  NA      NA  44102.000
#> 62     4    9               <NA>          <NA>    NA  NA  NA      NA  36248.595
#> 63     4   10               <NA>          <NA>    NA  NA  NA      NA  18350.251
#> 64     4   11               <NA>          <NA>    NA  NA  NA      NA 289267.000
#> 65     4   12               <NA>          <NA>    NA  NA  NA      NA 102005.691
#> 66     5    0          potential            <=  71.5  67  74      78         NA
#> 67     5    1            overall            <=  63.5  68  69      73         NA
#> 68     5    2               <NA>          <NA>    NA  NA  NA      NA   2026.420
#> 69     5    3          defending            <=  55.5  70  71      72         NA
#> 70     5    4               <NA>          <NA>    NA  NA  NA      NA  34130.720
#> 71     5    5               <NA>          <NA>    NA  NA  NA      NA  23844.260
#> 72     5    6               <NA>          <NA>    NA  NA  NA      NA  10398.270
#> 73     5    7               <NA>          <NA>    NA  NA  NA      NA  14664.230
#> 74     5    8            overall            <=  62.5  75  76      77         NA
#> 75     5    9               <NA>          <NA>    NA  NA  NA      NA  19785.220
#> 76     5   10               <NA>          <NA>    NA  NA  NA      NA 143840.300
#> 77     5   11               <NA>          <NA>    NA  NA  NA      NA 125105.113
#> 78     5   12               <NA>          <NA>    NA  NA  NA      NA  68996.456
#> 79     6    0            overall            <=  70.5  80  90      91         NA
#> 80     6    1          potential            <=  76.5  81  85      89         NA
#> 81     6    2            overall            <=  63.5  82  83      84         NA
#> 82     6    3               <NA>          <NA>    NA  NA  NA      NA   7741.378
#> 83     6    4               <NA>          <NA>    NA  NA  NA      NA  26837.494
#> 84     6    5               <NA>          <NA>    NA  NA  NA      NA  17901.106
#> 85     6    6            overall            <=  61.5  86  87      88         NA
#> 86     6    7               <NA>          <NA>    NA  NA  NA      NA  21366.198
#> 87     6    8               <NA>          <NA>    NA  NA  NA      NA  45777.770
#> 88     6    9               <NA>          <NA>    NA  NA  NA      NA  42147.154
#> 89     6   10               <NA>          <NA>    NA  NA  NA      NA  20438.191
#> 90     6   11               <NA>          <NA>    NA  NA  NA      NA 342956.270
#> 91     6   12               <NA>          <NA>    NA  NA  NA      NA 104852.595
#> 92     7    0            overall            <=  70.5  93 100     104         NA
#> 93     7    1          potential            <=  77.5  94  98      99         NA
#> 94     7    2          dribbling            <=  63.5  95  96      97         NA
#> 95     7    3               <NA>          <NA>    NA  NA  NA      NA  16308.440
#> 96     7    4               <NA>          <NA>    NA  NA  NA      NA  30033.632
#> 97     7    5               <NA>          <NA>    NA  NA  NA      NA   6567.240
#> 98     7    6               <NA>          <NA>    NA  NA  NA      NA  44699.993
#> 99     7    7               <NA>          <NA>    NA  NA  NA      NA  21539.463
#> 100    7    8          potential            <=  71.5 101 102     103         NA
#> 101    7    9               <NA>          <NA>    NA  NA  NA      NA  14160.247
#> 102    7   10               <NA>          <NA>    NA  NA  NA      NA 362160.643
#> 103    7   11               <NA>          <NA>    NA  NA  NA      NA 337090.289
#> 104    7   12               <NA>          <NA>    NA  NA  NA      NA 102058.615
#> 105    8    0            overall            <=  69.5 106 113     117         NA
#> 106    8    1            overall            <=  61.5 107 108     112         NA
#> 107    8    2               <NA>          <NA>    NA  NA  NA      NA   7410.516
#> 108    8    3          potential            <=  66.5 109 110     111         NA
#> 109    8    4               <NA>          <NA>    NA  NA  NA      NA  22677.596
#> 110    8    5               <NA>          <NA>    NA  NA  NA      NA  33099.149
#> 111    8    6               <NA>          <NA>    NA  NA  NA      NA  30544.548
#> 112    8    7               <NA>          <NA>    NA  NA  NA      NA  22623.153
#> 113    8    8          potential            <=  72.5 114 115     116         NA
#> 114    8    9               <NA>          <NA>    NA  NA  NA      NA  16612.635
#> 115    8   10               <NA>          <NA>    NA  NA  NA      NA 329444.579
#> 116    8   11               <NA>          <NA>    NA  NA  NA      NA 257443.576
#> 117    8   12               <NA>          <NA>    NA  NA  NA      NA  97085.303
#> 118    9    0            overall            <=  71.5 119 129     130         NA
#> 119    9    1            overall            <=  61.5 120 121     128         NA
#> 120    9    2               <NA>          <NA>    NA  NA  NA      NA  10936.545
#> 121    9    3          defending            <=  52.5 122 126     127         NA
#> 122    9    4          potential            <=  65.5 123 124     125         NA
#> 123    9    5               <NA>          <NA>    NA  NA  NA      NA  25219.426
#> 124    9    6               <NA>          <NA>    NA  NA  NA      NA  38559.141
#> 125    9    7               <NA>          <NA>    NA  NA  NA      NA  36755.248
#> 126    9    8               <NA>          <NA>    NA  NA  NA      NA  28728.673
#> 127    9    9               <NA>          <NA>    NA  NA  NA      NA  18045.960
#> 128    9   10               <NA>          <NA>    NA  NA  NA      NA  24548.439
#> 129    9   11               <NA>          <NA>    NA  NA  NA      NA 350000.121
#> 130    9   12               <NA>          <NA>    NA  NA  NA      NA  96055.932
#> 131   10    0            overall            <=  71.5 132 142     143         NA
#> 132   10    1            overall            <=  60.5 133 134     141         NA
#> 133   10    2               <NA>          <NA>    NA  NA  NA      NA  11342.890
#> 134   10    3          potential            <=  76.5 135 139     140         NA
#> 135   10    4                age            <=  29.5 136 137     138         NA
#> 136   10    5               <NA>          <NA>    NA  NA  NA      NA  31961.941
#> 137   10    6               <NA>          <NA>    NA  NA  NA      NA  13877.542
#> 138   10    7               <NA>          <NA>    NA  NA  NA      NA  27656.509
#> 139   10    8               <NA>          <NA>    NA  NA  NA      NA  44242.211
#> 140   10    9               <NA>          <NA>    NA  NA  NA      NA  29779.960
#> 141   10   10               <NA>          <NA>    NA  NA  NA      NA  24981.377
#> 142   10   11               <NA>          <NA>    NA  NA  NA      NA 318500.109
#> 143   10   12               <NA>          <NA>    NA  NA  NA      NA  90950.025
#> 144   11    0          potential            <=  75.5 145 152     156         NA
#> 145   11    1           shooting            <=  56.5 146 147     151         NA
#> 146   11    2               <NA>          <NA>    NA  NA  NA      NA  22646.696
#> 147   11    3                age            <=  30.5 148 149     150         NA
#> 148   11    4               <NA>          <NA>    NA  NA  NA      NA  36710.668
#> 149   11    5               <NA>          <NA>    NA  NA  NA      NA  10150.098
#> 150   11    6               <NA>          <NA>    NA  NA  NA      NA  30984.358
#> 151   11    7               <NA>          <NA>    NA  NA  NA      NA  13941.521
#> 152   11    8            overall            <=  68.5 153 154     155         NA
#> 153   11    9               <NA>          <NA>    NA  NA  NA      NA  34228.634
#> 154   11   10               <NA>          <NA>    NA  NA  NA      NA 440150.098
#> 155   11   11               <NA>          <NA>    NA  NA  NA      NA 319057.392
#> 156   11   12               <NA>          <NA>    NA  NA  NA      NA  97169.416
#> 157   12    0            overall            <=  72.5 158 168     169         NA
#> 158   12    1            overall            <=  60.5 159 160     167         NA
#> 159   12    2               <NA>          <NA>    NA  NA  NA      NA  14943.931
#> 160   12    3                age            <=  29.5 161 165     166         NA
#> 161   12    4          defending            <=  62.5 162 163     164         NA
#> 162   12    5               <NA>          <NA>    NA  NA  NA      NA  36076.106
#> 163   12    6               <NA>          <NA>    NA  NA  NA      NA  30005.555
#> 164   12    7               <NA>          <NA>    NA  NA  NA      NA  24654.362
#> 165   12    8               <NA>          <NA>    NA  NA  NA      NA  16353.811
#> 166   12    9               <NA>          <NA>    NA  NA  NA      NA  29794.263
#> 167   12   10               <NA>          <NA>    NA  NA  NA      NA  26156.882
#> 168   12   11               <NA>          <NA>    NA  NA  NA      NA 349635.088
#> 169   12   12               <NA>          <NA>    NA  NA  NA      NA  84028.297
#> 170   13    0            overall            <=  73.5 171 181     182         NA
#> 171   13    1            overall            <=  60.5 172 173     180         NA
#> 172   13    2               <NA>          <NA>    NA  NA  NA      NA  16949.538
#> 173   13    3                age            <=  30.5 174 178     179         NA
#> 174   13    4          potential            <=  65.5 175 176     177         NA
#> 175   13    5               <NA>          <NA>    NA  NA  NA      NA  27319.274
#> 176   13    6               <NA>          <NA>    NA  NA  NA      NA  35533.395
#> 177   13    7               <NA>          <NA>    NA  NA  NA      NA  34686.883
#> 178   13    8               <NA>          <NA>    NA  NA  NA      NA   9472.479
#> 179   13    9               <NA>          <NA>    NA  NA  NA      NA  30405.744
#> 180   13   10               <NA>          <NA>    NA  NA  NA      NA  27304.824
#> 181   13   11               <NA>          <NA>    NA  NA  NA      NA 468171.579
#> 182   13   12               <NA>          <NA>    NA  NA  NA      NA  90113.503
#> 183   14    0            overall            <=  73.5 184 194     195         NA
#> 184   14    1            overall            <=  59.5 185 186     193         NA
#> 185   14    2               <NA>          <NA>    NA  NA  NA      NA  16754.585
#> 186   14    3          potential            <=  65.5 187 188     192         NA
#> 187   14    4               <NA>          <NA>    NA  NA  NA      NA  25754.585
#> 188   14    5          defending            <=  57.5 189 190     191         NA
#> 189   14    6               <NA>          <NA>    NA  NA  NA      NA  37684.639
#> 190   14    7               <NA>          <NA>    NA  NA  NA      NA  30903.591
#> 191   14    8               <NA>          <NA>    NA  NA  NA      NA  21118.273
#> 192   14    9               <NA>          <NA>    NA  NA  NA      NA  31855.480
#> 193   14   10               <NA>          <NA>    NA  NA  NA      NA  29024.546
#> 194   14   11               <NA>          <NA>    NA  NA  NA      NA 424854.421
#> 195   14   12               <NA>          <NA>    NA  NA  NA      NA  86846.286
#> 196   15    0            overall            <=  72.5 197 207     208         NA
#> 197   15    1            overall            <=  59.5 198 199     206         NA
#> 198   15    2               <NA>          <NA>    NA  NA  NA      NA  18579.126
#> 199   15    3          potential            <=  65.5 200 201     205         NA
#> 200   15    4               <NA>          <NA>    NA  NA  NA      NA  26474.992
#> 201   15    5          defending            <=  65.5 202 203     204         NA
#> 202   15    6               <NA>          <NA>    NA  NA  NA      NA  35841.246
#> 203   15    7               <NA>          <NA>    NA  NA  NA      NA  23863.801
#> 204   15    8               <NA>          <NA>    NA  NA  NA      NA  27023.759
#> 205   15    9               <NA>          <NA>    NA  NA  NA      NA  31763.481
#> 206   15   10               <NA>          <NA>    NA  NA  NA      NA  29183.397
#> 207   15   11               <NA>          <NA>    NA  NA  NA      NA 276212.922
#> 208   15   12               <NA>          <NA>    NA  NA  NA      NA  74134.715
#> 209   16    0          potential            <=  77.5 210 220     221         NA
#> 210   16    1            overall            <=  59.5 211 212     219         NA
#> 211   16    2               <NA>          <NA>    NA  NA  NA      NA  20221.213
#> 212   16    3                age            <=  31.5 213 217     218         NA
#> 213   16    4          potential            <=  64.5 214 215     216         NA
#> 214   16    5               <NA>          <NA>    NA  NA  NA      NA  27137.563
#> 215   16    6               <NA>          <NA>    NA  NA  NA      NA  34895.857
#> 216   16    7               <NA>          <NA>    NA  NA  NA      NA  34205.020
#> 217   16    8               <NA>          <NA>    NA  NA  NA      NA   2399.371
#> 218   16    9               <NA>          <NA>    NA  NA  NA      NA  29594.233
#> 219   16   10               <NA>          <NA>    NA  NA  NA      NA  27856.469
#> 220   16   11               <NA>          <NA>    NA  NA  NA      NA 211747.687
#> 221   16   12               <NA>          <NA>    NA  NA  NA      NA  59588.218
#> 222   17    0            overall            <=  73.5 223 233     234         NA
#> 223   17    1            overall            <=  59.5 224 225     232         NA
#> 224   17    2               <NA>          <NA>    NA  NA  NA      NA  21699.092
#> 225   17    3          potential            <=  64.5 226 227     231         NA
#> 226   17    4               <NA>          <NA>    NA  NA  NA      NA  27113.737
#> 227   17    5                age            <=  31.5 228 229     230         NA
#> 228   17    6               <NA>          <NA>    NA  NA  NA      NA  33929.685
#> 229   17    7               <NA>          <NA>    NA  NA  NA      NA   1226.234
#> 230   17    8               <NA>          <NA>    NA  NA  NA      NA  30784.573
#> 231   17    9               <NA>          <NA>    NA  NA  NA      NA  30427.107
#> 232   17   10               <NA>          <NA>    NA  NA  NA      NA  28811.722
#> 233   17   11               <NA>          <NA>    NA  NA  NA      NA 344072.918
#> 234   17   12               <NA>          <NA>    NA  NA  NA      NA  75864.602
#> 235   18    0            overall            <=  74.5 236 246     247         NA
#> 236   18    1            overall            <=  58.5 237 238     245         NA
#> 237   18    2               <NA>          <NA>    NA  NA  NA      NA  22029.183
#> 238   18    3          potential            <=  65.5 239 240     244         NA
#> 239   18    4               <NA>          <NA>    NA  NA  NA      NA  28444.939
#> 240   18    5                age            <=  32.5 241 242     243         NA
#> 241   18    6               <NA>          <NA>    NA  NA  NA      NA  34018.313
#> 242   18    7               <NA>          <NA>    NA  NA  NA      NA -35477.604
#> 243   18    8               <NA>          <NA>    NA  NA  NA      NA  28989.237
#> 244   18    9               <NA>          <NA>    NA  NA  NA      NA  28909.098
#> 245   18   10               <NA>          <NA>    NA  NA  NA      NA  27901.283
#> 246   18   11               <NA>          <NA>    NA  NA  NA      NA 447008.218
#> 247   18   12               <NA>          <NA>    NA  NA  NA      NA  76970.592
#> 248   19    0            overall            <=  73.5 249 259     260         NA
#> 249   19    1            overall            <=  59.5 250 251     258         NA
#> 250   19    2               <NA>          <NA>    NA  NA  NA      NA  24326.265
#> 251   19    3                age            <=  32.5 252 256     257         NA
#> 252   19    4          potential            <=  64.5 253 254     255         NA
#> 253   19    5               <NA>          <NA>    NA  NA  NA      NA  29429.791
#> 254   19    6               <NA>          <NA>    NA  NA  NA      NA  34007.847
#> 255   19    7               <NA>          <NA>    NA  NA  NA      NA  33611.029
#> 256   19    8               <NA>          <NA>    NA  NA  NA      NA  13623.081
#> 257   19    9               <NA>          <NA>    NA  NA  NA      NA  32007.922
#> 258   19   10               <NA>          <NA>    NA  NA  NA      NA  30503.044
#> 259   19   11               <NA>          <NA>    NA  NA  NA      NA 271964.805
#> 260   19   12               <NA>          <NA>    NA  NA  NA      NA  65484.483
#>     Cover
#> 1     501
#> 2       0
#> 3       0
#> 4       0
#> 5       0
#> 6       0
#> 7       0
#> 8       0
#> 9     501
#> 10      0
#> 11    501
#> 12      0
#> 13      0
#> 14    501
#> 15      0
#> 16      0
#> 17      0
#> 18      0
#> 19      0
#> 20      0
#> 21      0
#> 22    501
#> 23      0
#> 24    501
#> 25      0
#> 26      0
#> 27    501
#> 28      0
#> 29      0
#> 30      0
#> 31      0
#> 32      0
#> 33      0
#> 34      0
#> 35      0
#> 36      0
#> 37      0
#> 38    501
#> 39      0
#> 40    501
#> 41      0
#> 42      0
#> 43      0
#> 44      0
#> 45      0
#> 46      0
#> 47      0
#> 48      0
#> 49      0
#> 50      0
#> 51    501
#> 52      0
#> 53    501
#> 54      0
#> 55      0
#> 56      0
#> 57      0
#> 58      0
#> 59      0
#> 60      0
#> 61      0
#> 62      0
#> 63      0
#> 64    501
#> 65      0
#> 66    501
#> 67      0
#> 68      0
#> 69      0
#> 70      0
#> 71      0
#> 72      0
#> 73      0
#> 74    501
#> 75      0
#> 76    501
#> 77      0
#> 78      0
#> 79    501
#> 80      0
#> 81      0
#> 82      0
#> 83      0
#> 84      0
#> 85      0
#> 86      0
#> 87      0
#> 88      0
#> 89      0
#> 90    501
#> 91      0
#> 92    501
#> 93      0
#> 94      0
#> 95      0
#> 96      0
#> 97      0
#> 98      0
#> 99      0
#> 100   501
#> 101     0
#> 102   501
#> 103     0
#> 104     0
#> 105   501
#> 106     0
#> 107     0
#> 108     0
#> 109     0
#> 110     0
#> 111     0
#> 112     0
#> 113   501
#> 114     0
#> 115   501
#> 116     0
#> 117     0
#> 118   501
#> 119     0
#> 120     0
#> 121     0
#> 122     0
#> 123     0
#> 124     0
#> 125     0
#> 126     0
#> 127     0
#> 128     0
#> 129   501
#> 130     0
#> 131   501
#> 132     0
#> 133     0
#> 134     0
#> 135     0
#> 136     0
#> 137     0
#> 138     0
#> 139     0
#> 140     0
#> 141     0
#> 142   501
#> 143     0
#> 144   501
#> 145     0
#> 146     0
#> 147     0
#> 148     0
#> 149     0
#> 150     0
#> 151     0
#> 152   501
#> 153     0
#> 154   501
#> 155     0
#> 156     0
#> 157   501
#> 158     0
#> 159     0
#> 160     0
#> 161     0
#> 162     0
#> 163     0
#> 164     0
#> 165     0
#> 166     0
#> 167     0
#> 168   501
#> 169     0
#> 170   501
#> 171     0
#> 172     0
#> 173     0
#> 174     0
#> 175     0
#> 176     0
#> 177     0
#> 178     0
#> 179     0
#> 180     0
#> 181   501
#> 182     0
#> 183   501
#> 184     0
#> 185     0
#> 186     0
#> 187     0
#> 188     0
#> 189     0
#> 190     0
#> 191     0
#> 192     0
#> 193     0
#> 194   501
#> 195     0
#> 196   501
#> 197     0
#> 198     0
#> 199     0
#> 200     0
#> 201     0
#> 202     0
#> 203     0
#> 204     0
#> 205     0
#> 206     0
#> 207   501
#> 208     0
#> 209   501
#> 210     0
#> 211     0
#> 212     0
#> 213     0
#> 214     0
#> 215     0
#> 216     0
#> 217     0
#> 218     0
#> 219     0
#> 220   501
#> 221     0
#> 222   501
#> 223     0
#> 224     0
#> 225     0
#> 226     0
#> 227     0
#> 228     0
#> 229     0
#> 230     0
#> 231     0
#> 232     0
#> 233   501
#> 234     0
#> 235   501
#> 236     0
#> 237     0
#> 238     0
#> 239     0
#> 240     0
#> 241     0
#> 242     0
#> 243     0
#> 244     0
#> 245     0
#> 246   501
#> 247     0
#> 248   501
#> 249     0
#> 250     0
#> 251     0
#> 252     0
#> 253     0
#> 254     0
#> 255     0
#> 256     0
#> 257     0
#> 258     0
#> 259   501
#> 260     0
# }