Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node   Feature Decision.type Split Yes  No Missing Prediction Cover
#> 1      0    0   overall            <=  65.5   2   9      13         NA   501
#> 2      0    1 potential            <=  73.5   3   4       8         NA     0
#> 3      0    2      <NA>          <NA>    NA  NA  NA      NA  -7000.000     0
#> 4      0    3   overall            <=  62.5   5   6       7         NA     0
#> 5      0    4      <NA>          <NA>    NA  NA  NA      NA   7500.000     0
#> 6      0    5      <NA>          <NA>    NA  NA  NA      NA  42500.000     0
#> 7      0    6      <NA>          <NA>    NA  NA  NA      NA  24887.387     0
#> 8      0    7      <NA>          <NA>    NA  NA  NA      NA  -1113.447     0
#> 9      0    8 potential            <=  68.5  10  11      12         NA   501
#> 10     0    9      <NA>          <NA>    NA  NA  NA      NA  30000.000     0
#> 11     0   10      <NA>          <NA>    NA  NA  NA      NA 205000.000   501
#> 12     0   11      <NA>          <NA>    NA  NA  NA      NA 178803.245     0
#> 13     0   12      <NA>          <NA>    NA  NA  NA      NA  95941.952     0
#> 14     1    0   overall            <=  66.5  15  22      26         NA   501
#> 15     1    1 potential            <=  73.5  16  17      21         NA     0
#> 16     1    2      <NA>          <NA>    NA  NA  NA      NA   -800.000     0
#> 17     1    3   overall            <=  62.5  18  19      20         NA     0
#> 18     1    4      <NA>          <NA>    NA  NA  NA      NA  10250.000     0
#> 19     1    5      <NA>          <NA>    NA  NA  NA      NA  44250.000     0
#> 20     1    6      <NA>          <NA>    NA  NA  NA      NA  29080.189     0
#> 21     1    7      <NA>          <NA>    NA  NA  NA      NA   4857.081     0
#> 22     1    8 potential            <=  69.5  23  24      25         NA   501
#> 23     1    9      <NA>          <NA>    NA  NA  NA      NA  35500.000     0
#> 24     1   10      <NA>          <NA>    NA  NA  NA      NA 248000.000   501
#> 25     1   11      <NA>          <NA>    NA  NA  NA      NA 212292.237     0
#> 26     1   12      <NA>          <NA>    NA  NA  NA      NA 104273.427     0
#> 27     2    0   overall            <=  67.5  28  35      39         NA   501
#> 28     2    1   overall            <=  63.5  29  30      34         NA     0
#> 29     2    2      <NA>          <NA>    NA  NA  NA      NA  -3220.000     0
#> 30     2    3       age            <=  24.5  31  32      33         NA     0
#> 31     2    4      <NA>          <NA>    NA  NA  NA      NA  40280.000     0
#> 32     2    5      <NA>          <NA>    NA  NA  NA      NA  22780.000     0
#> 33     2    6      <NA>          <NA>    NA  NA  NA      NA  30229.584     0
#> 34     2    7      <NA>          <NA>    NA  NA  NA      NA  11178.660     0
#> 35     2    8       age            <=  32.5  36  37      38         NA   501
#> 36     2    9      <NA>          <NA>    NA  NA  NA      NA 236700.000   435
#> 37     2   10      <NA>          <NA>    NA  NA  NA      NA  36700.000    66
#> 38     2   11      <NA>          <NA>    NA  NA  NA      NA 215189.583     0
#> 39     2   12      <NA>          <NA>    NA  NA  NA      NA  96899.411     0
#> 40     3    0   overall            <=  69.5  41  51      52         NA   501
#> 41     3    1 potential            <=  74.5  42  46      50         NA     0
#> 42     3    2   overall            <=  64.5  43  44      45         NA     0
#> 43     3    3      <NA>          <NA>    NA  NA  NA      NA   1602.000     0
#> 44     3    4      <NA>          <NA>    NA  NA  NA      NA  26172.000     0
#> 45     3    5      <NA>          <NA>    NA  NA  NA      NA  11504.667     0
#> 46     3    6   overall            <=  61.5  47  48      49         NA     0
#> 47     3    7      <NA>          <NA>    NA  NA  NA      NA   9047.000     0
#> 48     3    8      <NA>          <NA>    NA  NA  NA      NA  46172.000     0
#> 49     3    9      <NA>          <NA>    NA  NA  NA      NA  37533.298     0
#> 50     3   10      <NA>          <NA>    NA  NA  NA      NA  15798.049     0
#> 51     3   11      <NA>          <NA>    NA  NA  NA      NA 311530.000   501
#> 52     3   12      <NA>          <NA>    NA  NA  NA      NA 107504.401     0
#> 53     4    0   overall            <=  69.5  54  64      65         NA   501
#> 54     4    1 potential            <=  75.5  55  59      63         NA     0
#> 55     4    2   overall            <=  63.5  56  57      58         NA     0
#> 56     4    3      <NA>          <NA>    NA  NA  NA      NA   2941.800     0
#> 57     4    4      <NA>          <NA>    NA  NA  NA      NA  27412.800     0
#> 58     4    5      <NA>          <NA>    NA  NA  NA      NA  14953.223     0
#> 59     4    6   overall            <=  61.5  60  61      62         NA     0
#> 60     4    7      <NA>          <NA>    NA  NA  NA      NA  16642.300     0
#> 61     4    8      <NA>          <NA>    NA  NA  NA      NA  51679.800     0
#> 62     4    9      <NA>          <NA>    NA  NA  NA      NA  44594.052     0
#> 63     4   10      <NA>          <NA>    NA  NA  NA      NA  18732.769     0
#> 64     4   11      <NA>          <NA>    NA  NA  NA      NA 298877.000   501
#> 65     4   12      <NA>          <NA>    NA  NA  NA      NA 105114.916     0
#> 66     5    0 potential            <=  71.5  67  71      78         NA   501
#> 67     5    1  shooting            <=  59.5  68  69      70         NA     0
#> 68     5    2      <NA>          <NA>    NA  NA  NA      NA  12063.520     0
#> 69     5    3      <NA>          <NA>    NA  NA  NA      NA  29443.520     0
#> 70     5    4      <NA>          <NA>    NA  NA  NA      NA   -852.380     0
#> 71     5    5   overall            <=  62.5  72  73      77         NA   501
#> 72     5    6      <NA>          <NA>    NA  NA  NA      NA  20978.070     0
#> 73     5    7   overall            <=  69.5  74  75      76         NA   501
#> 74     5    8      <NA>          <NA>    NA  NA  NA      NA  40193.520     0
#> 75     5    9      <NA>          <NA>    NA  NA  NA      NA 352489.300   501
#> 76     5   10      <NA>          <NA>    NA  NA  NA      NA 243412.676     0
#> 77     5   11      <NA>          <NA>    NA  NA  NA      NA 211069.509     0
#> 78     5   12      <NA>          <NA>    NA  NA  NA      NA 109723.235     0
#> 79     6    0 potential            <=  71.5  80  84      91         NA   501
#> 80     6    1 dribbling            <=  63.5  81  82      83         NA     0
#> 81     6    2      <NA>          <NA>    NA  NA  NA      NA   9441.268     0
#> 82     6    3      <NA>          <NA>    NA  NA  NA      NA  29977.168     0
#> 83     6    4      <NA>          <NA>    NA  NA  NA      NA   2732.858     0
#> 84     6    5   overall            <=  61.5  85  86      90         NA   501
#> 85     6    6      <NA>          <NA>    NA  NA  NA      NA  12549.813     0
#> 86     6    7 gk_diving            <=  73.5  87  88      89         NA   501
#> 87     6    8      <NA>          <NA>    NA  NA  NA      NA  23173.268     0
#> 88     6    9      <NA>          <NA>    NA  NA  NA      NA 420740.370    56
#> 89     6   10      <NA>          <NA>    NA  NA  NA      NA 110740.370   445
#> 90     6   11      <NA>          <NA>    NA  NA  NA      NA 107391.093     0
#> 91     6   12      <NA>          <NA>    NA  NA  NA      NA  59541.770     0
#> 92     7    0   overall            <=  70.5  93 103     104         NA   501
#> 93     7    1   overall            <=  62.5  94  95     102         NA     0
#> 94     7    2      <NA>          <NA>    NA  NA  NA      NA   8125.282     0
#> 95     7    3 potential            <=  76.5  96 100     101         NA     0
#> 96     7    4 defending            <=  57.5  97  98      99         NA     0
#> 97     7    5      <NA>          <NA>    NA  NA  NA      NA  34158.451     0
#> 98     7    6      <NA>          <NA>    NA  NA  NA      NA  24739.451     0
#> 99     7    7      <NA>          <NA>    NA  NA  NA      NA  16834.841     0
#> 100    7    8      <NA>          <NA>    NA  NA  NA      NA  49418.431     0
#> 101    7    9      <NA>          <NA>    NA  NA  NA      NA  30496.718     0
#> 102    7   10      <NA>          <NA>    NA  NA  NA      NA  21982.340     0
#> 103    7   11      <NA>          <NA>    NA  NA  NA      NA 313166.333   501
#> 104    7   12      <NA>          <NA>    NA  NA  NA      NA  98769.015     0
#> 105    8    0   overall            <=  70.5 106 116     117         NA   501
#> 106    8    1   overall            <=  63.5 107 108     115         NA     0
#> 107    8    2      <NA>          <NA>    NA  NA  NA      NA  13684.613     0
#> 108    8    3 potential            <=  77.5 109 113     114         NA     0
#> 109    8    4 defending            <=  61.5 110 111     112         NA     0
#> 110    8    5      <NA>          <NA>    NA  NA  NA      NA  34285.606     0
#> 111    8    6      <NA>          <NA>    NA  NA  NA      NA  24047.096     0
#> 112    8    7      <NA>          <NA>    NA  NA  NA      NA  18651.357     0
#> 113    8    8      <NA>          <NA>    NA  NA  NA      NA  60476.588     0
#> 114    8    9      <NA>          <NA>    NA  NA  NA      NA  31647.554     0
#> 115    8   10      <NA>          <NA>    NA  NA  NA      NA  23475.020     0
#> 116    8   11      <NA>          <NA>    NA  NA  NA      NA 289728.055   501
#> 117    8   12      <NA>          <NA>    NA  NA  NA      NA  94677.823     0
#> 118    9    0   overall            <=  71.5 119 129     130         NA   501
#> 119    9    1   overall            <=  62.5 120 121     128         NA     0
#> 120    9    2      <NA>          <NA>    NA  NA  NA      NA  12816.152     0
#> 121    9    3       age            <=  28.5 122 126     127         NA     0
#> 122    9    4 defending            <=  57.5 123 124     125         NA     0
#> 123    9    5      <NA>          <NA>    NA  NA  NA      NA  39357.045     0
#> 124    9    6      <NA>          <NA>    NA  NA  NA      NA  32819.780     0
#> 125    9    7      <NA>          <NA>    NA  NA  NA      NA  27206.852     0
#> 126    9    8      <NA>          <NA>    NA  NA  NA      NA  17907.627     0
#> 127    9    9      <NA>          <NA>    NA  NA  NA      NA  30336.320     0
#> 128    9   10      <NA>          <NA>    NA  NA  NA      NA  24057.260     0
#> 129    9   11      <NA>          <NA>    NA  NA  NA      NA 359876.894   501
#> 130    9   12      <NA>          <NA>    NA  NA  NA      NA  96887.386     0
#> 131   10    0 potential            <=  74.5 132 139     143         NA   501
#> 132   10    1   overall            <=  61.5 133 134     138         NA     0
#> 133   10    2      <NA>          <NA>    NA  NA  NA      NA  11996.968     0
#> 134   10    3       age            <=  29.5 135 136     137         NA     0
#> 135   10    4      <NA>          <NA>    NA  NA  NA      NA  33684.498     0
#> 136   10    5      <NA>          <NA>    NA  NA  NA      NA  13894.090     0
#> 137   10    6      <NA>          <NA>    NA  NA  NA      NA  27375.294     0
#> 138   10    7      <NA>          <NA>    NA  NA  NA      NA  22644.059     0
#> 139   10    8   overall            <=  68.5 140 141     142         NA   501
#> 140   10    9      <NA>          <NA>    NA  NA  NA      NA  34627.371     0
#> 141   10   10      <NA>          <NA>    NA  NA  NA      NA 377389.205   501
#> 142   10   11      <NA>          <NA>    NA  NA  NA      NA 260914.341     0
#> 143   10   12      <NA>          <NA>    NA  NA  NA      NA  95071.551     0
#> 144   11    0   overall            <=  72.5 145 155     156         NA   501
#> 145   11    1   overall            <=  60.5 146 147     154         NA     0
#> 146   11    2      <NA>          <NA>    NA  NA  NA      NA  13297.271     0
#> 147   11    3       age            <=  28.5 148 152     153         NA     0
#> 148   11    4 defending            <=  52.5 149 150     151         NA     0
#> 149   11    5      <NA>          <NA>    NA  NA  NA      NA  38112.497     0
#> 150   11    6      <NA>          <NA>    NA  NA  NA      NA  32116.763     0
#> 151   11    7      <NA>          <NA>    NA  NA  NA      NA  28117.717     0
#> 152   11    8      <NA>          <NA>    NA  NA  NA      NA  18726.681     0
#> 153   11    9      <NA>          <NA>    NA  NA  NA      NA  29854.242     0
#> 154   11   10      <NA>          <NA>    NA  NA  NA      NA  25869.411     0
#> 155   11   11      <NA>          <NA>    NA  NA  NA      NA 393150.284   501
#> 156   11   12      <NA>          <NA>    NA  NA  NA      NA  92581.989     0
#> 157   12    0   overall            <=  72.5 158 168     169         NA   501
#> 158   12    1   overall            <=  60.5 159 160     167         NA     0
#> 159   12    2      <NA>          <NA>    NA  NA  NA      NA  15505.113     0
#> 160   12    3       age            <=  30.5 161 165     166         NA     0
#> 161   12    4 potential            <=  65.5 162 163     164         NA     0
#> 162   12    5      <NA>          <NA>    NA  NA  NA      NA  27375.746     0
#> 163   12    6      <NA>          <NA>    NA  NA  NA      NA  35258.758     0
#> 164   12    7      <NA>          <NA>    NA  NA  NA      NA  34335.975     0
#> 165   12    8      <NA>          <NA>    NA  NA  NA      NA  11548.305     0
#> 166   12    9      <NA>          <NA>    NA  NA  NA      NA  30546.102     0
#> 167   12   10      <NA>          <NA>    NA  NA  NA      NA  26823.442     0
#> 168   12   11      <NA>          <NA>    NA  NA  NA      NA 357335.256   501
#> 169   12   12      <NA>          <NA>    NA  NA  NA      NA  85989.361     0
#> 170   13    0   overall            <=  73.5 171 181     182         NA   501
#> 171   13    1   overall            <=  59.5 172 173     180         NA     0
#> 172   13    2      <NA>          <NA>    NA  NA  NA      NA  15454.601     0
#> 173   13    3       age            <=  30.5 174 178     179         NA     0
#> 174   13    4 defending            <=  53.5 175 176     177         NA     0
#> 175   13    5      <NA>          <NA>    NA  NA  NA      NA  37517.965     0
#> 176   13    6      <NA>          <NA>    NA  NA  NA      NA  31822.133     0
#> 177   13    7      <NA>          <NA>    NA  NA  NA      NA  27825.926     0
#> 178   13    8      <NA>          <NA>    NA  NA  NA      NA  14151.173     0
#> 179   13    9      <NA>          <NA>    NA  NA  NA      NA  30503.838     0
#> 180   13   10      <NA>          <NA>    NA  NA  NA      NA  27633.474     0
#> 181   13   11      <NA>          <NA>    NA  NA  NA      NA 425101.730   501
#> 182   13   12      <NA>          <NA>    NA  NA  NA      NA  84433.293     0
#> 183   14    0 potential            <=  76.5 184 194     195         NA   501
#> 184   14    1   overall            <=  59.5 185 186     193         NA     0
#> 185   14    2      <NA>          <NA>    NA  NA  NA      NA  17409.141     0
#> 186   14    3 potential            <=  65.5 187 188     192         NA     0
#> 187   14    4      <NA>          <NA>    NA  NA  NA      NA  25947.270     0
#> 188   14    5 defending            <=  63.5 189 190     191         NA     0
#> 189   14    6      <NA>          <NA>    NA  NA  NA      NA  36137.488     0
#> 190   14    7      <NA>          <NA>    NA  NA  NA      NA  23129.283     0
#> 191   14    8      <NA>          <NA>    NA  NA  NA      NA  26598.846     0
#> 192   14    9      <NA>          <NA>    NA  NA  NA      NA  30339.475     0
#> 193   14   10      <NA>          <NA>    NA  NA  NA      NA  27859.151     0
#> 194   14   11      <NA>          <NA>    NA  NA  NA      NA 156091.557   501
#> 195   14   12      <NA>          <NA>    NA  NA  NA      NA  53831.159     0
#> 196   15    0   overall            <=  72.5 197 207     208         NA   501
#> 197   15    1   overall            <=  60.5 198 199     206         NA     0
#> 198   15    2      <NA>          <NA>    NA  NA  NA      NA  20168.227     0
#> 199   15    3       age            <=  31.5 200 204     205         NA     0
#> 200   15    4 potential            <=  64.5 201 202     203         NA     0
#> 201   15    5      <NA>          <NA>    NA  NA  NA      NA  27614.543     0
#> 202   15    6      <NA>          <NA>    NA  NA  NA      NA  33720.108     0
#> 203   15    7      <NA>          <NA>    NA  NA  NA      NA  33255.303     0
#> 204   15    8      <NA>          <NA>    NA  NA  NA      NA  12793.141     0
#> 205   15    9      <NA>          <NA>    NA  NA  NA      NA  30837.863     0
#> 206   15   10      <NA>          <NA>    NA  NA  NA      NA  28249.782     0
#> 207   15   11      <NA>          <NA>    NA  NA  NA      NA 323982.402   501
#> 208   15   12      <NA>          <NA>    NA  NA  NA      NA  80251.349     0
#> 209   16    0   overall            <=  73.5 210 220     221         NA   501
#> 210   16    1   overall            <=  59.5 211 212     219         NA     0
#> 211   16    2      <NA>          <NA>    NA  NA  NA      NA  20613.835     0
#> 212   16    3 potential            <=  64.5 213 214     218         NA     0
#> 213   16    4      <NA>          <NA>    NA  NA  NA      NA  27124.448     0
#> 214   16    5 defending            <=  57.5 215 216     217         NA     0
#> 215   16    6      <NA>          <NA>    NA  NA  NA      NA  36070.040     0
#> 216   16    7      <NA>          <NA>    NA  NA  NA      NA  30620.434     0
#> 217   16    8      <NA>          <NA>    NA  NA  NA      NA  27522.890     0
#> 218   16    9      <NA>          <NA>    NA  NA  NA      NA  32097.852     0
#> 219   16   10      <NA>          <NA>    NA  NA  NA      NA  29817.453     0
#> 220   16   11      <NA>          <NA>    NA  NA  NA      NA 395084.161   501
#> 221   16   12      <NA>          <NA>    NA  NA  NA      NA  81695.797     0
#> 222   17    0   overall            <=  73.5 223 233     234         NA   501
#> 223   17    1   overall            <=  59.5 224 225     232         NA     0
#> 224   17    2      <NA>          <NA>    NA  NA  NA      NA  21090.021     0
#> 225   17    3 potential            <=  65.5 226 227     231         NA     0
#> 226   17    4      <NA>          <NA>    NA  NA  NA      NA  28965.996     0
#> 227   17    5 defending            <=  65.5 228 229     230         NA     0
#> 228   17    6      <NA>          <NA>    NA  NA  NA      NA  35066.206     0
#> 229   17    7      <NA>          <NA>    NA  NA  NA      NA  19854.007     0
#> 230   17    8      <NA>          <NA>    NA  NA  NA      NA  24762.577     0
#> 231   17    9      <NA>          <NA>    NA  NA  NA      NA  30611.646     0
#> 232   17   10      <NA>          <NA>    NA  NA  NA      NA  28817.663     0
#> 233   17   11      <NA>          <NA>    NA  NA  NA      NA 344075.745   501
#> 234   17   12      <NA>          <NA>    NA  NA  NA      NA  76766.972     0
#> 235   18    0   overall            <=  73.5 236 246     247         NA   501
#> 236   18    1   overall            <=  58.5 237 238     245         NA     0
#> 237   18    2      <NA>          <NA>    NA  NA  NA      NA  22278.709     0
#> 238   18    3 potential            <=  65.5 239 240     244         NA     0
#> 239   18    4      <NA>          <NA>    NA  NA  NA      NA  29514.940     0
#> 240   18    5       age            <=  30.5 241 242     243         NA     0
#> 241   18    6      <NA>          <NA>    NA  NA  NA      NA  34663.977     0
#> 242   18    7      <NA>          <NA>    NA  NA  NA      NA   7027.571     0
#> 243   18    8      <NA>          <NA>    NA  NA  NA      NA  31205.716     0
#> 244   18    9      <NA>          <NA>    NA  NA  NA      NA  30948.993     0
#> 245   18   10      <NA>          <NA>    NA  NA  NA      NA  29593.914     0
#> 246   18   11      <NA>          <NA>    NA  NA  NA      NA 328168.171   501
#> 247   18   12      <NA>          <NA>    NA  NA  NA      NA  73143.480     0
#> 248   19    0 potential            <=  78.5 249 256     260         NA   501
#> 249   19    1   overall            <=  58.5 250 251     255         NA     0
#> 250   19    2      <NA>          <NA>    NA  NA  NA      NA  23753.148     0
#> 251   19    3 potential            <=  64.5 252 253     254         NA     0
#> 252   19    4      <NA>          <NA>    NA  NA  NA      NA  28956.123     0
#> 253   19    5      <NA>          <NA>    NA  NA  NA      NA  33421.631     0
#> 254   19    6      <NA>          <NA>    NA  NA  NA      NA  32960.941     0
#> 255   19    7      <NA>          <NA>    NA  NA  NA      NA  31607.678     0
#> 256   19    8   overall            <=  64.5 257 258     259         NA   501
#> 257   19    9      <NA>          <NA>    NA  NA  NA      NA  24514.523     0
#> 258   19   10      <NA>          <NA>    NA  NA  NA      NA 348851.354   501
#> 259   19   11      <NA>          <NA>    NA  NA  NA      NA 315201.084     0
#> 260   19   12      <NA>          <NA>    NA  NA  NA      NA  70489.671     0