Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node            Feature Decision.type Split Yes  No Missing
#> 1      0    0            overall            <=  65.5   2   9      13
#> 2      0    1          potential            <=  73.5   3   4       8
#> 3      0    2               <NA>          <NA>    NA  NA  NA      NA
#> 4      0    3            overall            <=  62.5   5   6       7
#> 5      0    4               <NA>          <NA>    NA  NA  NA      NA
#> 6      0    5               <NA>          <NA>    NA  NA  NA      NA
#> 7      0    6               <NA>          <NA>    NA  NA  NA      NA
#> 8      0    7               <NA>          <NA>    NA  NA  NA      NA
#> 9      0    8          potential            <=  68.5  10  11      12
#> 10     0    9               <NA>          <NA>    NA  NA  NA      NA
#> 11     0   10               <NA>          <NA>    NA  NA  NA      NA
#> 12     0   11               <NA>          <NA>    NA  NA  NA      NA
#> 13     0   12               <NA>          <NA>    NA  NA  NA      NA
#> 14     1    0            overall            <=  66.5  15  22      26
#> 15     1    1          potential            <=  73.5  16  17      21
#> 16     1    2               <NA>          <NA>    NA  NA  NA      NA
#> 17     1    3            overall            <=  63.5  18  19      20
#> 18     1    4               <NA>          <NA>    NA  NA  NA      NA
#> 19     1    5               <NA>          <NA>    NA  NA  NA      NA
#> 20     1    6               <NA>          <NA>    NA  NA  NA      NA
#> 21     1    7               <NA>          <NA>    NA  NA  NA      NA
#> 22     1    8          potential            <=  69.5  23  24      25
#> 23     1    9               <NA>          <NA>    NA  NA  NA      NA
#> 24     1   10               <NA>          <NA>    NA  NA  NA      NA
#> 25     1   11               <NA>          <NA>    NA  NA  NA      NA
#> 26     1   12               <NA>          <NA>    NA  NA  NA      NA
#> 27     2    0            overall            <=  67.5  28  35      39
#> 28     2    1            overall            <=  63.5  29  30      34
#> 29     2    2               <NA>          <NA>    NA  NA  NA      NA
#> 30     2    3                age            <=  23.5  31  32      33
#> 31     2    4               <NA>          <NA>    NA  NA  NA      NA
#> 32     2    5               <NA>          <NA>    NA  NA  NA      NA
#> 33     2    6               <NA>          <NA>    NA  NA  NA      NA
#> 34     2    7               <NA>          <NA>    NA  NA  NA      NA
#> 35     2    8                age            <=  32.5  36  37      38
#> 36     2    9               <NA>          <NA>    NA  NA  NA      NA
#> 37     2   10               <NA>          <NA>    NA  NA  NA      NA
#> 38     2   11               <NA>          <NA>    NA  NA  NA      NA
#> 39     2   12               <NA>          <NA>    NA  NA  NA      NA
#> 40     3    0            overall            <=  69.5  41  51      52
#> 41     3    1          potential            <=  74.5  42  46      50
#> 42     3    2            overall            <=  64.5  43  44      45
#> 43     3    3               <NA>          <NA>    NA  NA  NA      NA
#> 44     3    4               <NA>          <NA>    NA  NA  NA      NA
#> 45     3    5               <NA>          <NA>    NA  NA  NA      NA
#> 46     3    6            overall            <=  61.5  47  48      49
#> 47     3    7               <NA>          <NA>    NA  NA  NA      NA
#> 48     3    8               <NA>          <NA>    NA  NA  NA      NA
#> 49     3    9               <NA>          <NA>    NA  NA  NA      NA
#> 50     3   10               <NA>          <NA>    NA  NA  NA      NA
#> 51     3   11               <NA>          <NA>    NA  NA  NA      NA
#> 52     3   12               <NA>          <NA>    NA  NA  NA      NA
#> 53     4    0            overall            <=  69.5  54  64      65
#> 54     4    1          potential            <=  75.5  55  62      63
#> 55     4    2            overall            <=  63.5  56  57      61
#> 56     4    3               <NA>          <NA>    NA  NA  NA      NA
#> 57     4    4          defending            <=  55.5  58  59      60
#> 58     4    5               <NA>          <NA>    NA  NA  NA      NA
#> 59     4    6               <NA>          <NA>    NA  NA  NA      NA
#> 60     4    7               <NA>          <NA>    NA  NA  NA      NA
#> 61     4    8               <NA>          <NA>    NA  NA  NA      NA
#> 62     4    9               <NA>          <NA>    NA  NA  NA      NA
#> 63     4   10               <NA>          <NA>    NA  NA  NA      NA
#> 64     4   11               <NA>          <NA>    NA  NA  NA      NA
#> 65     4   12               <NA>          <NA>    NA  NA  NA      NA
#> 66     5    0          potential            <=  71.5  67  71      78
#> 67     5    1           shooting            <=  59.5  68  69      70
#> 68     5    2               <NA>          <NA>    NA  NA  NA      NA
#> 69     5    3               <NA>          <NA>    NA  NA  NA      NA
#> 70     5    4               <NA>          <NA>    NA  NA  NA      NA
#> 71     5    5            overall            <=  62.5  72  73      77
#> 72     5    6               <NA>          <NA>    NA  NA  NA      NA
#> 73     5    7            overall            <=  69.5  74  75      76
#> 74     5    8               <NA>          <NA>    NA  NA  NA      NA
#> 75     5    9               <NA>          <NA>    NA  NA  NA      NA
#> 76     5   10               <NA>          <NA>    NA  NA  NA      NA
#> 77     5   11               <NA>          <NA>    NA  NA  NA      NA
#> 78     5   12               <NA>          <NA>    NA  NA  NA      NA
#> 79     6    0            overall            <=  70.5  80  90      91
#> 80     6    1            overall            <=  62.5  81  82      89
#> 81     6    2               <NA>          <NA>    NA  NA  NA      NA
#> 82     6    3                age            <=  24.5  83  87      88
#> 83     6    4          defending            <=  53.5  84  85      86
#> 84     6    5               <NA>          <NA>    NA  NA  NA      NA
#> 85     6    6               <NA>          <NA>    NA  NA  NA      NA
#> 86     6    7               <NA>          <NA>    NA  NA  NA      NA
#> 87     6    8               <NA>          <NA>    NA  NA  NA      NA
#> 88     6    9               <NA>          <NA>    NA  NA  NA      NA
#> 89     6   10               <NA>          <NA>    NA  NA  NA      NA
#> 90     6   11               <NA>          <NA>    NA  NA  NA      NA
#> 91     6   12               <NA>          <NA>    NA  NA  NA      NA
#> 92     7    0          potential            <=  72.5  93  97     104
#> 93     7    1          dribbling            <=  63.5  94  95      96
#> 94     7    2               <NA>          <NA>    NA  NA  NA      NA
#> 95     7    3               <NA>          <NA>    NA  NA  NA      NA
#> 96     7    4               <NA>          <NA>    NA  NA  NA      NA
#> 97     7    5            overall            <=  61.5  98  99     103
#> 98     7    6               <NA>          <NA>    NA  NA  NA      NA
#> 99     7    7            overall            <=  70.5 100 101     102
#> 100    7    8               <NA>          <NA>    NA  NA  NA      NA
#> 101    7    9               <NA>          <NA>    NA  NA  NA      NA
#> 102    7   10               <NA>          <NA>    NA  NA  NA      NA
#> 103    7   11               <NA>          <NA>    NA  NA  NA      NA
#> 104    7   12               <NA>          <NA>    NA  NA  NA      NA
#> 105    8    0          potential            <=  74.5 106 113     117
#> 106    8    1            overall            <=  62.5 107 108     112
#> 107    8    2               <NA>          <NA>    NA  NA  NA      NA
#> 108    8    3                age            <=  30.5 109 110     111
#> 109    8    4               <NA>          <NA>    NA  NA  NA      NA
#> 110    8    5               <NA>          <NA>    NA  NA  NA      NA
#> 111    8    6               <NA>          <NA>    NA  NA  NA      NA
#> 112    8    7               <NA>          <NA>    NA  NA  NA      NA
#> 113    8    8            overall            <=  61.5 114 115     116
#> 114    8    9               <NA>          <NA>    NA  NA  NA      NA
#> 115    8   10               <NA>          <NA>    NA  NA  NA      NA
#> 116    8   11               <NA>          <NA>    NA  NA  NA      NA
#> 117    8   12               <NA>          <NA>    NA  NA  NA      NA
#> 118    9    0            overall            <=  71.5 119 129     130
#> 119    9    1            overall            <=  60.5 120 121     128
#> 120    9    2               <NA>          <NA>    NA  NA  NA      NA
#> 121    9    3                age            <=  29.5 122 126     127
#> 122    9    4           shooting            <=  54.5 123 124     125
#> 123    9    5               <NA>          <NA>    NA  NA  NA      NA
#> 124    9    6               <NA>          <NA>    NA  NA  NA      NA
#> 125    9    7               <NA>          <NA>    NA  NA  NA      NA
#> 126    9    8               <NA>          <NA>    NA  NA  NA      NA
#> 127    9    9               <NA>          <NA>    NA  NA  NA      NA
#> 128    9   10               <NA>          <NA>    NA  NA  NA      NA
#> 129    9   11               <NA>          <NA>    NA  NA  NA      NA
#> 130    9   12               <NA>          <NA>    NA  NA  NA      NA
#> 131   10    0            overall            <=  71.5 132 142     143
#> 132   10    1            overall            <=  63.5 133 134     141
#> 133   10    2               <NA>          <NA>    NA  NA  NA      NA
#> 134   10    3          defending            <=  57.5 135 136     140
#> 135   10    4               <NA>          <NA>    NA  NA  NA      NA
#> 136   10    5          potential            <=  77.5 137 138     139
#> 137   10    6               <NA>          <NA>    NA  NA  NA      NA
#> 138   10    7               <NA>          <NA>    NA  NA  NA      NA
#> 139   10    8               <NA>          <NA>    NA  NA  NA      NA
#> 140   10    9               <NA>          <NA>    NA  NA  NA      NA
#> 141   10   10               <NA>          <NA>    NA  NA  NA      NA
#> 142   10   11               <NA>          <NA>    NA  NA  NA      NA
#> 143   10   12               <NA>          <NA>    NA  NA  NA      NA
#> 144   11    0            overall            <=  72.5 145 155     156
#> 145   11    1            overall            <=  60.5 146 147     154
#> 146   11    2               <NA>          <NA>    NA  NA  NA      NA
#> 147   11    3                age            <=  28.5 148 152     153
#> 148   11    4           shooting            <=  61.5 149 150     151
#> 149   11    5               <NA>          <NA>    NA  NA  NA      NA
#> 150   11    6               <NA>          <NA>    NA  NA  NA      NA
#> 151   11    7               <NA>          <NA>    NA  NA  NA      NA
#> 152   11    8               <NA>          <NA>    NA  NA  NA      NA
#> 153   11    9               <NA>          <NA>    NA  NA  NA      NA
#> 154   11   10               <NA>          <NA>    NA  NA  NA      NA
#> 155   11   11               <NA>          <NA>    NA  NA  NA      NA
#> 156   11   12               <NA>          <NA>    NA  NA  NA      NA
#> 157   12    0            overall            <=  72.5 158 168     169
#> 158   12    1            overall            <=  60.5 159 160     167
#> 159   12    2               <NA>          <NA>    NA  NA  NA      NA
#> 160   12    3                age            <=  29.5 161 165     166
#> 161   12    4          potential            <=  65.5 162 163     164
#> 162   12    5               <NA>          <NA>    NA  NA  NA      NA
#> 163   12    6               <NA>          <NA>    NA  NA  NA      NA
#> 164   12    7               <NA>          <NA>    NA  NA  NA      NA
#> 165   12    8               <NA>          <NA>    NA  NA  NA      NA
#> 166   12    9               <NA>          <NA>    NA  NA  NA      NA
#> 167   12   10               <NA>          <NA>    NA  NA  NA      NA
#> 168   12   11               <NA>          <NA>    NA  NA  NA      NA
#> 169   12   12               <NA>          <NA>    NA  NA  NA      NA
#> 170   13    0          potential            <=  77.5 171 178     182
#> 171   13    1 skill_ball_control            <=  60.5 172 173     177
#> 172   13    2               <NA>          <NA>    NA  NA  NA      NA
#> 173   13    3          potential            <=  65.5 174 175     176
#> 174   13    4               <NA>          <NA>    NA  NA  NA      NA
#> 175   13    5               <NA>          <NA>    NA  NA  NA      NA
#> 176   13    6               <NA>          <NA>    NA  NA  NA      NA
#> 177   13    7               <NA>          <NA>    NA  NA  NA      NA
#> 178   13    8            overall            <=  64.5 179 180     181
#> 179   13    9               <NA>          <NA>    NA  NA  NA      NA
#> 180   13   10               <NA>          <NA>    NA  NA  NA      NA
#> 181   13   11               <NA>          <NA>    NA  NA  NA      NA
#> 182   13   12               <NA>          <NA>    NA  NA  NA      NA
#> 183   14    0            overall            <=  73.5 184 194     195
#> 184   14    1            overall            <=  59.5 185 186     193
#> 185   14    2               <NA>          <NA>    NA  NA  NA      NA
#> 186   14    3          potential            <=  65.5 187 188     192
#> 187   14    4               <NA>          <NA>    NA  NA  NA      NA
#> 188   14    5          defending            <=  63.5 189 190     191
#> 189   14    6               <NA>          <NA>    NA  NA  NA      NA
#> 190   14    7               <NA>          <NA>    NA  NA  NA      NA
#> 191   14    8               <NA>          <NA>    NA  NA  NA      NA
#> 192   14    9               <NA>          <NA>    NA  NA  NA      NA
#> 193   14   10               <NA>          <NA>    NA  NA  NA      NA
#> 194   14   11               <NA>          <NA>    NA  NA  NA      NA
#> 195   14   12               <NA>          <NA>    NA  NA  NA      NA
#> 196   15    0            overall            <=  73.5 197 207     208
#> 197   15    1            overall            <=  59.5 198 199     206
#> 198   15    2               <NA>          <NA>    NA  NA  NA      NA
#> 199   15    3                age            <=  30.5 200 204     205
#> 200   15    4          potential            <=  64.5 201 202     203
#> 201   15    5               <NA>          <NA>    NA  NA  NA      NA
#> 202   15    6               <NA>          <NA>    NA  NA  NA      NA
#> 203   15    7               <NA>          <NA>    NA  NA  NA      NA
#> 204   15    8               <NA>          <NA>    NA  NA  NA      NA
#> 205   15    9               <NA>          <NA>    NA  NA  NA      NA
#> 206   15   10               <NA>          <NA>    NA  NA  NA      NA
#> 207   15   11               <NA>          <NA>    NA  NA  NA      NA
#> 208   15   12               <NA>          <NA>    NA  NA  NA      NA
#> 209   16    0            overall            <=  73.5 210 220     221
#> 210   16    1            overall            <=  59.5 211 212     219
#> 211   16    2               <NA>          <NA>    NA  NA  NA      NA
#> 212   16    3          potential            <=  65.5 213 214     218
#> 213   16    4               <NA>          <NA>    NA  NA  NA      NA
#> 214   16    5          defending            <=  64.5 215 216     217
#> 215   16    6               <NA>          <NA>    NA  NA  NA      NA
#> 216   16    7               <NA>          <NA>    NA  NA  NA      NA
#> 217   16    8               <NA>          <NA>    NA  NA  NA      NA
#> 218   16    9               <NA>          <NA>    NA  NA  NA      NA
#> 219   16   10               <NA>          <NA>    NA  NA  NA      NA
#> 220   16   11               <NA>          <NA>    NA  NA  NA      NA
#> 221   16   12               <NA>          <NA>    NA  NA  NA      NA
#> 222   17    0            overall            <=  74.5 223 233     234
#> 223   17    1            overall            <=  59.5 224 225     232
#> 224   17    2               <NA>          <NA>    NA  NA  NA      NA
#> 225   17    3                age            <=  30.5 226 230     231
#> 226   17    4          potential            <=  64.5 227 228     229
#> 227   17    5               <NA>          <NA>    NA  NA  NA      NA
#> 228   17    6               <NA>          <NA>    NA  NA  NA      NA
#> 229   17    7               <NA>          <NA>    NA  NA  NA      NA
#> 230   17    8               <NA>          <NA>    NA  NA  NA      NA
#> 231   17    9               <NA>          <NA>    NA  NA  NA      NA
#> 232   17   10               <NA>          <NA>    NA  NA  NA      NA
#> 233   17   11               <NA>          <NA>    NA  NA  NA      NA
#> 234   17   12               <NA>          <NA>    NA  NA  NA      NA
#> 235   18    0            overall            <=  73.5 236 246     247
#> 236   18    1            overall            <=  59.5 237 238     245
#> 237   18    2               <NA>          <NA>    NA  NA  NA      NA
#> 238   18    3                age            <=  31.5 239 243     244
#> 239   18    4          potential            <=  64.5 240 241     242
#> 240   18    5               <NA>          <NA>    NA  NA  NA      NA
#> 241   18    6               <NA>          <NA>    NA  NA  NA      NA
#> 242   18    7               <NA>          <NA>    NA  NA  NA      NA
#> 243   18    8               <NA>          <NA>    NA  NA  NA      NA
#> 244   18    9               <NA>          <NA>    NA  NA  NA      NA
#> 245   18   10               <NA>          <NA>    NA  NA  NA      NA
#> 246   18   11               <NA>          <NA>    NA  NA  NA      NA
#> 247   18   12               <NA>          <NA>    NA  NA  NA      NA
#> 248   19    0            overall            <=  73.5 249 259     260
#> 249   19    1            overall            <=  58.5 250 251     258
#> 250   19    2               <NA>          <NA>    NA  NA  NA      NA
#> 251   19    3          potential            <=  64.5 252 253     257
#> 252   19    4               <NA>          <NA>    NA  NA  NA      NA
#> 253   19    5          defending            <=  65.5 254 255     256
#> 254   19    6               <NA>          <NA>    NA  NA  NA      NA
#> 255   19    7               <NA>          <NA>    NA  NA  NA      NA
#> 256   19    8               <NA>          <NA>    NA  NA  NA      NA
#> 257   19    9               <NA>          <NA>    NA  NA  NA      NA
#> 258   19   10               <NA>          <NA>    NA  NA  NA      NA
#> 259   19   11               <NA>          <NA>    NA  NA  NA      NA
#> 260   19   12               <NA>          <NA>    NA  NA  NA      NA
#>      Prediction Cover
#> 1            NA   501
#> 2            NA     0
#> 3    -7000.0000     0
#> 4            NA     0
#> 5     7500.0000     0
#> 6    45000.0000     0
#> 7    25611.9110     0
#> 8     -997.7114     0
#> 9            NA   501
#> 10   28750.0000     0
#> 11  205000.0000   501
#> 12  177438.8532     0
#> 13   96391.6730     0
#> 14           NA   501
#> 15           NA     0
#> 16    -800.0000     0
#> 17           NA     0
#> 18   12750.0000     0
#> 19   48000.0000     0
#> 20   28022.3416     0
#> 21    4501.4877     0
#> 22           NA   501
#> 23   33125.0000     0
#> 24  248000.0000   501
#> 25  211402.3194     0
#> 26  102597.7022     0
#> 27           NA   501
#> 28           NA     0
#> 29   -3220.0000     0
#> 30           NA     0
#> 31   42700.0000     0
#> 32   25280.0000     0
#> 33   31308.7796     0
#> 34   11135.3957     0
#> 35           NA   501
#> 36  236700.0000   435
#> 37   26700.0000    66
#> 38  213560.2541     0
#> 39   96566.2611     0
#> 40           NA   501
#> 41           NA     0
#> 42           NA     0
#> 43    1602.0000     0
#> 44   26530.0000     0
#> 45   11726.4713     0
#> 46           NA     0
#> 47   11297.0000     0
#> 48   45930.0000     0
#> 49   37907.6667     0
#> 50   16140.5734     0
#> 51  317530.0000   501
#> 52  109799.1763     0
#> 53           NA   501
#> 54           NA     0
#> 55           NA     0
#> 56    1941.8000     0
#> 57           NA     0
#> 58   36777.0000     0
#> 59   24434.4000     0
#> 60   12099.0000     0
#> 61   14641.4387     0
#> 62   43954.0000     0
#> 63   18482.8120     0
#> 64  288777.0000   501
#> 65  102360.0761     0
#> 66           NA   501
#> 67           NA     0
#> 68   12247.6200     0
#> 69   30796.9600     0
#> 70    -752.3800     0
#> 71           NA   501
#> 72   18247.6200     0
#> 73           NA   501
#> 74   40747.6200     0
#> 75  343899.3000   501
#> 76  233752.5357     0
#> 77  200117.1000     0
#> 78  102894.4854     0
#> 79           NA   501
#> 80           NA     0
#> 81    5522.8580     0
#> 82           NA     0
#> 83           NA     0
#> 84   44196.5380     0
#> 85   38258.7980     0
#> 86   27065.1380     0
#> 87   23319.6040     0
#> 88   29997.5936     0
#> 89   20700.9040     0
#> 90  322009.3700   501
#> 91  100157.4532     0
#> 92           NA   501
#> 93           NA     0
#> 94   14091.6376     0
#> 95   31009.6436     0
#> 96    5770.5722     0
#> 97           NA   501
#> 98   17870.5722     0
#> 99           NA   501
#> 100  41431.5522     0
#> 101 443308.4330   501
#> 102 271979.4598     0
#> 103 244448.1922     0
#> 104 111960.9343     0
#> 105          NA   501
#> 106          NA     0
#> 107   9561.4084     0
#> 108          NA     0
#> 109  34886.6792     0
#> 110   8908.6792     0
#> 111  28593.0075     0
#> 112  21837.6417     0
#> 113          NA   501
#> 114  21607.7950     0
#> 115 193477.5897   501
#> 116 178907.0984     0
#> 117  69479.2364     0
#> 118          NA   501
#> 119          NA     0
#> 120   9105.2676     0
#> 121          NA     0
#> 122          NA     0
#> 123  30605.2676     0
#> 124  37070.6632     0
#> 125  23237.3741     0
#> 126  16164.4525     0
#> 127  29356.4014     0
#> 128  24045.1855     0
#> 129 344559.3763   501
#> 130  93696.2614     0
#> 131          NA   501
#> 132          NA     0
#> 133  17898.2013     0
#> 134          NA     0
#> 135  37995.8790     0
#> 136          NA     0
#> 137  25905.2424     0
#> 138  52989.5173     0
#> 139  28258.0718     0
#> 140  14580.5072     0
#> 141  25192.9534     0
#> 142 305673.8931   501
#> 143  86973.0313     0
#> 144          NA   501
#> 145          NA     0
#> 146  13404.9207     0
#> 147          NA     0
#> 148          NA     0
#> 149  33849.9549     0
#> 150  41870.6577     0
#> 151  25123.8166     0
#> 152  20165.9086     0
#> 153  30908.1085     0
#> 154  26670.5683     0
#> 155 378606.5038   501
#> 156  89748.7018     0
#> 157          NA   501
#> 158          NA     0
#> 159  15564.4286     0
#> 160          NA     0
#> 161          NA     0
#> 162  26111.4349     0
#> 163  35762.5500     0
#> 164  34818.8854     0
#> 165  18468.0097     0
#> 166  31333.7101     0
#> 167  27560.2415     0
#> 168 394245.8534   501
#> 169  92599.8932     0
#> 170          NA   501
#> 171          NA     0
#> 172  21221.2446     0
#> 173          NA     0
#> 174  26976.6741     0
#> 175  34883.1769     0
#> 176  34014.3517     0
#> 177  28139.2747     0
#> 178          NA   501
#> 179  28849.7337     0
#> 180 408321.2681   501
#> 181 361097.0784     0
#> 182  85812.1277     0
#> 183          NA   501
#> 184          NA     0
#> 185  16885.8613     0
#> 186          NA     0
#> 187  26617.1807     0
#> 188          NA     0
#> 189  36671.5458     0
#> 190  24938.8263     0
#> 191  20072.8798     0
#> 192  30876.6457     0
#> 193  28151.7069     0
#> 194 420989.1413   501
#> 195  87513.6170     0
#> 196          NA   501
#> 197          NA     0
#> 198  18697.2752     0
#> 199          NA     0
#> 200          NA     0
#> 201  27131.1780     0
#> 202  34692.1895     0
#> 203  34090.7777     0
#> 204  15276.4668     0
#> 205  31198.5714     0
#> 206  28772.7075     0
#> 207 382390.2271   501
#> 208  80699.0354     0
#> 209          NA   501
#> 210          NA     0
#> 211  20327.5477     0
#> 212          NA     0
#> 213  28474.1126     0
#> 214          NA     0
#> 215  36325.3451     0
#> 216  25067.6004     0
#> 217  23200.3538     0
#> 218  31706.6608     0
#> 219  29547.0143     0
#> 220 347651.2044   501
#> 221  76954.5979     0
#> 222          NA   501
#> 223          NA     0
#> 224  21811.6963     0
#> 225          NA     0
#> 226          NA     0
#> 227  28655.9198     0
#> 228  35604.3505     0
#> 229  35113.5597     0
#> 230  14904.2807     0
#> 231  31688.6735     0
#> 232  29910.9404     0
#> 233 426886.0840   501
#> 234  77431.5452     0
#> 235          NA   501
#> 236          NA     0
#> 237  23113.6233     0
#> 238          NA     0
#> 239          NA     0
#> 240  29510.4604     0
#> 241  34627.4253     0
#> 242  34193.4457     0
#> 243  16077.2361     0
#> 244  32068.3022     0
#> 245  30434.4459     0
#> 246 277197.4756   501
#> 247  66318.9045     0
#> 248          NA   501
#> 249          NA     0
#> 250  22831.3929     0
#> 251          NA     0
#> 252  28674.6008     0
#> 253          NA     0
#> 254  35642.0099     0
#> 255  24414.1939     0
#> 256  27577.4203     0
#> 257  32245.5031     0
#> 258  30798.2899     0
#> 259 252977.7280   501
#> 260  62888.9845     0
# }