Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node   Feature Decision.type Split Yes  No Missing  Prediction Cover
#> 1      0    0   overall            <=  65.5   2   9      13          NA   501
#> 2      0    1 potential            <=  73.5   3   4       8          NA     0
#> 3      0    2      <NA>          <NA>    NA  NA  NA      NA  -7000.0000     0
#> 4      0    3   overall            <=  62.5   5   6       7          NA     0
#> 5      0    4      <NA>          <NA>    NA  NA  NA      NA   7500.0000     0
#> 6      0    5      <NA>          <NA>    NA  NA  NA      NA  45000.0000     0
#> 7      0    6      <NA>          <NA>    NA  NA  NA      NA  25488.2045     0
#> 8      0    7      <NA>          <NA>    NA  NA  NA      NA  -1161.9171     0
#> 9      0    8 potential            <=  68.5  10  11      12          NA   501
#> 10     0    9      <NA>          <NA>    NA  NA  NA      NA  27500.0000     0
#> 11     0   10      <NA>          <NA>    NA  NA  NA      NA 205000.0000   501
#> 12     0   11      <NA>          <NA>    NA  NA  NA      NA 178663.3967     0
#> 13     0   12      <NA>          <NA>    NA  NA  NA      NA  95116.1506     0
#> 14     1    0   overall            <=  66.5  15  22      26          NA   501
#> 15     1    1 potential            <=  71.5  16  17      21          NA     0
#> 16     1    2      <NA>          <NA>    NA  NA  NA      NA  -1800.0000     0
#> 17     1    3   overall            <=  63.5  18  19      20          NA     0
#> 18     1    4      <NA>          <NA>    NA  NA  NA      NA  11700.0000     0
#> 19     1    5      <NA>          <NA>    NA  NA  NA      NA  43000.0000     0
#> 20     1    6      <NA>          <NA>    NA  NA  NA      NA  24402.1091     0
#> 21     1    7      <NA>          <NA>    NA  NA  NA      NA   5757.7722     0
#> 22     1    8 potential            <=  69.5  23  24      25          NA   501
#> 23     1    9      <NA>          <NA>    NA  NA  NA      NA  35500.0000     0
#> 24     1   10      <NA>          <NA>    NA  NA  NA      NA 238000.0000   501
#> 25     1   11      <NA>          <NA>    NA  NA  NA      NA 203864.0210     0
#> 26     1   12      <NA>          <NA>    NA  NA  NA      NA 100529.6860     0
#> 27     2    0   overall            <=  67.5  28  35      39          NA   501
#> 28     2    1   overall            <=  63.5  29  30      34          NA     0
#> 29     2    2      <NA>          <NA>    NA  NA  NA      NA  -3120.0000     0
#> 30     2    3       age            <=  24.5  31  32      33          NA     0
#> 31     2    4      <NA>          <NA>    NA  NA  NA      NA  38200.0000     0
#> 32     2    5      <NA>          <NA>    NA  NA  NA      NA  22880.0000     0
#> 33     2    6      <NA>          <NA>    NA  NA  NA      NA  29594.6642     0
#> 34     2    7      <NA>          <NA>    NA  NA  NA      NA  10520.3935     0
#> 35     2    8       age            <=  32.5  36  37      38          NA   501
#> 36     2    9      <NA>          <NA>    NA  NA  NA      NA 227700.0000   435
#> 37     2   10      <NA>          <NA>    NA  NA  NA      NA  27700.0000    66
#> 38     2   11      <NA>          <NA>    NA  NA  NA      NA 207663.6647     0
#> 39     2   12      <NA>          <NA>    NA  NA  NA      NA  93635.9449     0
#> 40     3    0   overall            <=  69.5  41  51      52          NA   501
#> 41     3    1 potential            <=  74.5  42  46      50          NA     0
#> 42     3    2   overall            <=  64.5  43  44      45          NA     0
#> 43     3    3      <NA>          <NA>    NA  NA  NA      NA   1692.0000     0
#> 44     3    4      <NA>          <NA>    NA  NA  NA      NA  26880.0000     0
#> 45     3    5      <NA>          <NA>    NA  NA  NA      NA  11809.4360     0
#> 46     3    6   overall            <=  61.5  47  48      49          NA     0
#> 47     3    7      <NA>          <NA>    NA  NA  NA      NA   8892.0000     0
#> 48     3    8      <NA>          <NA>    NA  NA  NA      NA  47380.0000     0
#> 49     3    9      <NA>          <NA>    NA  NA  NA      NA  39100.1650     0
#> 50     3   10      <NA>          <NA>    NA  NA  NA      NA  16243.1995     0
#> 51     3   11      <NA>          <NA>    NA  NA  NA      NA 318430.0000   501
#> 52     3   12      <NA>          <NA>    NA  NA  NA      NA 111240.7570     0
#> 53     4    0   overall            <=  69.5  54  64      65          NA   501
#> 54     4    1 potential            <=  74.5  55  59      63          NA     0
#> 55     4    2   overall            <=  63.5  56  57      58          NA     0
#> 56     4    3      <NA>          <NA>    NA  NA  NA      NA   1022.8000     0
#> 57     4    4      <NA>          <NA>    NA  NA  NA      NA  27422.8000     0
#> 58     4    5      <NA>          <NA>    NA  NA  NA      NA  13734.4648     0
#> 59     4    6   overall            <=  61.5  60  61      62          NA     0
#> 60     4    7      <NA>          <NA>    NA  NA  NA      NA  16502.8000     0
#> 61     4    8      <NA>          <NA>    NA  NA  NA      NA  46142.0000     0
#> 62     4    9      <NA>          <NA>    NA  NA  NA      NA  39549.0853     0
#> 63     4   10      <NA>          <NA>    NA  NA  NA      NA  17919.1850     0
#> 64     4   11      <NA>          <NA>    NA  NA  NA      NA 290087.0000   501
#> 65     4   12      <NA>          <NA>    NA  NA  NA      NA 101782.2518     0
#> 66     5    0 potential            <=  71.5  67  71      78          NA   501
#> 67     5    1  shooting            <=  59.5  68  69      70          NA     0
#> 68     5    2      <NA>          <NA>    NA  NA  NA      NA  12420.5200     0
#> 69     5    3      <NA>          <NA>    NA  NA  NA      NA  30661.7200     0
#> 70     5    4      <NA>          <NA>    NA  NA  NA      NA  -1579.4800     0
#> 71     5    5   overall            <=  62.5  72  73      77          NA   501
#> 72     5    6      <NA>          <NA>    NA  NA  NA      NA  20620.5200     0
#> 73     5    7 potential            <=  75.5  74  75      76          NA   501
#> 74     5    8      <NA>          <NA>    NA  NA  NA      NA  46949.7200     0
#> 75     5    9      <NA>          <NA>    NA  NA  NA      NA 404578.3000   501
#> 76     5   10      <NA>          <NA>    NA  NA  NA      NA 242575.6099     0
#> 77     5   11      <NA>          <NA>    NA  NA  NA      NA 210606.7698     0
#> 78     5   12      <NA>          <NA>    NA  NA  NA      NA 108492.1130     0
#> 79     6    0   overall            <=  69.5  80  87      91          NA   501
#> 80     6    1 dribbling            <=  63.5  81  85      86          NA     0
#> 81     6    2   overall            <=  62.5  82  83      84          NA     0
#> 82     6    3      <NA>          <NA>    NA  NA  NA      NA   5678.4680     0
#> 83     6    4      <NA>          <NA>    NA  NA  NA      NA  26969.6680     0
#> 84     6    5      <NA>          <NA>    NA  NA  NA      NA  15230.0052     0
#> 85     6    6      <NA>          <NA>    NA  NA  NA      NA  30433.5480     0
#> 86     6    7      <NA>          <NA>    NA  NA  NA      NA   4078.4680     0
#> 87     6    8 potential            <=  71.5  88  89      90          NA   501
#> 88     6    9      <NA>          <NA>    NA  NA  NA      NA  16836.2480     0
#> 89     6   10      <NA>          <NA>    NA  NA  NA      NA 317620.4700   501
#> 90     6   11      <NA>          <NA>    NA  NA  NA      NA 274911.9005     0
#> 91     6   12      <NA>          <NA>    NA  NA  NA      NA  97575.0008     0
#> 92     7    0   overall            <=  70.5  93 100     104          NA   501
#> 93     7    1 dribbling            <=  63.5  94  95      99          NA     0
#> 94     7    2      <NA>          <NA>    NA  NA  NA      NA  16481.5012     0
#> 95     7    3       age            <=  29.5  96  97      98          NA     0
#> 96     7    4      <NA>          <NA>    NA  NA  NA      NA  33635.1132     0
#> 97     7    5      <NA>          <NA>    NA  NA  NA      NA  16828.5032     0
#> 98     7    6      <NA>          <NA>    NA  NA  NA      NA  31164.5763     0
#> 99     7    7      <NA>          <NA>    NA  NA  NA      NA   6411.8212     0
#> 100    7    8       age            <=  33.5 101 102     103          NA   501
#> 101    7    9      <NA>          <NA>    NA  NA  NA      NA 375121.2810   462
#> 102    7   10      <NA>          <NA>    NA  NA  NA      NA    974.8921    39
#> 103    7   11      <NA>          <NA>    NA  NA  NA      NA 348186.4561     0
#> 104    7   12      <NA>          <NA>    NA  NA  NA      NA 107226.2230     0
#> 105    8    0   overall            <=  71.5 106 116     117          NA   501
#> 106    8    1   overall            <=  61.5 107 108     115          NA     0
#> 107    8    2      <NA>          <NA>    NA  NA  NA      NA   8421.5205     0
#> 108    8    3       age            <=  29.5 109 113     114          NA     0
#> 109    8    4 defending            <=  54.5 110 111     112          NA     0
#> 110    8    5      <NA>          <NA>    NA  NA  NA      NA  38768.6819     0
#> 111    8    6      <NA>          <NA>    NA  NA  NA      NA  31459.6311     0
#> 112    8    7      <NA>          <NA>    NA  NA  NA      NA  23029.4391     0
#> 113    8    8      <NA>          <NA>    NA  NA  NA      NA  14789.4391     0
#> 114    8    9      <NA>          <NA>    NA  NA  NA      NA  29475.3952     0
#> 115    8   10      <NA>          <NA>    NA  NA  NA      NA  23107.4489     0
#> 116    8   11      <NA>          <NA>    NA  NA  NA      NA 391109.1529   501
#> 117    8   12      <NA>          <NA>    NA  NA  NA      NA 104608.2092     0
#> 118    9    0   overall            <=  70.5 119 126     130          NA   501
#> 119    9    1   overall            <=  60.5 120 121     125          NA     0
#> 120    9    2      <NA>          <NA>    NA  NA  NA      NA   9120.3190     0
#> 121    9    3 defending            <=  52.5 122 123     124          NA     0
#> 122    9    4      <NA>          <NA>    NA  NA  NA      NA  35456.4829     0
#> 123    9    5      <NA>          <NA>    NA  NA  NA      NA  28120.3190     0
#> 124    9    6      <NA>          <NA>    NA  NA  NA      NA  20081.0952     0
#> 125    9    7      <NA>          <NA>    NA  NA  NA      NA  24208.6435     0
#> 126    9    8 potential            <=  72.5 127 128     129          NA   501
#> 127    9    9      <NA>          <NA>    NA  NA  NA      NA  24003.9216     0
#> 128    9   10      <NA>          <NA>    NA  NA  NA      NA 319735.3796   501
#> 129    9   11      <NA>          <NA>    NA  NA  NA      NA 271590.1486     0
#> 130    9   12      <NA>          <NA>    NA  NA  NA      NA  88388.7013     0
#> 131   10    0   overall            <=  71.5 132 142     143          NA   501
#> 132   10    1   overall            <=  61.5 133 134     141          NA     0
#> 133   10    2      <NA>          <NA>    NA  NA  NA      NA  12820.2871     0
#> 134   10    3       age            <=  29.5 135 139     140          NA     0
#> 135   10    4 defending            <=  55.5 136 137     138          NA     0
#> 136   10    5      <NA>          <NA>    NA  NA  NA      NA  37869.9546     0
#> 137   10    6      <NA>          <NA>    NA  NA  NA      NA  31875.3561     0
#> 138   10    7      <NA>          <NA>    NA  NA  NA      NA  24484.7367     0
#> 139   10    8      <NA>          <NA>    NA  NA  NA      NA  17761.1590     0
#> 140   10    9      <NA>          <NA>    NA  NA  NA      NA  30177.4699     0
#> 141   10   10      <NA>          <NA>    NA  NA  NA      NA  24954.2210     0
#> 142   10   11      <NA>          <NA>    NA  NA  NA      NA 298676.4805   501
#> 143   10   12      <NA>          <NA>    NA  NA  NA      NA  85485.2076     0
#> 144   11    0   overall            <=  72.5 145 155     156          NA   501
#> 145   11    1   overall            <=  60.5 146 147     154          NA     0
#> 146   11    2      <NA>          <NA>    NA  NA  NA      NA  13926.2584     0
#> 147   11    3       age            <=  28.5 148 152     153          NA     0
#> 148   11    4  shooting            <=  63.5 149 150     151          NA     0
#> 149   11    5      <NA>          <NA>    NA  NA  NA      NA  33479.0205     0
#> 150   11    6      <NA>          <NA>    NA  NA  NA      NA  44548.2577     0
#> 151   11    7      <NA>          <NA>    NA  NA  NA      NA  26769.9120     0
#> 152   11    8      <NA>          <NA>    NA  NA  NA      NA  19805.8594     0
#> 153   11    9      <NA>          <NA>    NA  NA  NA      NA  30836.5261     0
#> 154   11   10      <NA>          <NA>    NA  NA  NA      NA  26711.1937     0
#> 155   11   11      <NA>          <NA>    NA  NA  NA      NA 414894.1936   501
#> 156   11   12      <NA>          <NA>    NA  NA  NA      NA  95224.0703     0
#> 157   12    0   overall            <=  72.5 158 168     169          NA   501
#> 158   12    1   overall            <=  59.5 159 160     167          NA     0
#> 159   12    2      <NA>          <NA>    NA  NA  NA      NA  14033.6326     0
#> 160   12    3 potential            <=  65.5 161 162     166          NA     0
#> 161   12    4      <NA>          <NA>    NA  NA  NA      NA  23214.1879     0
#> 162   12    5 defending            <=  55.5 163 164     165          NA     0
#> 163   12    6      <NA>          <NA>    NA  NA  NA      NA  38214.1879     0
#> 164   12    7      <NA>          <NA>    NA  NA  NA      NA  31219.6712     0
#> 165   12    8      <NA>          <NA>    NA  NA  NA      NA  25397.2808     0
#> 166   12    9      <NA>          <NA>    NA  NA  NA      NA  31865.0497     0
#> 167   12   10      <NA>          <NA>    NA  NA  NA      NA  28375.2037     0
#> 168   12   11      <NA>          <NA>    NA  NA  NA      NA 376904.7742   501
#> 169   12   12      <NA>          <NA>    NA  NA  NA      NA  92215.7225     0
#> 170   13    0 potential            <=  73.5 171 178     182          NA   501
#> 171   13    1   overall            <=  62.5 172 173     177          NA     0
#> 172   13    2      <NA>          <NA>    NA  NA  NA      NA  20130.2693     0
#> 173   13    3       age            <=  30.5 174 175     176          NA     0
#> 174   13    4      <NA>          <NA>    NA  NA  NA      NA  34977.9513     0
#> 175   13    5      <NA>          <NA>    NA  NA  NA      NA  11111.8181     0
#> 176   13    6      <NA>          <NA>    NA  NA  NA      NA  28848.9307     0
#> 177   13    7      <NA>          <NA>    NA  NA  NA      NA  25557.4476     0
#> 178   13    8   overall            <=  69.5 179 180     181          NA   501
#> 179   13    9      <NA>          <NA>    NA  NA  NA      NA  31324.2138     0
#> 180   13   10      <NA>          <NA>    NA  NA  NA      NA 242714.2968   501
#> 181   13   11      <NA>          <NA>    NA  NA  NA      NA 158484.4829     0
#> 182   13   12      <NA>          <NA>    NA  NA  NA      NA  72683.3470     0
#> 183   14    0   overall            <=  73.5 184 194     195          NA   501
#> 184   14    1   overall            <=  59.5 185 186     193          NA     0
#> 185   14    2      <NA>          <NA>    NA  NA  NA      NA  17617.2424     0
#> 186   14    3 potential            <=  65.5 187 188     192          NA     0
#> 187   14    4      <NA>          <NA>    NA  NA  NA      NA  25276.1646     0
#> 188   14    5 defending            <=  65.5 189 190     191          NA     0
#> 189   14    6      <NA>          <NA>    NA  NA  NA      NA  36233.9135     0
#> 190   14    7      <NA>          <NA>    NA  NA  NA      NA  19131.7245     0
#> 191   14    8      <NA>          <NA>    NA  NA  NA      NA  25327.3606     0
#> 192   14    9      <NA>          <NA>    NA  NA  NA      NA  30719.0330     0
#> 193   14   10      <NA>          <NA>    NA  NA  NA      NA  28268.7237     0
#> 194   14   11      <NA>          <NA>    NA  NA  NA      NA 421942.8671   501
#> 195   14   12      <NA>          <NA>    NA  NA  NA      NA  84741.7297     0
#> 196   15    0   overall            <=  73.5 197 207     208          NA   501
#> 197   15    1   overall            <=  59.5 198 199     206          NA     0
#> 198   15    2      <NA>          <NA>    NA  NA  NA      NA  19355.5181     0
#> 199   15    3       age            <=  30.5 200 204     205          NA     0
#> 200   15    4 potential            <=  65.5 201 202     203          NA     0
#> 201   15    5      <NA>          <NA>    NA  NA  NA      NA  27816.2942     0
#> 202   15    6      <NA>          <NA>    NA  NA  NA      NA  35271.5826     0
#> 203   15    7      <NA>          <NA>    NA  NA  NA      NA  34434.4927     0
#> 204   15    8      <NA>          <NA>    NA  NA  NA      NA  13759.8348     0
#> 205   15    9      <NA>          <NA>    NA  NA  NA      NA  31084.8397     0
#> 206   15   10      <NA>          <NA>    NA  NA  NA      NA  28853.0354     0
#> 207   15   11      <NA>          <NA>    NA  NA  NA      NA 383248.5804   501
#> 208   15   12      <NA>          <NA>    NA  NA  NA      NA  79381.2546     0
#> 209   16    0   overall            <=  74.5 210 220     221          NA   501
#> 210   16    1   overall            <=  59.5 211 212     219          NA     0
#> 211   16    2      <NA>          <NA>    NA  NA  NA      NA  20919.9663     0
#> 212   16    3       age            <=  30.5 213 217     218          NA     0
#> 213   16    4 potential            <=  65.5 214 215     216          NA     0
#> 214   16    5      <NA>          <NA>    NA  NA  NA      NA  28389.9410     0
#> 215   16    6      <NA>          <NA>    NA  NA  NA      NA  35244.4244     0
#> 216   16    7      <NA>          <NA>    NA  NA  NA      NA  34456.6387     0
#> 217   16    8      <NA>          <NA>    NA  NA  NA      NA  15251.8483     0
#> 218   16    9      <NA>          <NA>    NA  NA  NA      NA  31240.8176     0
#> 219   16   10      <NA>          <NA>    NA  NA  NA      NA  29343.1991     0
#> 220   16   11      <NA>          <NA>    NA  NA  NA      NA 448423.7224   501
#> 221   16   12      <NA>          <NA>    NA  NA  NA      NA  77308.8657     0
#> 222   17    0 potential            <=  78.5 223 233     234          NA   501
#> 223   17    1   overall            <=  59.5 224 225     232          NA     0
#> 224   17    2      <NA>          <NA>    NA  NA  NA      NA  22327.9697     0
#> 225   17    3       age            <=  31.5 226 230     231          NA     0
#> 226   17    4 potential            <=  64.5 227 228     229          NA     0
#> 227   17    5      <NA>          <NA>    NA  NA  NA      NA  27610.2351     0
#> 228   17    6      <NA>          <NA>    NA  NA  NA      NA  35059.4195     0
#> 229   17    7      <NA>          <NA>    NA  NA  NA      NA  34431.0621     0
#> 230   17    8      <NA>          <NA>    NA  NA  NA      NA   9825.5708     0
#> 231   17    9      <NA>          <NA>    NA  NA  NA      NA  31016.1736     0
#> 232   17   10      <NA>          <NA>    NA  NA  NA      NA  29416.4373     0
#> 233   17   11      <NA>          <NA>    NA  NA  NA      NA 257081.3501   501
#> 234   17   12      <NA>          <NA>    NA  NA  NA      NA  59957.7638     0
#> 235   18    0   overall            <=  73.5 236 246     247          NA   501
#> 236   18    1   overall            <=  58.5 237 238     245          NA     0
#> 237   18    2      <NA>          <NA>    NA  NA  NA      NA  22595.1727     0
#> 238   18    3 potential            <=  64.5 239 240     244          NA     0
#> 239   18    4      <NA>          <NA>    NA  NA  NA      NA  27557.1960     0
#> 240   18    5 defending            <=  65.5 241 242     243          NA     0
#> 241   18    6      <NA>          <NA>    NA  NA  NA      NA  35355.8682     0
#> 242   18    7      <NA>          <NA>    NA  NA  NA      NA  20389.2209     0
#> 243   18    8      <NA>          <NA>    NA  NA  NA      NA  28006.3189     0
#> 244   18    9      <NA>          <NA>    NA  NA  NA      NA  31287.0827     0
#> 245   18   10      <NA>          <NA>    NA  NA  NA      NA  29968.6469     0
#> 246   18   11      <NA>          <NA>    NA  NA  NA      NA 284873.2151   501
#> 247   18   12      <NA>          <NA>    NA  NA  NA      NA  66423.4309     0
#> 248   19    0   overall            <=  62.5 249 253     260          NA   501
#> 249   19    1 potential            <=  70.5 250 251     252          NA     0
#> 250   19    2      <NA>          <NA>    NA  NA  NA      NA  26413.2863     0
#> 251   19    3      <NA>          <NA>    NA  NA  NA      NA  34493.2213     0
#> 252   19    4      <NA>          <NA>    NA  NA  NA      NA  28893.6384     0
#> 253   19    5   overall            <=  74.5 254 258     259          NA   501
#> 254   19    6       age            <=  31.5 255 256     257          NA     0
#> 255   19    7      <NA>          <NA>    NA  NA  NA      NA  34735.4740     0
#> 256   19    8      <NA>          <NA>    NA  NA  NA      NA  13545.9046     0
#> 257   19    9      <NA>          <NA>    NA  NA  NA      NA  31982.6806     0
#> 258   19   10      <NA>          <NA>    NA  NA  NA      NA 409885.8936   501
#> 259   19   11      <NA>          <NA>    NA  NA  NA      NA  93228.4674     0
#> 260   19   12      <NA>          <NA>    NA  NA  NA      NA  75066.3207     0
# }