Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node            Feature Decision.type Split Yes  No Missing
#> 1      0    0            overall            <=  65.5   2   9      13
#> 2      0    1          potential            <=  73.5   3   4       8
#> 3      0    2               <NA>          <NA>    NA  NA  NA      NA
#> 4      0    3            overall            <=  62.5   5   6       7
#> 5      0    4               <NA>          <NA>    NA  NA  NA      NA
#> 6      0    5               <NA>          <NA>    NA  NA  NA      NA
#> 7      0    6               <NA>          <NA>    NA  NA  NA      NA
#> 8      0    7               <NA>          <NA>    NA  NA  NA      NA
#> 9      0    8          potential            <=  68.5  10  11      12
#> 10     0    9               <NA>          <NA>    NA  NA  NA      NA
#> 11     0   10               <NA>          <NA>    NA  NA  NA      NA
#> 12     0   11               <NA>          <NA>    NA  NA  NA      NA
#> 13     0   12               <NA>          <NA>    NA  NA  NA      NA
#> 14     1    0            overall            <=  66.5  15  22      26
#> 15     1    1            overall            <=  63.5  16  17      21
#> 16     1    2               <NA>          <NA>    NA  NA  NA      NA
#> 17     1    3          potential            <=  71.5  18  19      20
#> 18     1    4               <NA>          <NA>    NA  NA  NA      NA
#> 19     1    5               <NA>          <NA>    NA  NA  NA      NA
#> 20     1    6               <NA>          <NA>    NA  NA  NA      NA
#> 21     1    7               <NA>          <NA>    NA  NA  NA      NA
#> 22     1    8          potential            <=  69.5  23  24      25
#> 23     1    9               <NA>          <NA>    NA  NA  NA      NA
#> 24     1   10               <NA>          <NA>    NA  NA  NA      NA
#> 25     1   11               <NA>          <NA>    NA  NA  NA      NA
#> 26     1   12               <NA>          <NA>    NA  NA  NA      NA
#> 27     2    0            overall            <=  67.5  28  35      39
#> 28     2    1          potential            <=  74.5  29  30      34
#> 29     2    2               <NA>          <NA>    NA  NA  NA      NA
#> 30     2    3            overall            <=  62.5  31  32      33
#> 31     2    4               <NA>          <NA>    NA  NA  NA      NA
#> 32     2    5               <NA>          <NA>    NA  NA  NA      NA
#> 33     2    6               <NA>          <NA>    NA  NA  NA      NA
#> 34     2    7               <NA>          <NA>    NA  NA  NA      NA
#> 35     2    8                age            <=  32.5  36  37      38
#> 36     2    9               <NA>          <NA>    NA  NA  NA      NA
#> 37     2   10               <NA>          <NA>    NA  NA  NA      NA
#> 38     2   11               <NA>          <NA>    NA  NA  NA      NA
#> 39     2   12               <NA>          <NA>    NA  NA  NA      NA
#> 40     3    0            overall            <=  69.5  41  51      52
#> 41     3    1          potential            <=  75.5  42  46      50
#> 42     3    2            overall            <=  63.5  43  44      45
#> 43     3    3               <NA>          <NA>    NA  NA  NA      NA
#> 44     3    4               <NA>          <NA>    NA  NA  NA      NA
#> 45     3    5               <NA>          <NA>    NA  NA  NA      NA
#> 46     3    6            overall            <=  61.5  47  48      49
#> 47     3    7               <NA>          <NA>    NA  NA  NA      NA
#> 48     3    8               <NA>          <NA>    NA  NA  NA      NA
#> 49     3    9               <NA>          <NA>    NA  NA  NA      NA
#> 50     3   10               <NA>          <NA>    NA  NA  NA      NA
#> 51     3   11               <NA>          <NA>    NA  NA  NA      NA
#> 52     3   12               <NA>          <NA>    NA  NA  NA      NA
#> 53     4    0            overall            <=  69.5  54  64      65
#> 54     4    1          potential            <=  73.5  55  59      63
#> 55     4    2            overall            <=  64.5  56  57      58
#> 56     4    3               <NA>          <NA>    NA  NA  NA      NA
#> 57     4    4               <NA>          <NA>    NA  NA  NA      NA
#> 58     4    5               <NA>          <NA>    NA  NA  NA      NA
#> 59     4    6            overall            <=  61.5  60  61      62
#> 60     4    7               <NA>          <NA>    NA  NA  NA      NA
#> 61     4    8               <NA>          <NA>    NA  NA  NA      NA
#> 62     4    9               <NA>          <NA>    NA  NA  NA      NA
#> 63     4   10               <NA>          <NA>    NA  NA  NA      NA
#> 64     4   11               <NA>          <NA>    NA  NA  NA      NA
#> 65     4   12               <NA>          <NA>    NA  NA  NA      NA
#> 66     5    0          potential            <=  71.5  67  74      78
#> 67     5    1            overall            <=  64.5  68  69      73
#> 68     5    2               <NA>          <NA>    NA  NA  NA      NA
#> 69     5    3                age            <=  28.5  70  71      72
#> 70     5    4               <NA>          <NA>    NA  NA  NA      NA
#> 71     5    5               <NA>          <NA>    NA  NA  NA      NA
#> 72     5    6               <NA>          <NA>    NA  NA  NA      NA
#> 73     5    7               <NA>          <NA>    NA  NA  NA      NA
#> 74     5    8            overall            <=  62.5  75  76      77
#> 75     5    9               <NA>          <NA>    NA  NA  NA      NA
#> 76     5   10               <NA>          <NA>    NA  NA  NA      NA
#> 77     5   11               <NA>          <NA>    NA  NA  NA      NA
#> 78     5   12               <NA>          <NA>    NA  NA  NA      NA
#> 79     6    0            overall            <=  70.5  80  87      91
#> 80     6    1          dribbling            <=  63.5  81  82      86
#> 81     6    2               <NA>          <NA>    NA  NA  NA      NA
#> 82     6    3          potential            <=  77.5  83  84      85
#> 83     6    4               <NA>          <NA>    NA  NA  NA      NA
#> 84     6    5               <NA>          <NA>    NA  NA  NA      NA
#> 85     6    6               <NA>          <NA>    NA  NA  NA      NA
#> 86     6    7               <NA>          <NA>    NA  NA  NA      NA
#> 87     6    8                age            <=  33.5  88  89      90
#> 88     6    9               <NA>          <NA>    NA  NA  NA      NA
#> 89     6   10               <NA>          <NA>    NA  NA  NA      NA
#> 90     6   11               <NA>          <NA>    NA  NA  NA      NA
#> 91     6   12               <NA>          <NA>    NA  NA  NA      NA
#> 92     7    0            overall            <=  70.5  93 103     104
#> 93     7    1            overall            <=  63.5  94  95     102
#> 94     7    2               <NA>          <NA>    NA  NA  NA      NA
#> 95     7    3          potential            <=  77.5  96 100     101
#> 96     7    4          defending            <=  55.5  97  98      99
#> 97     7    5               <NA>          <NA>    NA  NA  NA      NA
#> 98     7    6               <NA>          <NA>    NA  NA  NA      NA
#> 99     7    7               <NA>          <NA>    NA  NA  NA      NA
#> 100    7    8               <NA>          <NA>    NA  NA  NA      NA
#> 101    7    9               <NA>          <NA>    NA  NA  NA      NA
#> 102    7   10               <NA>          <NA>    NA  NA  NA      NA
#> 103    7   11               <NA>          <NA>    NA  NA  NA      NA
#> 104    7   12               <NA>          <NA>    NA  NA  NA      NA
#> 105    8    0            overall            <=  71.5 106 116     117
#> 106    8    1 skill_ball_control            <=  61.5 107 108     115
#> 107    8    2               <NA>          <NA>    NA  NA  NA      NA
#> 108    8    3          potential            <=  77.5 109 113     114
#> 109    8    4                age            <=  29.5 110 111     112
#> 110    8    5               <NA>          <NA>    NA  NA  NA      NA
#> 111    8    6               <NA>          <NA>    NA  NA  NA      NA
#> 112    8    7               <NA>          <NA>    NA  NA  NA      NA
#> 113    8    8               <NA>          <NA>    NA  NA  NA      NA
#> 114    8    9               <NA>          <NA>    NA  NA  NA      NA
#> 115    8   10               <NA>          <NA>    NA  NA  NA      NA
#> 116    8   11               <NA>          <NA>    NA  NA  NA      NA
#> 117    8   12               <NA>          <NA>    NA  NA  NA      NA
#> 118    9    0          potential            <=  73.5 119 126     130
#> 119    9    1            overall            <=  61.5 120 121     125
#> 120    9    2               <NA>          <NA>    NA  NA  NA      NA
#> 121    9    3                age            <=  30.5 122 123     124
#> 122    9    4               <NA>          <NA>    NA  NA  NA      NA
#> 123    9    5               <NA>          <NA>    NA  NA  NA      NA
#> 124    9    6               <NA>          <NA>    NA  NA  NA      NA
#> 125    9    7               <NA>          <NA>    NA  NA  NA      NA
#> 126    9    8            overall            <=  68.5 127 128     129
#> 127    9    9               <NA>          <NA>    NA  NA  NA      NA
#> 128    9   10               <NA>          <NA>    NA  NA  NA      NA
#> 129    9   11               <NA>          <NA>    NA  NA  NA      NA
#> 130    9   12               <NA>          <NA>    NA  NA  NA      NA
#> 131   10    0            overall            <=  71.5 132 142     143
#> 132   10    1            overall            <=  60.5 133 134     141
#> 133   10    2               <NA>          <NA>    NA  NA  NA      NA
#> 134   10    3                age            <=  27.5 135 139     140
#> 135   10    4          defending            <=  50.5 136 137     138
#> 136   10    5               <NA>          <NA>    NA  NA  NA      NA
#> 137   10    6               <NA>          <NA>    NA  NA  NA      NA
#> 138   10    7               <NA>          <NA>    NA  NA  NA      NA
#> 139   10    8               <NA>          <NA>    NA  NA  NA      NA
#> 140   10    9               <NA>          <NA>    NA  NA  NA      NA
#> 141   10   10               <NA>          <NA>    NA  NA  NA      NA
#> 142   10   11               <NA>          <NA>    NA  NA  NA      NA
#> 143   10   12               <NA>          <NA>    NA  NA  NA      NA
#> 144   11    0            overall            <=  72.5 145 155     156
#> 145   11    1            overall            <=  60.5 146 147     154
#> 146   11    2               <NA>          <NA>    NA  NA  NA      NA
#> 147   11    3                age            <=  27.5 148 152     153
#> 148   11    4           shooting            <=  52.5 149 150     151
#> 149   11    5               <NA>          <NA>    NA  NA  NA      NA
#> 150   11    6               <NA>          <NA>    NA  NA  NA      NA
#> 151   11    7               <NA>          <NA>    NA  NA  NA      NA
#> 152   11    8               <NA>          <NA>    NA  NA  NA      NA
#> 153   11    9               <NA>          <NA>    NA  NA  NA      NA
#> 154   11   10               <NA>          <NA>    NA  NA  NA      NA
#> 155   11   11               <NA>          <NA>    NA  NA  NA      NA
#> 156   11   12               <NA>          <NA>    NA  NA  NA      NA
#> 157   12    0          potential            <=  75.5 158 165     169
#> 158   12    1           shooting            <=  58.5 159 163     164
#> 159   12    2            overall            <=  61.5 160 161     162
#> 160   12    3               <NA>          <NA>    NA  NA  NA      NA
#> 161   12    4               <NA>          <NA>    NA  NA  NA      NA
#> 162   12    5               <NA>          <NA>    NA  NA  NA      NA
#> 163   12    6               <NA>          <NA>    NA  NA  NA      NA
#> 164   12    7               <NA>          <NA>    NA  NA  NA      NA
#> 165   12    8            overall            <=  69.5 166 167     168
#> 166   12    9               <NA>          <NA>    NA  NA  NA      NA
#> 167   12   10               <NA>          <NA>    NA  NA  NA      NA
#> 168   12   11               <NA>          <NA>    NA  NA  NA      NA
#> 169   12   12               <NA>          <NA>    NA  NA  NA      NA
#> 170   13    0            overall            <=  72.5 171 181     182
#> 171   13    1            overall            <=  60.5 172 173     180
#> 172   13    2               <NA>          <NA>    NA  NA  NA      NA
#> 173   13    3                age            <=  28.5 174 178     179
#> 174   13    4          defending            <=  55.5 175 176     177
#> 175   13    5               <NA>          <NA>    NA  NA  NA      NA
#> 176   13    6               <NA>          <NA>    NA  NA  NA      NA
#> 177   13    7               <NA>          <NA>    NA  NA  NA      NA
#> 178   13    8               <NA>          <NA>    NA  NA  NA      NA
#> 179   13    9               <NA>          <NA>    NA  NA  NA      NA
#> 180   13   10               <NA>          <NA>    NA  NA  NA      NA
#> 181   13   11               <NA>          <NA>    NA  NA  NA      NA
#> 182   13   12               <NA>          <NA>    NA  NA  NA      NA
#> 183   14    0          potential            <=  76.5 184 194     195
#> 184   14    1            overall            <=  59.5 185 186     193
#> 185   14    2               <NA>          <NA>    NA  NA  NA      NA
#> 186   14    3          potential            <=  65.5 187 188     192
#> 187   14    4               <NA>          <NA>    NA  NA  NA      NA
#> 188   14    5          defending            <=  63.5 189 190     191
#> 189   14    6               <NA>          <NA>    NA  NA  NA      NA
#> 190   14    7               <NA>          <NA>    NA  NA  NA      NA
#> 191   14    8               <NA>          <NA>    NA  NA  NA      NA
#> 192   14    9               <NA>          <NA>    NA  NA  NA      NA
#> 193   14   10               <NA>          <NA>    NA  NA  NA      NA
#> 194   14   11               <NA>          <NA>    NA  NA  NA      NA
#> 195   14   12               <NA>          <NA>    NA  NA  NA      NA
#> 196   15    0            overall            <=  73.5 197 207     208
#> 197   15    1            overall            <=  59.5 198 199     206
#> 198   15    2               <NA>          <NA>    NA  NA  NA      NA
#> 199   15    3                age            <=  29.5 200 204     205
#> 200   15    4          potential            <=  65.5 201 202     203
#> 201   15    5               <NA>          <NA>    NA  NA  NA      NA
#> 202   15    6               <NA>          <NA>    NA  NA  NA      NA
#> 203   15    7               <NA>          <NA>    NA  NA  NA      NA
#> 204   15    8               <NA>          <NA>    NA  NA  NA      NA
#> 205   15    9               <NA>          <NA>    NA  NA  NA      NA
#> 206   15   10               <NA>          <NA>    NA  NA  NA      NA
#> 207   15   11               <NA>          <NA>    NA  NA  NA      NA
#> 208   15   12               <NA>          <NA>    NA  NA  NA      NA
#> 209   16    0            overall            <=  73.5 210 220     221
#> 210   16    1            overall            <=  59.5 211 212     219
#> 211   16    2               <NA>          <NA>    NA  NA  NA      NA
#> 212   16    3                age            <=  30.5 213 217     218
#> 213   16    4          defending            <=  54.5 214 215     216
#> 214   16    5               <NA>          <NA>    NA  NA  NA      NA
#> 215   16    6               <NA>          <NA>    NA  NA  NA      NA
#> 216   16    7               <NA>          <NA>    NA  NA  NA      NA
#> 217   16    8               <NA>          <NA>    NA  NA  NA      NA
#> 218   16    9               <NA>          <NA>    NA  NA  NA      NA
#> 219   16   10               <NA>          <NA>    NA  NA  NA      NA
#> 220   16   11               <NA>          <NA>    NA  NA  NA      NA
#> 221   16   12               <NA>          <NA>    NA  NA  NA      NA
#> 222   17    0            overall            <=  73.5 223 233     234
#> 223   17    1            overall            <=  59.5 224 225     232
#> 224   17    2               <NA>          <NA>    NA  NA  NA      NA
#> 225   17    3                age            <=  31.5 226 230     231
#> 226   17    4          potential            <=  64.5 227 228     229
#> 227   17    5               <NA>          <NA>    NA  NA  NA      NA
#> 228   17    6               <NA>          <NA>    NA  NA  NA      NA
#> 229   17    7               <NA>          <NA>    NA  NA  NA      NA
#> 230   17    8               <NA>          <NA>    NA  NA  NA      NA
#> 231   17    9               <NA>          <NA>    NA  NA  NA      NA
#> 232   17   10               <NA>          <NA>    NA  NA  NA      NA
#> 233   17   11               <NA>          <NA>    NA  NA  NA      NA
#> 234   17   12               <NA>          <NA>    NA  NA  NA      NA
#> 235   18    0            overall            <=  74.5 236 246     247
#> 236   18    1            overall            <=  59.5 237 238     245
#> 237   18    2               <NA>          <NA>    NA  NA  NA      NA
#> 238   18    3                age            <=  31.5 239 243     244
#> 239   18    4          potential            <=  64.5 240 241     242
#> 240   18    5               <NA>          <NA>    NA  NA  NA      NA
#> 241   18    6               <NA>          <NA>    NA  NA  NA      NA
#> 242   18    7               <NA>          <NA>    NA  NA  NA      NA
#> 243   18    8               <NA>          <NA>    NA  NA  NA      NA
#> 244   18    9               <NA>          <NA>    NA  NA  NA      NA
#> 245   18   10               <NA>          <NA>    NA  NA  NA      NA
#> 246   18   11               <NA>          <NA>    NA  NA  NA      NA
#> 247   18   12               <NA>          <NA>    NA  NA  NA      NA
#> 248   19    0          potential            <=  78.5 249 259     260
#> 249   19    1            overall            <=  58.5 250 251     258
#> 250   19    2               <NA>          <NA>    NA  NA  NA      NA
#> 251   19    3                age            <=  32.5 252 256     257
#> 252   19    4           shooting            <=  60.5 253 254     255
#> 253   19    5               <NA>          <NA>    NA  NA  NA      NA
#> 254   19    6               <NA>          <NA>    NA  NA  NA      NA
#> 255   19    7               <NA>          <NA>    NA  NA  NA      NA
#> 256   19    8               <NA>          <NA>    NA  NA  NA      NA
#> 257   19    9               <NA>          <NA>    NA  NA  NA      NA
#> 258   19   10               <NA>          <NA>    NA  NA  NA      NA
#> 259   19   11               <NA>          <NA>    NA  NA  NA      NA
#> 260   19   12               <NA>          <NA>    NA  NA  NA      NA
#>      Prediction Cover
#> 1            NA   501
#> 2            NA     0
#> 3    -7000.0000     0
#> 4            NA     0
#> 5    10000.0000     0
#> 6    45000.0000     0
#> 7    26834.8624     0
#> 8     -835.7211     0
#> 9            NA   501
#> 10   30000.0000     0
#> 11  205000.0000   501
#> 12  179691.9814     0
#> 13   96964.1099     0
#> 14           NA   501
#> 15           NA     0
#> 16   -6800.0000     0
#> 17           NA     0
#> 18   21700.0000     0
#> 19   43000.0000     0
#> 20   28688.6610     0
#> 21    5571.7830     0
#> 22           NA   501
#> 23   33000.0000     0
#> 24  248000.0000   501
#> 25  212681.7248     0
#> 26  104900.2517     0
#> 27           NA   501
#> 28           NA     0
#> 29    6830.0000     0
#> 30           NA     0
#> 31   14180.0000     0
#> 32   47200.0000     0
#> 33   35173.9424     0
#> 34   11073.3596     0
#> 35           NA   501
#> 36  236700.0000   435
#> 37   26700.0000    66
#> 38  213280.7470     0
#> 39   95195.5269     0
#> 40           NA   501
#> 41           NA     0
#> 42           NA     0
#> 43   -1803.0000     0
#> 44   26530.0000     0
#> 45   12317.2546     0
#> 46           NA     0
#> 47   11262.0000     0
#> 48   46960.0000     0
#> 49   39828.9916     0
#> 50   15920.4467     0
#> 51  326530.0000   501
#> 52  110880.8463     0
#> 53           NA   501
#> 54           NA     0
#> 55           NA     0
#> 56    3877.3000     0
#> 57   26694.0000     0
#> 58   12984.0052     0
#> 59           NA     0
#> 60   10677.3000     0
#> 61   42827.0000     0
#> 62   35587.1821     0
#> 63   17657.3262     0
#> 64  297377.0000   501
#> 65  103449.4528     0
#> 66           NA   501
#> 67           NA     0
#> 68    5989.5700     0
#> 69           NA     0
#> 70   32524.6000     0
#> 71   14707.6000     0
#> 72   23397.7956     0
#> 73   13824.1945     0
#> 74           NA   501
#> 75   17989.5700     0
#> 76  141139.3000   501
#> 77  122767.0528     0
#> 78   66561.1685     0
#> 79           NA   501
#> 80           NA     0
#> 81   15890.6130     0
#> 82           NA     0
#> 83   29690.6130     0
#> 84   53937.3700     0
#> 85   32691.3499     0
#> 86    4890.6130     0
#> 87           NA   501
#> 88  390525.3700   462
#> 89   11525.3700    39
#> 90  362723.5150     0
#> 91  109392.2241     0
#> 92           NA   501
#> 93           NA     0
#> 94   10801.5517     0
#> 95           NA     0
#> 96           NA     0
#> 97   34998.3087     0
#> 98   24468.3087     0
#> 99   13247.7787     0
#> 100  60993.6330     0
#> 101  30066.9891     0
#> 102  21384.1754     0
#> 103 304972.8330   501
#> 104  94771.5455     0
#> 105          NA   501
#> 106          NA     0
#> 107  14321.3965     0
#> 108          NA     0
#> 109          NA     0
#> 110  33014.4778     0
#> 111  13699.6478     0
#> 112  28985.7740     0
#> 113  59444.2697     0
#> 114  32678.2483     0
#> 115  23237.3600     0
#> 116 377975.5497   501
#> 117 102421.7245     0
#> 118          NA   501
#> 119          NA     0
#> 120   8389.2569     0
#> 121          NA     0
#> 122  32289.2569     0
#> 123   7239.2812     0
#> 124  26107.0513     0
#> 125  20625.4861     0
#> 126          NA   501
#> 127  31667.7569     0
#> 128 293677.9947   501
#> 129 200137.2913     0
#> 130  84777.5329     0
#> 131          NA   501
#> 132          NA     0
#> 133  10750.3312     0
#> 134          NA     0
#> 135          NA     0
#> 136  38030.2544     0
#> 137  32560.3312     0
#> 138  24643.3368     0
#> 139  22072.1321     0
#> 140  30039.8028     0
#> 141  25171.8574     0
#> 142 317810.1953   501
#> 143  90814.5528     0
#> 144          NA   501
#> 145          NA     0
#> 146  12475.2981     0
#> 147          NA     0
#> 148          NA     0
#> 149  32804.2981     0
#> 150  38434.9899     0
#> 151  24695.9975     0
#> 152  21808.1398     0
#> 153  30516.2767     0
#> 154  26231.0029     0
#> 155 439529.1757   501
#> 156 100442.8753     0
#> 157          NA   501
#> 158          NA     0
#> 159          NA     0
#> 160  16758.4601     0
#> 161  29398.7683     0
#> 162  24328.3271     0
#> 163  33999.7850     0
#> 164  14727.7683     0
#> 165          NA   501
#> 166  34934.1203     0
#> 167 449076.2582   501
#> 168 308091.7006     0
#> 169  97087.6009     0
#> 170          NA   501
#> 171          NA     0
#> 172  17251.9222     0
#> 173          NA     0
#> 174          NA     0
#> 175  38111.7366     0
#> 176  32149.5346     0
#> 177  26162.1731     0
#> 178  21237.4247     0
#> 179  30551.7355     0
#> 180  27300.6308     0
#> 181 349176.2797   501
#> 182  84357.0430     0
#> 183          NA   501
#> 184          NA     0
#> 185  17226.7300     0
#> 186          NA     0
#> 187  25965.3842     0
#> 188          NA     0
#> 189  36262.3068     0
#> 190  23053.8893     0
#> 191  24531.6879     0
#> 192  30211.7349     0
#> 193  27658.7264     0
#> 194 126251.0044   501
#> 195  48555.2405     0
#> 196          NA   501
#> 197          NA     0
#> 198  18004.0570     0
#> 199          NA     0
#> 200          NA     0
#> 201  28462.1706     0
#> 202  34369.5955     0
#> 203  33763.3986     0
#> 204  18189.0761     0
#> 205  30398.4536     0
#> 206  28028.5850     0
#> 207 376124.7737   501
#> 208  79905.9249     0
#> 209          NA   501
#> 210          NA     0
#> 211  19703.6513     0
#> 212          NA     0
#> 213          NA     0
#> 214  36207.1512     0
#> 215  31305.8477     0
#> 216  26604.5596     0
#> 217  15194.6232     0
#> 218  29978.6258     0
#> 219  28013.7484     0
#> 220 383013.4266   501
#> 221  80958.7709     0
#> 222          NA   501
#> 223          NA     0
#> 224  21536.3554     0
#> 225          NA     0
#> 226          NA     0
#> 227  28112.4219     0
#> 228  34037.5001     0
#> 229  33568.9031     0
#> 230  13581.6779     0
#> 231  31260.5194     0
#> 232  29494.5275     0
#> 233 348212.0839   501
#> 234  75912.4143     0
#> 235          NA   501
#> 236          NA     0
#> 237  23544.9966     0
#> 238          NA     0
#> 239          NA     0
#> 240  28768.7406     0
#> 241  34259.0747     0
#> 242  33816.9762     0
#> 243  13069.6400     0
#> 244  31331.5691     0
#> 245  29910.5702     0
#> 246 466890.8755   501
#> 247  80689.9864     0
#> 248          NA   501
#> 249          NA     0
#> 250  23028.2202     0
#> 251          NA     0
#> 252          NA     0
#> 253  32490.6577     0
#> 254  38982.3991     0
#> 255  28600.1842     0
#> 256   9754.5966     0
#> 257  31765.3752     0
#> 258  30467.4406     0
#> 259 266964.9681   501
#> 260  61236.1856     0
# }