Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node            Feature Decision.type Split Yes  No Missing Prediction
#> 1      0    0            overall            <=  65.5   2   9      13         NA
#> 2      0    1          potential            <=  74.5   3   4       8         NA
#> 3      0    2               <NA>          <NA>    NA  NA  NA      NA  -6000.000
#> 4      0    3            overall            <=  62.5   5   6       7         NA
#> 5      0    4               <NA>          <NA>    NA  NA  NA      NA   7500.000
#> 6      0    5               <NA>          <NA>    NA  NA  NA      NA  47500.000
#> 7      0    6               <NA>          <NA>    NA  NA  NA      NA  27465.812
#> 8      0    7               <NA>          <NA>    NA  NA  NA      NA  -1334.247
#> 9      0    8          potential            <=  68.5  10  11      12         NA
#> 10     0    9               <NA>          <NA>    NA  NA  NA      NA  27500.000
#> 11     0   10               <NA>          <NA>    NA  NA  NA      NA 205000.000
#> 12     0   11               <NA>          <NA>    NA  NA  NA      NA 178893.890
#> 13     0   12               <NA>          <NA>    NA  NA  NA      NA  96145.530
#> 14     1    0            overall            <=  66.5  15  22      26         NA
#> 15     1    1          potential            <=  71.5  16  17      21         NA
#> 16     1    2               <NA>          <NA>    NA  NA  NA      NA  -1900.000
#> 17     1    3            overall            <=  63.5  18  19      20         NA
#> 18     1    4               <NA>          <NA>    NA  NA  NA      NA  12750.000
#> 19     1    5               <NA>          <NA>    NA  NA  NA      NA  43750.000
#> 20     1    6               <NA>          <NA>    NA  NA  NA      NA  25626.571
#> 21     1    7               <NA>          <NA>    NA  NA  NA      NA   5912.765
#> 22     1    8          potential            <=  69.5  23  24      25         NA
#> 23     1    9               <NA>          <NA>    NA  NA  NA      NA  33250.000
#> 24     1   10               <NA>          <NA>    NA  NA  NA      NA 248000.000
#> 25     1   11               <NA>          <NA>    NA  NA  NA      NA 211062.214
#> 26     1   12               <NA>          <NA>    NA  NA  NA      NA 104054.070
#> 27     2    0            overall            <=  67.5  28  35      39         NA
#> 28     2    1            overall            <=  63.5  29  30      34         NA
#> 29     2    2               <NA>          <NA>    NA  NA  NA      NA  -3210.000
#> 30     2    3          potential            <=  73.5  31  32      33         NA
#> 31     2    4               <NA>          <NA>    NA  NA  NA      NA  25925.000
#> 32     2    5               <NA>          <NA>    NA  NA  NA      NA  46700.000
#> 33     2    6               <NA>          <NA>    NA  NA  NA      NA  31009.920
#> 34     2    7               <NA>          <NA>    NA  NA  NA      NA  11239.413
#> 35     2    8                age            <=  32.5  36  37      38         NA
#> 36     2    9               <NA>          <NA>    NA  NA  NA      NA 236700.000
#> 37     2   10               <NA>          <NA>    NA  NA  NA      NA  36700.000
#> 38     2   11               <NA>          <NA>    NA  NA  NA      NA 213862.268
#> 39     2   12               <NA>          <NA>    NA  NA  NA      NA  96088.703
#> 40     3    0            overall            <=  69.5  41  51      52         NA
#> 41     3    1          potential            <=  75.5  42  46      50         NA
#> 42     3    2            overall            <=  63.5  43  44      45         NA
#> 43     3    3               <NA>          <NA>    NA  NA  NA      NA  -1389.000
#> 44     3    4               <NA>          <NA>    NA  NA  NA      NA  26530.000
#> 45     3    5               <NA>          <NA>    NA  NA  NA      NA  12340.078
#> 46     3    6            overall            <=  61.5  47  48      49         NA
#> 47     3    7               <NA>          <NA>    NA  NA  NA      NA  16296.000
#> 48     3    8               <NA>          <NA>    NA  NA  NA      NA  49205.000
#> 49     3    9               <NA>          <NA>    NA  NA  NA      NA  42526.052
#> 50     3   10               <NA>          <NA>    NA  NA  NA      NA  16219.496
#> 51     3   11               <NA>          <NA>    NA  NA  NA      NA 326530.000
#> 52     3   12               <NA>          <NA>    NA  NA  NA      NA 111733.606
#> 53     4    0            overall            <=  69.5  54  64      65         NA
#> 54     4    1          potential            <=  74.5  55  59      63         NA
#> 55     4    2            overall            <=  64.5  56  57      58         NA
#> 56     4    3               <NA>          <NA>    NA  NA  NA      NA   5249.900
#> 57     4    4               <NA>          <NA>    NA  NA  NA      NA  27377.000
#> 58     4    5               <NA>          <NA>    NA  NA  NA      NA  14260.785
#> 59     4    6            overall            <=  62.5  60  61      62         NA
#> 60     4    7               <NA>          <NA>    NA  NA  NA      NA  18166.400
#> 61     4    8               <NA>          <NA>    NA  NA  NA      NA  45284.500
#> 62     4    9               <NA>          <NA>    NA  NA  NA      NA  37484.471
#> 63     4   10               <NA>          <NA>    NA  NA  NA      NA  18074.120
#> 64     4   11               <NA>          <NA>    NA  NA  NA      NA 287377.000
#> 65     4   12               <NA>          <NA>    NA  NA  NA      NA 100199.859
#> 66     5    0          potential            <=  71.5  67  74      78         NA
#> 67     5    1            overall            <=  63.5  68  69      73         NA
#> 68     5    2               <NA>          <NA>    NA  NA  NA      NA   3224.910
#> 69     5    3                age            <=  28.5  70  71      72         NA
#> 70     5    4               <NA>          <NA>    NA  NA  NA      NA  32519.510
#> 71     5    5               <NA>          <NA>    NA  NA  NA      NA  14614.300
#> 72     5    6               <NA>          <NA>    NA  NA  NA      NA  23852.860
#> 73     5    7               <NA>          <NA>    NA  NA  NA      NA  13908.570
#> 74     5    8            overall            <=  62.5  75  76      77         NA
#> 75     5    9               <NA>          <NA>    NA  NA  NA      NA  16759.910
#> 76     5   10               <NA>          <NA>    NA  NA  NA      NA 132139.300
#> 77     5   11               <NA>          <NA>    NA  NA  NA      NA 114926.281
#> 78     5   12               <NA>          <NA>    NA  NA  NA      NA  62809.144
#> 79     6    0            overall            <=  70.5  80  87      91         NA
#> 80     6    1          dribbling            <=  63.5  81  82      86         NA
#> 81     6    2               <NA>          <NA>    NA  NA  NA      NA  16083.919
#> 82     6    3          potential            <=  77.5  83  84      85         NA
#> 83     6    4               <NA>          <NA>    NA  NA  NA      NA  29862.349
#> 84     6    5               <NA>          <NA>    NA  NA  NA      NA  59542.120
#> 85     6    6               <NA>          <NA>    NA  NA  NA      NA  33504.417
#> 86     6    7               <NA>          <NA>    NA  NA  NA      NA   4367.120
#> 87     6    8                age            <=  32.5  88  89      90         NA
#> 88     6    9               <NA>          <NA>    NA  NA  NA      NA 442425.370
#> 89     6   10               <NA>          <NA>    NA  NA  NA      NA  28301.620
#> 90     6   11               <NA>          <NA>    NA  NA  NA      NA 393104.677
#> 91     6   12               <NA>          <NA>    NA  NA  NA      NA 120105.851
#> 92     7    0            overall            <=  70.5  93 100     104         NA
#> 93     7    1 skill_ball_control            <=  62.5  94  95      99         NA
#> 94     7    2               <NA>          <NA>    NA  NA  NA      NA  14294.027
#> 95     7    3          potential            <=  77.5  96  97      98         NA
#> 96     7    4               <NA>          <NA>    NA  NA  NA      NA  29968.614
#> 97     7    5               <NA>          <NA>    NA  NA  NA      NA  60912.908
#> 98     7    6               <NA>          <NA>    NA  NA  NA      NA  33632.632
#> 99     7    7               <NA>          <NA>    NA  NA  NA      NA  22314.368
#> 100    7    8          potential            <=  72.5 101 102     103         NA
#> 101    7    9               <NA>          <NA>    NA  NA  NA      NA  31682.833
#> 102    7   10               <NA>          <NA>    NA  NA  NA      NA 413095.208
#> 103    7   11               <NA>          <NA>    NA  NA  NA      NA 351129.494
#> 104    7   12               <NA>          <NA>    NA  NA  NA      NA 109348.375
#> 105    8    0            overall            <=  71.5 106 116     117         NA
#> 106    8    1            overall            <=  61.5 107 108     115         NA
#> 107    8    2               <NA>          <NA>    NA  NA  NA      NA   7536.304
#> 108    8    3          potential            <=  76.5 109 113     114         NA
#> 109    8    4          defending            <=  52.5 110 111     112         NA
#> 110    8    5               <NA>          <NA>    NA  NA  NA      NA  34919.323
#> 111    8    6               <NA>          <NA>    NA  NA  NA      NA  26367.274
#> 112    8    7               <NA>          <NA>    NA  NA  NA      NA  14691.965
#> 113    8    8               <NA>          <NA>    NA  NA  NA      NA  47345.074
#> 114    8    9               <NA>          <NA>    NA  NA  NA      NA  30542.369
#> 115    8   10               <NA>          <NA>    NA  NA  NA      NA  23569.162
#> 116    8   11               <NA>          <NA>    NA  NA  NA      NA 363873.312
#> 117    8   12               <NA>          <NA>    NA  NA  NA      NA  99866.701
#> 118    9    0            overall            <=  71.5 119 129     130         NA
#> 119    9    1            overall            <=  61.5 120 121     128         NA
#> 120    9    2               <NA>          <NA>    NA  NA  NA      NA  10282.674
#> 121    9    3                age            <=  29.5 122 126     127         NA
#> 122    9    4           shooting            <=  57.5 123 124     125         NA
#> 123    9    5               <NA>          <NA>    NA  NA  NA      NA  32436.413
#> 124    9    6               <NA>          <NA>    NA  NA  NA      NA  38805.233
#> 125    9    7               <NA>          <NA>    NA  NA  NA      NA  23567.108
#> 126    9    8               <NA>          <NA>    NA  NA  NA      NA  14630.322
#> 127    9    9               <NA>          <NA>    NA  NA  NA      NA  30025.121
#> 128    9   10               <NA>          <NA>    NA  NA  NA      NA  23893.857
#> 129    9   11               <NA>          <NA>    NA  NA  NA      NA 330985.981
#> 130    9   12               <NA>          <NA>    NA  NA  NA      NA  91434.635
#> 131   10    0          potential            <=  73.5 132 139     143         NA
#> 132   10    1           shooting            <=  59.5 133 134     138         NA
#> 133   10    2               <NA>          <NA>    NA  NA  NA      NA  20764.227
#> 134   10    3                age            <=  30.5 135 136     137         NA
#> 135   10    4               <NA>          <NA>    NA  NA  NA      NA  38349.174
#> 136   10    5               <NA>          <NA>    NA  NA  NA      NA  12231.336
#> 137   10    6               <NA>          <NA>    NA  NA  NA      NA  31947.030
#> 138   10    7               <NA>          <NA>    NA  NA  NA      NA  10754.406
#> 139   10    8            overall            <=  68.5 140 141     142         NA
#> 140   10    9               <NA>          <NA>    NA  NA  NA      NA  33441.972
#> 141   10   10               <NA>          <NA>    NA  NA  NA      NA 251387.383
#> 142   10   11               <NA>          <NA>    NA  NA  NA      NA 173355.083
#> 143   10   12               <NA>          <NA>    NA  NA  NA      NA  76517.282
#> 144   11    0            overall            <=  72.5 145 155     156         NA
#> 145   11    1            overall            <=  60.5 146 147     154         NA
#> 146   11    2               <NA>          <NA>    NA  NA  NA      NA  13006.304
#> 147   11    3                age            <=  29.5 148 152     153         NA
#> 148   11    4          potential            <=  65.5 149 150     151         NA
#> 149   11    5               <NA>          <NA>    NA  NA  NA      NA  26006.304
#> 150   11    6               <NA>          <NA>    NA  NA  NA      NA  35479.801
#> 151   11    7               <NA>          <NA>    NA  NA  NA      NA  34567.491
#> 152   11    8               <NA>          <NA>    NA  NA  NA      NA  17438.893
#> 153   11    9               <NA>          <NA>    NA  NA  NA      NA  30836.001
#> 154   11   10               <NA>          <NA>    NA  NA  NA      NA  26511.989
#> 155   11   11               <NA>          <NA>    NA  NA  NA      NA 379748.645
#> 156   11   12               <NA>          <NA>    NA  NA  NA      NA  90480.330
#> 157   12    0          potential            <=  75.5 158 165     169         NA
#> 158   12    1           shooting            <=  58.5 159 163     164         NA
#> 159   12    2            overall            <=  60.5 160 161     162         NA
#> 160   12    3               <NA>          <NA>    NA  NA  NA      NA  15205.674
#> 161   12    4               <NA>          <NA>    NA  NA  NA      NA  29460.594
#> 162   12    5               <NA>          <NA>    NA  NA  NA      NA  24682.767
#> 163   12    6               <NA>          <NA>    NA  NA  NA      NA  35260.625
#> 164   12    7               <NA>          <NA>    NA  NA  NA      NA  15035.197
#> 165   12    8            overall            <=  69.5 166 167     168         NA
#> 166   12    9               <NA>          <NA>    NA  NA  NA      NA  35709.249
#> 167   12   10               <NA>          <NA>    NA  NA  NA      NA 445273.780
#> 168   12   11               <NA>          <NA>    NA  NA  NA      NA 304474.833
#> 169   12   12               <NA>          <NA>    NA  NA  NA      NA  96490.482
#> 170   13    0            overall            <=  73.5 171 181     182         NA
#> 171   13    1            overall            <=  60.5 172 173     180         NA
#> 172   13    2               <NA>          <NA>    NA  NA  NA      NA  17185.106
#> 173   13    3                age            <=  30.5 174 178     179         NA
#> 174   13    4          potential            <=  65.5 175 176     177         NA
#> 175   13    5               <NA>          <NA>    NA  NA  NA      NA  27793.685
#> 176   13    6               <NA>          <NA>    NA  NA  NA      NA  35586.598
#> 177   13    7               <NA>          <NA>    NA  NA  NA      NA  34774.934
#> 178   13    8               <NA>          <NA>    NA  NA  NA      NA  11941.444
#> 179   13    9               <NA>          <NA>    NA  NA  NA      NA  30949.008
#> 180   13   10               <NA>          <NA>    NA  NA  NA      NA  27771.091
#> 181   13   11               <NA>          <NA>    NA  NA  NA      NA 415658.777
#> 182   13   12               <NA>          <NA>    NA  NA  NA      NA  85323.963
#> 183   14    0            overall            <=  72.5 184 194     195         NA
#> 184   14    1            overall            <=  59.5 185 186     193         NA
#> 185   14    2               <NA>          <NA>    NA  NA  NA      NA  16966.596
#> 186   14    3          defending            <=  54.5 187 191     192         NA
#> 187   14    4          potential            <=  65.5 188 189     190         NA
#> 188   14    5               <NA>          <NA>    NA  NA  NA      NA  28613.624
#> 189   14    6               <NA>          <NA>    NA  NA  NA      NA  38208.472
#> 190   14    7               <NA>          <NA>    NA  NA  NA      NA  36896.320
#> 191   14    8               <NA>          <NA>    NA  NA  NA      NA  29840.105
#> 192   14    9               <NA>          <NA>    NA  NA  NA      NA  24663.241
#> 193   14   10               <NA>          <NA>    NA  NA  NA      NA  29088.241
#> 194   14   11               <NA>          <NA>    NA  NA  NA      NA 277592.899
#> 195   14   12               <NA>          <NA>    NA  NA  NA      NA  73492.235
#> 196   15    0          potential            <=  77.5 197 207     208         NA
#> 197   15    1            overall            <=  59.5 198 199     206         NA
#> 198   15    2               <NA>          <NA>    NA  NA  NA      NA  18769.936
#> 199   15    3                age            <=  31.5 200 204     205         NA
#> 200   15    4          potential            <=  64.5 201 202     203         NA
#> 201   15    5               <NA>          <NA>    NA  NA  NA      NA  26482.585
#> 202   15    6               <NA>          <NA>    NA  NA  NA      NA  34913.289
#> 203   15    7               <NA>          <NA>    NA  NA  NA      NA  34213.653
#> 204   15    8               <NA>          <NA>    NA  NA  NA      NA   9517.171
#> 205   15    9               <NA>          <NA>    NA  NA  NA      NA  30670.731
#> 206   15   10               <NA>          <NA>    NA  NA  NA      NA  28437.082
#> 207   15   11               <NA>          <NA>    NA  NA  NA      NA 191921.234
#> 208   15   12               <NA>          <NA>    NA  NA  NA      NA  55914.011
#> 209   16    0            overall            <=  73.5 210 220     221         NA
#> 210   16    1            overall            <=  59.5 211 212     219         NA
#> 211   16    2               <NA>          <NA>    NA  NA  NA      NA  19745.251
#> 212   16    3                age            <=  30.5 213 217     218         NA
#> 213   16    4          defending            <=  54.5 214 215     216         NA
#> 214   16    5               <NA>          <NA>    NA  NA  NA      NA  36712.373
#> 215   16    6               <NA>          <NA>    NA  NA  NA      NA  31397.487
#> 216   16    7               <NA>          <NA>    NA  NA  NA      NA  27198.652
#> 217   16    8               <NA>          <NA>    NA  NA  NA      NA  15897.864
#> 218   16    9               <NA>          <NA>    NA  NA  NA      NA  30328.552
#> 219   16   10               <NA>          <NA>    NA  NA  NA      NA  28390.803
#> 220   16   11               <NA>          <NA>    NA  NA  NA      NA 376229.111
#> 221   16   12               <NA>          <NA>    NA  NA  NA      NA  78136.362
#> 222   17    0            overall            <=  73.5 223 233     234         NA
#> 223   17    1            overall            <=  59.5 224 225     232         NA
#> 224   17    2               <NA>          <NA>    NA  NA  NA      NA  21918.417
#> 225   17    3          potential            <=  64.5 226 227     231         NA
#> 226   17    4               <NA>          <NA>    NA  NA  NA      NA  26719.345
#> 227   17    5          defending            <=  63.5 228 229     230         NA
#> 228   17    6               <NA>          <NA>    NA  NA  NA      NA  35520.017
#> 229   17    7               <NA>          <NA>    NA  NA  NA      NA  23516.432
#> 230   17    8               <NA>          <NA>    NA  NA  NA      NA  27173.705
#> 231   17    9               <NA>          <NA>    NA  NA  NA      NA  30810.275
#> 232   17   10               <NA>          <NA>    NA  NA  NA      NA  29113.511
#> 233   17   11               <NA>          <NA>    NA  NA  NA      NA 342106.200
#> 234   17   12               <NA>          <NA>    NA  NA  NA      NA  73430.454
#> 235   18    0            overall            <=  74.5 236 246     247         NA
#> 236   18    1            overall            <=  59.5 237 238     245         NA
#> 237   18    2               <NA>          <NA>    NA  NA  NA      NA  23226.576
#> 238   18    3                age            <=  31.5 239 243     244         NA
#> 239   18    4          potential            <=  64.5 240 241     242         NA
#> 240   18    5               <NA>          <NA>    NA  NA  NA      NA  28430.834
#> 241   18    6               <NA>          <NA>    NA  NA  NA      NA  34589.659
#> 242   18    7               <NA>          <NA>    NA  NA  NA      NA  34118.967
#> 243   18    8               <NA>          <NA>    NA  NA  NA      NA  12262.562
#> 244   18    9               <NA>          <NA>    NA  NA  NA      NA  31685.702
#> 245   18   10               <NA>          <NA>    NA  NA  NA      NA  30131.538
#> 246   18   11               <NA>          <NA>    NA  NA  NA      NA 427096.374
#> 247   18   12               <NA>          <NA>    NA  NA  NA      NA  75305.346
#> 248   19    0            overall            <=  74.5 249 259     260         NA
#> 249   19    1            overall            <=  58.5 250 251     258         NA
#> 250   19    2               <NA>          <NA>    NA  NA  NA      NA  23403.918
#> 251   19    3          potential            <=  65.5 252 253     257         NA
#> 252   19    4               <NA>          <NA>    NA  NA  NA      NA  29265.673
#> 253   19    5          defending            <=  65.5 254 255     256         NA
#> 254   19    6               <NA>          <NA>    NA  NA  NA      NA  36196.787
#> 255   19    7               <NA>          <NA>    NA  NA  NA      NA  21505.788
#> 256   19    8               <NA>          <NA>    NA  NA  NA      NA  27209.170
#> 257   19    9               <NA>          <NA>    NA  NA  NA      NA  31777.025
#> 258   19   10               <NA>          <NA>    NA  NA  NA      NA  30522.871
#> 259   19   11               <NA>          <NA>    NA  NA  NA      NA 422185.942
#> 260   19   12               <NA>          <NA>    NA  NA  NA      NA  75693.336
#>     Cover
#> 1     501
#> 2       0
#> 3       0
#> 4       0
#> 5       0
#> 6       0
#> 7       0
#> 8       0
#> 9     501
#> 10      0
#> 11    501
#> 12      0
#> 13      0
#> 14    501
#> 15      0
#> 16      0
#> 17      0
#> 18      0
#> 19      0
#> 20      0
#> 21      0
#> 22    501
#> 23      0
#> 24    501
#> 25      0
#> 26      0
#> 27    501
#> 28      0
#> 29      0
#> 30      0
#> 31      0
#> 32      0
#> 33      0
#> 34      0
#> 35    501
#> 36    435
#> 37     66
#> 38      0
#> 39      0
#> 40    501
#> 41      0
#> 42      0
#> 43      0
#> 44      0
#> 45      0
#> 46      0
#> 47      0
#> 48      0
#> 49      0
#> 50      0
#> 51    501
#> 52      0
#> 53    501
#> 54      0
#> 55      0
#> 56      0
#> 57      0
#> 58      0
#> 59      0
#> 60      0
#> 61      0
#> 62      0
#> 63      0
#> 64    501
#> 65      0
#> 66    501
#> 67      0
#> 68      0
#> 69      0
#> 70      0
#> 71      0
#> 72      0
#> 73      0
#> 74    501
#> 75      0
#> 76    501
#> 77      0
#> 78      0
#> 79    501
#> 80      0
#> 81      0
#> 82      0
#> 83      0
#> 84      0
#> 85      0
#> 86      0
#> 87    501
#> 88    435
#> 89     66
#> 90      0
#> 91      0
#> 92    501
#> 93      0
#> 94      0
#> 95      0
#> 96      0
#> 97      0
#> 98      0
#> 99      0
#> 100   501
#> 101     0
#> 102   501
#> 103     0
#> 104     0
#> 105   501
#> 106     0
#> 107     0
#> 108     0
#> 109     0
#> 110     0
#> 111     0
#> 112     0
#> 113     0
#> 114     0
#> 115     0
#> 116   501
#> 117     0
#> 118   501
#> 119     0
#> 120     0
#> 121     0
#> 122     0
#> 123     0
#> 124     0
#> 125     0
#> 126     0
#> 127     0
#> 128     0
#> 129   501
#> 130     0
#> 131   501
#> 132     0
#> 133     0
#> 134     0
#> 135     0
#> 136     0
#> 137     0
#> 138     0
#> 139   501
#> 140     0
#> 141   501
#> 142     0
#> 143     0
#> 144   501
#> 145     0
#> 146     0
#> 147     0
#> 148     0
#> 149     0
#> 150     0
#> 151     0
#> 152     0
#> 153     0
#> 154     0
#> 155   501
#> 156     0
#> 157   501
#> 158     0
#> 159     0
#> 160     0
#> 161     0
#> 162     0
#> 163     0
#> 164     0
#> 165   501
#> 166     0
#> 167   501
#> 168     0
#> 169     0
#> 170   501
#> 171     0
#> 172     0
#> 173     0
#> 174     0
#> 175     0
#> 176     0
#> 177     0
#> 178     0
#> 179     0
#> 180     0
#> 181   501
#> 182     0
#> 183   501
#> 184     0
#> 185     0
#> 186     0
#> 187     0
#> 188     0
#> 189     0
#> 190     0
#> 191     0
#> 192     0
#> 193     0
#> 194   501
#> 195     0
#> 196   501
#> 197     0
#> 198     0
#> 199     0
#> 200     0
#> 201     0
#> 202     0
#> 203     0
#> 204     0
#> 205     0
#> 206     0
#> 207   501
#> 208     0
#> 209   501
#> 210     0
#> 211     0
#> 212     0
#> 213     0
#> 214     0
#> 215     0
#> 216     0
#> 217     0
#> 218     0
#> 219     0
#> 220   501
#> 221     0
#> 222   501
#> 223     0
#> 224     0
#> 225     0
#> 226     0
#> 227     0
#> 228     0
#> 229     0
#> 230     0
#> 231     0
#> 232     0
#> 233   501
#> 234     0
#> 235   501
#> 236     0
#> 237     0
#> 238     0
#> 239     0
#> 240     0
#> 241     0
#> 242     0
#> 243     0
#> 244     0
#> 245     0
#> 246   501
#> 247     0
#> 248   501
#> 249     0
#> 250     0
#> 251     0
#> 252     0
#> 253     0
#> 254     0
#> 255     0
#> 256     0
#> 257     0
#> 258     0
#> 259   501
#> 260     0
# }