Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node   Feature Decision.type Split Yes  No Missing Prediction Cover
#> 1      0    0   overall            <=  65.5   2   9      13         NA   501
#> 2      0    1 potential            <=  73.5   3   4       8         NA     0
#> 3      0    2      <NA>          <NA>    NA  NA  NA      NA  -7000.000     0
#> 4      0    3   overall            <=  62.5   5   6       7         NA     0
#> 5      0    4      <NA>          <NA>    NA  NA  NA      NA   7500.000     0
#> 6      0    5      <NA>          <NA>    NA  NA  NA      NA  45000.000     0
#> 7      0    6      <NA>          <NA>    NA  NA  NA      NA  25216.535     0
#> 8      0    7      <NA>          <NA>    NA  NA  NA      NA  -1173.036     0
#> 9      0    8 potential            <=  68.5  10  11      12         NA   501
#> 10     0    9      <NA>          <NA>    NA  NA  NA      NA  30000.000     0
#> 11     0   10      <NA>          <NA>    NA  NA  NA      NA 205000.000   501
#> 12     0   11      <NA>          <NA>    NA  NA  NA      NA 179208.283     0
#> 13     0   12      <NA>          <NA>    NA  NA  NA      NA  96054.054     0
#> 14     1    0   overall            <=  66.5  15  22      26         NA   501
#> 15     1    1 potential            <=  73.5  16  17      21         NA     0
#> 16     1    2      <NA>          <NA>    NA  NA  NA      NA   -800.000     0
#> 17     1    3   overall            <=  63.5  18  19      20         NA     0
#> 18     1    4      <NA>          <NA>    NA  NA  NA      NA  14000.000     0
#> 19     1    5      <NA>          <NA>    NA  NA  NA      NA  48000.000     0
#> 20     1    6      <NA>          <NA>    NA  NA  NA      NA  28900.546     0
#> 21     1    7      <NA>          <NA>    NA  NA  NA      NA   4845.201     0
#> 22     1    8 potential            <=  69.5  23  24      25         NA   501
#> 23     1    9      <NA>          <NA>    NA  NA  NA      NA  35500.000     0
#> 24     1   10      <NA>          <NA>    NA  NA  NA      NA 248000.000   501
#> 25     1   11      <NA>          <NA>    NA  NA  NA      NA 211690.751     0
#> 26     1   12      <NA>          <NA>    NA  NA  NA      NA 102734.139     0
#> 27     2    0   overall            <=  67.5  28  35      39         NA   501
#> 28     2    1   overall            <=  63.5  29  30      34         NA     0
#> 29     2    2      <NA>          <NA>    NA  NA  NA      NA  -3220.000     0
#> 30     2    3       age            <=  24.5  31  32      33         NA     0
#> 31     2    4      <NA>          <NA>    NA  NA  NA      NA  40280.000     0
#> 32     2    5      <NA>          <NA>    NA  NA  NA      NA  22780.000     0
#> 33     2    6      <NA>          <NA>    NA  NA  NA      NA  30250.852     0
#> 34     2    7      <NA>          <NA>    NA  NA  NA      NA  10997.143     0
#> 35     2    8 potential            <=  71.5  36  37      38         NA   501
#> 36     2    9      <NA>          <NA>    NA  NA  NA      NA  40450.000     0
#> 37     2   10      <NA>          <NA>    NA  NA  NA      NA 336700.000   501
#> 38     2   11      <NA>          <NA>    NA  NA  NA      NA 258390.666     0
#> 39     2   12      <NA>          <NA>    NA  NA  NA      NA 116272.710     0
#> 40     3    0   overall            <=  69.5  41  51      52         NA   501
#> 41     3    1   overall            <=  64.5  42  46      50         NA     0
#> 42     3    2 potential            <=  75.5  43  44      45         NA     0
#> 43     3    3      <NA>          <NA>    NA  NA  NA      NA   1602.000     0
#> 44     3    4      <NA>          <NA>    NA  NA  NA      NA  34672.000     0
#> 45     3    5      <NA>          <NA>    NA  NA  NA      NA   4759.854     0
#> 46     3    6 defending            <=  59.5  47  48      49         NA     0
#> 47     3    7      <NA>          <NA>    NA  NA  NA      NA  41502.000     0
#> 48     3    8      <NA>          <NA>    NA  NA  NA      NA  29172.000     0
#> 49     3    9      <NA>          <NA>    NA  NA  NA      NA  12802.000     0
#> 50     3   10      <NA>          <NA>    NA  NA  NA      NA  16734.299     0
#> 51     3   11      <NA>          <NA>    NA  NA  NA      NA 316530.000   501
#> 52     3   12      <NA>          <NA>    NA  NA  NA      NA 110160.074     0
#> 53     4    0   overall            <=  69.5  54  64      65         NA   501
#> 54     4    1   overall            <=  63.5  55  56      63         NA     0
#> 55     4    2      <NA>          <NA>    NA  NA  NA      NA   3941.800     0
#> 56     4    3 potential            <=  67.5  57  58      62         NA     0
#> 57     4    4      <NA>          <NA>    NA  NA  NA      NA  22846.800     0
#> 58     4    5 defending            <=  59.5  59  60      61         NA     0
#> 59     4    6      <NA>          <NA>    NA  NA  NA      NA  44021.800     0
#> 60     4    7      <NA>          <NA>    NA  NA  NA      NA  34134.800     0
#> 61     4    8      <NA>          <NA>    NA  NA  NA      NA  18891.800     0
#> 62     4    9      <NA>          <NA>    NA  NA  NA      NA  32871.343     0
#> 63     4   10      <NA>          <NA>    NA  NA  NA      NA  18823.139     0
#> 64     4   11      <NA>          <NA>    NA  NA  NA      NA 278377.000   501
#> 65     4   12      <NA>          <NA>    NA  NA  NA      NA  99878.694     0
#> 66     5    0   overall            <=  68.5  67  74      78         NA   501
#> 67     5    1   overall            <=  62.5  68  69      73         NA     0
#> 68     5    2      <NA>          <NA>    NA  NA  NA      NA   3047.620     0
#> 69     5    3       age            <=  22.5  70  71      72         NA     0
#> 70     5    4      <NA>          <NA>    NA  NA  NA      NA  43047.620     0
#> 71     5    5      <NA>          <NA>    NA  NA  NA      NA  27100.120     0
#> 72     5    6      <NA>          <NA>    NA  NA  NA      NA  31361.316     0
#> 73     5    7      <NA>          <NA>    NA  NA  NA      NA  18707.123     0
#> 74     5    8 potential            <=  73.5  75  76      77         NA   501
#> 75     5    9      <NA>          <NA>    NA  NA  NA      NA  44039.300     0
#> 76     5   10      <NA>          <NA>    NA  NA  NA      NA 394039.300   501
#> 77     5   11      <NA>          <NA>    NA  NA  NA      NA 269687.363     0
#> 78     5   12      <NA>          <NA>    NA  NA  NA      NA 108756.823     0
#> 79     6    0   overall            <=  70.5  80  90      91         NA   501
#> 80     6    1   overall            <=  62.5  81  82      89         NA     0
#> 81     6    2      <NA>          <NA>    NA  NA  NA      NA   6242.858     0
#> 82     6    3 potential            <=  66.5  83  84      88         NA     0
#> 83     6    4      <NA>          <NA>    NA  NA  NA      NA  21337.608     0
#> 84     6    5 defending            <=  59.5  85  86      87         NA     0
#> 85     6    6      <NA>          <NA>    NA  NA  NA      NA  42604.530     0
#> 86     6    7      <NA>          <NA>    NA  NA  NA      NA  32242.858     0
#> 87     6    8      <NA>          <NA>    NA  NA  NA      NA  19922.608     0
#> 88     6    9      <NA>          <NA>    NA  NA  NA      NA  32591.572     0
#> 89     6   10      <NA>          <NA>    NA  NA  NA      NA  22447.899     0
#> 90     6   11      <NA>          <NA>    NA  NA  NA      NA 308135.370   501
#> 91     6   12      <NA>          <NA>    NA  NA  NA      NA  95878.238     0
#> 92     7    0   overall            <=  70.5  93 103     104         NA   501
#> 93     7    1   overall            <=  61.5  94  95     102         NA     0
#> 94     7    2      <NA>          <NA>    NA  NA  NA      NA   7118.572     0
#> 95     7    3       age            <=  21.5  96  97     101         NA     0
#> 96     7    4      <NA>          <NA>    NA  NA  NA      NA  43412.405     0
#> 97     7    5 potential            <=  65.5  98  99     100         NA     0
#> 98     7    6      <NA>          <NA>    NA  NA  NA      NA  19118.572     0
#> 99     7    7      <NA>          <NA>    NA  NA  NA      NA  32014.536     0
#> 100    7    8      <NA>          <NA>    NA  NA  NA      NA  29519.529     0
#> 101    7    9      <NA>          <NA>    NA  NA  NA      NA  31988.371     0
#> 102    7   10      <NA>          <NA>    NA  NA  NA      NA  24178.479     0
#> 103    7   11      <NA>          <NA>    NA  NA  NA      NA 280821.833   501
#> 104    7   12      <NA>          <NA>    NA  NA  NA      NA  91688.121     0
#> 105    8    0 potential            <=  72.5 106 113     117         NA   501
#> 106    8    1   overall            <=  64.5 107 108     112         NA     0
#> 107    8    2      <NA>          <NA>    NA  NA  NA      NA  14801.490     0
#> 108    8    3       age            <=  30.5 109 110     111         NA     0
#> 109    8    4      <NA>          <NA>    NA  NA  NA      NA  36938.569     0
#> 110    8    5      <NA>          <NA>    NA  NA  NA      NA   5257.442     0
#> 111    8    6      <NA>          <NA>    NA  NA  NA      NA  27440.479     0
#> 112    8    7      <NA>          <NA>    NA  NA  NA      NA  20799.213     0
#> 113    8    8   overall            <=  60.5 114 115     116         NA   501
#> 114    8    9      <NA>          <NA>    NA  NA  NA      NA  18169.715     0
#> 115    8   10      <NA>          <NA>    NA  NA  NA      NA  86239.650   501
#> 116    8   11      <NA>          <NA>    NA  NA  NA      NA  80484.988     0
#> 117    8   12      <NA>          <NA>    NA  NA  NA      NA  45596.991     0
#> 118    9    0   overall            <=  71.5 119 126     130         NA   501
#> 119    9    1 dribbling            <=  63.5 120 124     125         NA     0
#> 120    9    2   overall            <=  61.5 121 122     123         NA     0
#> 121    9    3      <NA>          <NA>    NA  NA  NA      NA  11926.566     0
#> 122    9    4      <NA>          <NA>    NA  NA  NA      NA  30603.416     0
#> 123    9    5      <NA>          <NA>    NA  NA  NA      NA  22055.343     0
#> 124    9    6      <NA>          <NA>    NA  NA  NA      NA  35617.633     0
#> 125    9    7      <NA>          <NA>    NA  NA  NA      NA   9950.747     0
#> 126    9    8       age            <=  33.5 127 128     129         NA   501
#> 127    9    9      <NA>          <NA>    NA  NA  NA      NA 401115.685   462
#> 128    9   10      <NA>          <NA>    NA  NA  NA      NA -38884.315    39
#> 129    9   11      <NA>          <NA>    NA  NA  NA      NA 368866.689     0
#> 130    9   12      <NA>          <NA>    NA  NA  NA      NA 100483.302     0
#> 131   10    0   overall            <=  71.5 132 142     143         NA   501
#> 132   10    1 dribbling            <=  63.5 133 140     141         NA     0
#> 133   10    2   overall            <=  60.5 134 135     139         NA     0
#> 134   10    3      <NA>          <NA>    NA  NA  NA      NA  12233.909     0
#> 135   10    4 potential            <=  66.5 136 137     138         NA     0
#> 136   10    5      <NA>          <NA>    NA  NA  NA      NA  23166.224     0
#> 137   10    6      <NA>          <NA>    NA  NA  NA      NA  33525.060     0
#> 138   10    7      <NA>          <NA>    NA  NA  NA      NA  30517.958     0
#> 139   10    8      <NA>          <NA>    NA  NA  NA      NA  23335.655     0
#> 140   10    9      <NA>          <NA>    NA  NA  NA      NA  35468.436     0
#> 141   10   10      <NA>          <NA>    NA  NA  NA      NA  13431.491     0
#> 142   10   11      <NA>          <NA>    NA  NA  NA      NA 314504.116   501
#> 143   10   12      <NA>          <NA>    NA  NA  NA      NA  90639.389     0
#> 144   11    0 potential            <=  75.5 145 152     156         NA   501
#> 145   11    1  shooting            <=  58.5 146 147     151         NA     0
#> 146   11    2      <NA>          <NA>    NA  NA  NA      NA  24141.782     0
#> 147   11    3       age            <=  31.5 148 149     150         NA     0
#> 148   11    4      <NA>          <NA>    NA  NA  NA      NA  39387.955     0
#> 149   11    5      <NA>          <NA>    NA  NA  NA      NA   5430.850     0
#> 150   11    6      <NA>          <NA>    NA  NA  NA      NA  34094.840     0
#> 151   11    7      <NA>          <NA>    NA  NA  NA      NA  13388.342     0
#> 152   11    8   overall            <=  60.5 153 154     155         NA   501
#> 153   11    9      <NA>          <NA>    NA  NA  NA      NA  21936.696     0
#> 154   11   10      <NA>          <NA>    NA  NA  NA      NA 186553.705   501
#> 155   11   11      <NA>          <NA>    NA  NA  NA      NA 176749.847     0
#> 156   11   12      <NA>          <NA>    NA  NA  NA      NA  62910.821     0
#> 157   12    0   overall            <=  72.5 158 168     169         NA   501
#> 158   12    1   overall            <=  63.5 159 160     167         NA     0
#> 159   12    2      <NA>          <NA>    NA  NA  NA      NA  21061.095     0
#> 160   12    3       age            <=  30.5 161 165     166         NA     0
#> 161   12    4 defending            <=  57.5 162 163     164         NA     0
#> 162   12    5      <NA>          <NA>    NA  NA  NA      NA  39533.320     0
#> 163   12    6      <NA>          <NA>    NA  NA  NA      NA  32539.634     0
#> 164   12    7      <NA>          <NA>    NA  NA  NA      NA  25310.575     0
#> 165   12    8      <NA>          <NA>    NA  NA  NA      NA   8876.100     0
#> 166   12    9      <NA>          <NA>    NA  NA  NA      NA  30024.075     0
#> 167   12   10      <NA>          <NA>    NA  NA  NA      NA  26307.074     0
#> 168   12   11      <NA>          <NA>    NA  NA  NA      NA 386114.909   501
#> 169   12   12      <NA>          <NA>    NA  NA  NA      NA  91622.886     0
#> 170   13    0   overall            <=  72.5 171 181     182         NA   501
#> 171   13    1   overall            <=  59.5 172 173     180         NA     0
#> 172   13    2      <NA>          <NA>    NA  NA  NA      NA  14990.231     0
#> 173   13    3 potential            <=  65.5 174 175     179         NA     0
#> 174   13    4      <NA>          <NA>    NA  NA  NA      NA  25143.398     0
#> 175   13    5 defending            <=  55.5 176 177     178         NA     0
#> 176   13    6      <NA>          <NA>    NA  NA  NA      NA  37600.376     0
#> 177   13    7      <NA>          <NA>    NA  NA  NA      NA  31998.584     0
#> 178   13    8      <NA>          <NA>    NA  NA  NA      NA  27955.106     0
#> 179   13    9      <NA>          <NA>    NA  NA  NA      NA  32554.768     0
#> 180   13   10      <NA>          <NA>    NA  NA  NA      NA  29112.156     0
#> 181   13   11      <NA>          <NA>    NA  NA  NA      NA 336286.843   501
#> 182   13   12      <NA>          <NA>    NA  NA  NA      NA  82655.134     0
#> 183   14    0   overall            <=  73.5 184 194     195         NA   501
#> 184   14    1   overall            <=  59.5 185 186     193         NA     0
#> 185   14    2      <NA>          <NA>    NA  NA  NA      NA  16991.208     0
#> 186   14    3 potential            <=  65.5 187 188     192         NA     0
#> 187   14    4      <NA>          <NA>    NA  NA  NA      NA  26400.249     0
#> 188   14    5 defending            <=  54.5 189 190     191         NA     0
#> 189   14    6      <NA>          <NA>    NA  NA  NA      NA  37506.620     0
#> 190   14    7      <NA>          <NA>    NA  NA  NA      NA  32298.726     0
#> 191   14    8      <NA>          <NA>    NA  NA  NA      NA  26847.888     0
#> 192   14    9      <NA>          <NA>    NA  NA  NA      NA  32675.340     0
#> 193   14   10      <NA>          <NA>    NA  NA  NA      NA  29693.643     0
#> 194   14   11      <NA>          <NA>    NA  NA  NA      NA 406158.159   501
#> 195   14   12      <NA>          <NA>    NA  NA  NA      NA  85469.215     0
#> 196   15    0   overall            <=  73.5 197 207     208         NA   501
#> 197   15    1   overall            <=  59.5 198 199     206         NA     0
#> 198   15    2      <NA>          <NA>    NA  NA  NA      NA  18792.087     0
#> 199   15    3       age            <=  30.5 200 204     205         NA     0
#> 200   15    4  shooting            <=  52.5 201 202     203         NA     0
#> 201   15    5      <NA>          <NA>    NA  NA  NA      NA  32195.262     0
#> 202   15    6      <NA>          <NA>    NA  NA  NA      NA  37372.411     0
#> 203   15    7      <NA>          <NA>    NA  NA  NA      NA  28874.309     0
#> 204   15    8      <NA>          <NA>    NA  NA  NA      NA  10906.828     0
#> 205   15    9      <NA>          <NA>    NA  NA  NA      NA  30588.316     0
#> 206   15   10      <NA>          <NA>    NA  NA  NA      NA  28342.641     0
#> 207   15   11      <NA>          <NA>    NA  NA  NA      NA 419042.343   501
#> 208   15   12      <NA>          <NA>    NA  NA  NA      NA  86013.491     0
#> 209   16    0   overall            <=  63.5 210 211     221         NA   501
#> 210   16    1      <NA>          <NA>    NA  NA  NA      NA  26412.878     0
#> 211   16    2 potential            <=  77.5 212 219     220         NA   501
#> 212   16    3       age            <=  31.5 213 217     218         NA     0
#> 213   16    4  shooting            <=  64.5 214 215     216         NA     0
#> 214   16    5      <NA>          <NA>    NA  NA  NA      NA  34303.756     0
#> 215   16    6      <NA>          <NA>    NA  NA  NA      NA  50355.931     0
#> 216   16    7      <NA>          <NA>    NA  NA  NA      NA  28348.010     0
#> 217   16    8      <NA>          <NA>    NA  NA  NA      NA   -324.573     0
#> 218   16    9      <NA>          <NA>    NA  NA  NA      NA  31480.304     0
#> 219   16   10      <NA>          <NA>    NA  NA  NA      NA 330638.109   501
#> 220   16   11      <NA>          <NA>    NA  NA  NA      NA  99260.550     0
#> 221   16   12      <NA>          <NA>    NA  NA  NA      NA  74892.968     0
#> 222   17    0   overall            <=  73.5 223 233     234         NA   501
#> 223   17    1   overall            <=  59.5 224 225     232         NA     0
#> 224   17    2      <NA>          <NA>    NA  NA  NA      NA  21271.590     0
#> 225   17    3 potential            <=  64.5 226 227     231         NA     0
#> 226   17    4      <NA>          <NA>    NA  NA  NA      NA  26926.895     0
#> 227   17    5       age            <=  32.5 228 229     230         NA     0
#> 228   17    6      <NA>          <NA>    NA  NA  NA      NA  34489.602     0
#> 229   17    7      <NA>          <NA>    NA  NA  NA      NA  -1289.768     0
#> 230   17    8      <NA>          <NA>    NA  NA  NA      NA  31951.031     0
#> 231   17    9      <NA>          <NA>    NA  NA  NA      NA  31465.236     0
#> 232   17   10      <NA>          <NA>    NA  NA  NA      NA  29506.129     0
#> 233   17   11      <NA>          <NA>    NA  NA  NA      NA 351074.298   501
#> 234   17   12      <NA>          <NA>    NA  NA  NA      NA  77183.650     0
#> 235   18    0   overall            <=  74.5 236 246     247         NA   501
#> 236   18    1   overall            <=  59.5 237 238     245         NA     0
#> 237   18    2      <NA>          <NA>    NA  NA  NA      NA  22644.431     0
#> 238   18    3 potential            <=  65.5 239 240     244         NA     0
#> 239   18    4      <NA>          <NA>    NA  NA  NA      NA  29026.092     0
#> 240   18    5       age            <=  32.5 241 242     243         NA     0
#> 241   18    6      <NA>          <NA>    NA  NA  NA      NA  35017.683     0
#> 242   18    7      <NA>          <NA>    NA  NA  NA      NA -11640.345     0
#> 243   18    8      <NA>          <NA>    NA  NA  NA      NA  31792.814     0
#> 244   18    9      <NA>          <NA>    NA  NA  NA      NA  31414.732     0
#> 245   18   10      <NA>          <NA>    NA  NA  NA      NA  29827.767     0
#> 246   18   11      <NA>          <NA>    NA  NA  NA      NA 419466.868   501
#> 247   18   12      <NA>          <NA>    NA  NA  NA      NA  76640.737     0
#> 248   19    0   overall            <=  58.5 249 250     260         NA   501
#> 249   19    1      <NA>          <NA>    NA  NA  NA      NA  22879.988     0
#> 250   19    2   overall            <=  74.5 251 258     259         NA   501
#> 251   19    3 potential            <=  64.5 252 253     257         NA     0
#> 252   19    4      <NA>          <NA>    NA  NA  NA      NA  27879.988     0
#> 253   19    5 defending            <=  61.5 254 255     256         NA     0
#> 254   19    6      <NA>          <NA>    NA  NA  NA      NA  36257.311     0
#> 255   19    7      <NA>          <NA>    NA  NA  NA      NA  30117.147     0
#> 256   19    8      <NA>          <NA>    NA  NA  NA      NA  29136.661     0
#> 257   19    9      <NA>          <NA>    NA  NA  NA      NA  32808.448     0
#> 258   19   10      <NA>          <NA>    NA  NA  NA      NA 381020.181   501
#> 259   19   11      <NA>          <NA>    NA  NA  NA      NA  80607.384     0
#> 260   19   12      <NA>          <NA>    NA  NA  NA      NA  72945.350     0
# }