Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node     Feature Decision.type Split Yes  No Missing Prediction Cover
#> 1      0    0     overall            <=  65.5   2   9      13         NA   501
#> 2      0    1   potential            <=  74.5   3   4       8         NA     0
#> 3      0    2        <NA>          <NA>    NA  NA  NA      NA  -6000.000     0
#> 4      0    3     overall            <=  62.5   5   6       7         NA     0
#> 5      0    4        <NA>          <NA>    NA  NA  NA      NA   7500.000     0
#> 6      0    5        <NA>          <NA>    NA  NA  NA      NA  45000.000     0
#> 7      0    6        <NA>          <NA>    NA  NA  NA      NA  26119.792     0
#> 8      0    7        <NA>          <NA>    NA  NA  NA      NA  -1604.419     0
#> 9      0    8   potential            <=  68.5  10  11      12         NA   501
#> 10     0    9        <NA>          <NA>    NA  NA  NA      NA  27500.000     0
#> 11     0   10        <NA>          <NA>    NA  NA  NA      NA 205000.000   501
#> 12     0   11        <NA>          <NA>    NA  NA  NA      NA 178897.059     0
#> 13     0   12        <NA>          <NA>    NA  NA  NA      NA  95766.441     0
#> 14     1    0     overall            <=  66.5  15  22      26         NA   501
#> 15     1    1   potential            <=  73.5  16  17      21         NA     0
#> 16     1    2        <NA>          <NA>    NA  NA  NA      NA   -900.000     0
#> 17     1    3     overall            <=  63.5  18  19      20         NA     0
#> 18     1    4        <NA>          <NA>    NA  NA  NA      NA  16500.000     0
#> 19     1    5        <NA>          <NA>    NA  NA  NA      NA  48000.000     0
#> 20     1    6        <NA>          <NA>    NA  NA  NA      NA  30341.960     0
#> 21     1    7        <NA>          <NA>    NA  NA  NA      NA   5050.850     0
#> 22     1    8   potential            <=  69.5  23  24      25         NA   501
#> 23     1    9        <NA>          <NA>    NA  NA  NA      NA  35500.000     0
#> 24     1   10        <NA>          <NA>    NA  NA  NA      NA 248000.000   501
#> 25     1   11        <NA>          <NA>    NA  NA  NA      NA 211983.875     0
#> 26     1   12        <NA>          <NA>    NA  NA  NA      NA 104045.399     0
#> 27     2    0     overall            <=  67.5  28  35      39         NA   501
#> 28     2    1     overall            <=  63.5  29  30      34         NA     0
#> 29     2    2        <NA>          <NA>    NA  NA  NA      NA  -3310.000     0
#> 30     2    3         age            <=  23.5  31  32      33         NA     0
#> 31     2    4        <NA>          <NA>    NA  NA  NA      NA  42700.000     0
#> 32     2    5        <NA>          <NA>    NA  NA  NA      NA  24340.000     0
#> 33     2    6        <NA>          <NA>    NA  NA  NA      NA  30642.078     0
#> 34     2    7        <NA>          <NA>    NA  NA  NA      NA  11040.616     0
#> 35     2    8         age            <=  32.5  36  37      38         NA   501
#> 36     2    9        <NA>          <NA>    NA  NA  NA      NA 236700.000   435
#> 37     2   10        <NA>          <NA>    NA  NA  NA      NA  36700.000    66
#> 38     2   11        <NA>          <NA>    NA  NA  NA      NA 215025.899     0
#> 39     2   12        <NA>          <NA>    NA  NA  NA      NA  96103.305     0
#> 40     3    0     overall            <=  69.5  41  51      52         NA   501
#> 41     3    1   potential            <=  75.5  42  46      50         NA     0
#> 42     3    2     overall            <=  64.5  43  44      45         NA     0
#> 43     3    3        <NA>          <NA>    NA  NA  NA      NA   2521.000     0
#> 44     3    4        <NA>          <NA>    NA  NA  NA      NA  26256.000     0
#> 45     3    5        <NA>          <NA>    NA  NA  NA      NA  12200.182     0
#> 46     3    6     overall            <=  61.5  47  48      49         NA     0
#> 47     3    7        <NA>          <NA>    NA  NA  NA      NA  10931.000     0
#> 48     3    8        <NA>          <NA>    NA  NA  NA      NA  49430.000     0
#> 49     3    9        <NA>          <NA>    NA  NA  NA      NA  41846.132     0
#> 50     3   10        <NA>          <NA>    NA  NA  NA      NA  15930.498     0
#> 51     3   11        <NA>          <NA>    NA  NA  NA      NA 326530.000   501
#> 52     3   12        <NA>          <NA>    NA  NA  NA      NA 111261.672     0
#> 53     4    0     overall            <=  69.5  54  61      65         NA   501
#> 54     4    1     overall            <=  63.5  55  56      60         NA     0
#> 55     4    2        <NA>          <NA>    NA  NA  NA      NA   3768.900     0
#> 56     4    3         age            <=  23.5  57  58      59         NA     0
#> 57     4    4        <NA>          <NA>    NA  NA  NA      NA  44395.700     0
#> 58     4    5        <NA>          <NA>    NA  NA  NA      NA  24503.900     0
#> 59     4    6        <NA>          <NA>    NA  NA  NA      NA  30841.435     0
#> 60     4    7        <NA>          <NA>    NA  NA  NA      NA  17626.295     0
#> 61     4    8   potential            <=  70.5  62  63      64         NA   501
#> 62     4    9        <NA>          <NA>    NA  NA  NA      NA  27377.000     0
#> 63     4   10        <NA>          <NA>    NA  NA  NA      NA 327377.000   501
#> 64     4   11        <NA>          <NA>    NA  NA  NA      NA 305098.702     0
#> 65     4   12        <NA>          <NA>    NA  NA  NA      NA 104884.032     0
#> 66     5    0   potential            <=  71.5  67  71      78         NA   501
#> 67     5    1    shooting            <=  59.5  68  69      70         NA     0
#> 68     5    2        <NA>          <NA>    NA  NA  NA      NA  12892.010     0
#> 69     5    3        <NA>          <NA>    NA  NA  NA      NA  32330.010     0
#> 70     5    4        <NA>          <NA>    NA  NA  NA      NA  -1107.990     0
#> 71     5    5     overall            <=  61.5  72  73      77         NA   501
#> 72     5    6        <NA>          <NA>    NA  NA  NA      NA  13652.010     0
#> 73     5    7 gk_reflexes            <=  73.5  74  75      76         NA   501
#> 74     5    8        <NA>          <NA>    NA  NA  NA      NA  24228.330     0
#> 75     5    9        <NA>          <NA>    NA  NA  NA      NA 348139.300    56
#> 76     5   10        <NA>          <NA>    NA  NA  NA      NA 138139.300   445
#> 77     5   11        <NA>          <NA>    NA  NA  NA      NA 128841.711     0
#> 78     5   12        <NA>          <NA>    NA  NA  NA      NA  70142.240     0
#> 79     6    0     overall            <=  69.5  80  90      91         NA   501
#> 80     6    1     overall            <=  62.5  81  82      89         NA     0
#> 81     6    2        <NA>          <NA>    NA  NA  NA      NA   5102.809     0
#> 82     6    3   potential            <=  77.5  83  87      88         NA     0
#> 83     6    4   defending            <=  57.5  84  85      86         NA     0
#> 84     6    5        <NA>          <NA>    NA  NA  NA      NA  32890.809     0
#> 85     6    6        <NA>          <NA>    NA  NA  NA      NA  24664.809     0
#> 86     6    7        <NA>          <NA>    NA  NA  NA      NA  18190.809     0
#> 87     6    8        <NA>          <NA>    NA  NA  NA      NA  53233.500     0
#> 88     6    9        <NA>          <NA>    NA  NA  NA      NA  29601.003     0
#> 89     6   10        <NA>          <NA>    NA  NA  NA      NA  19634.262     0
#> 90     6   11        <NA>          <NA>    NA  NA  NA      NA 257825.370   501
#> 91     6   12        <NA>          <NA>    NA  NA  NA      NA  92350.445     0
#> 92     7    0     overall            <=  70.5  93 103     104         NA   501
#> 93     7    1   potential            <=  77.5  94 101     102         NA     0
#> 94     7    2     overall            <=  64.5  95  96     100         NA     0
#> 95     7    3        <NA>          <NA>    NA  NA  NA      NA  14016.528     0
#> 96     7    4   defending            <=  58.5  97  98      99         NA     0
#> 97     7    5        <NA>          <NA>    NA  NA  NA      NA  35431.928     0
#> 98     7    6        <NA>          <NA>    NA  NA  NA      NA  23198.328     0
#> 99     7    7        <NA>          <NA>    NA  NA  NA      NA  12745.728     0
#> 100    7    8        <NA>          <NA>    NA  NA  NA      NA  19967.635     0
#> 101    7    9        <NA>          <NA>    NA  NA  NA      NA  44910.150     0
#> 102    7   10        <NA>          <NA>    NA  NA  NA      NA  21863.442     0
#> 103    7   11        <NA>          <NA>    NA  NA  NA      NA 315542.833   501
#> 104    7   12        <NA>          <NA>    NA  NA  NA      NA  96962.330     0
#> 105    8    0     overall            <=  71.5 106 116     117         NA   501
#> 106    8    1     overall            <=  62.5 107 108     115         NA     0
#> 107    8    2        <NA>          <NA>    NA  NA  NA      NA  11114.875     0
#> 108    8    3         age            <=  28.5 109 113     114         NA     0
#> 109    8    4   defending            <=  53.5 110 111     112         NA     0
#> 110    8    5        <NA>          <NA>    NA  NA  NA      NA  39218.626     0
#> 111    8    6        <NA>          <NA>    NA  NA  NA      NA  31653.771     0
#> 112    8    7        <NA>          <NA>    NA  NA  NA      NA  25282.075     0
#> 113    8    8        <NA>          <NA>    NA  NA  NA      NA  17234.675     0
#> 114    8    9        <NA>          <NA>    NA  NA  NA      NA  29155.740     0
#> 115    8   10        <NA>          <NA>    NA  NA  NA      NA  22690.489     0
#> 116    8   11        <NA>          <NA>    NA  NA  NA      NA 387488.550   501
#> 117    8   12        <NA>          <NA>    NA  NA  NA      NA 101006.913     0
#> 118    9    0     overall            <=  71.5 119 126     130         NA   501
#> 119    9    1   dribbling            <=  63.5 120 121     125         NA     0
#> 120    9    2        <NA>          <NA>    NA  NA  NA      NA  20579.388     0
#> 121    9    3         age            <=  29.5 122 123     124         NA     0
#> 122    9    4        <NA>          <NA>    NA  NA  NA      NA  35801.804     0
#> 123    9    5        <NA>          <NA>    NA  NA  NA      NA  17857.102     0
#> 124    9    6        <NA>          <NA>    NA  NA  NA      NA  32875.229     0
#> 125    9    7        <NA>          <NA>    NA  NA  NA      NA  10979.388     0
#> 126    9    8   potential            <=  72.5 127 128     129         NA   501
#> 127    9    9        <NA>          <NA>    NA  NA  NA      NA   2239.695     0
#> 128    9   10        <NA>          <NA>    NA  NA  NA      NA 402239.695   501
#> 129    9   11        <NA>          <NA>    NA  NA  NA      NA 363975.395     0
#> 130    9   12        <NA>          <NA>    NA  NA  NA      NA  99469.743     0
#> 131   10    0   potential            <=  73.5 132 139     143         NA   501
#> 132   10    1     overall            <=  62.5 133 134     138         NA     0
#> 133   10    2        <NA>          <NA>    NA  NA  NA      NA  13021.449     0
#> 134   10    3         age            <=  30.5 135 136     137         NA     0
#> 135   10    4        <NA>          <NA>    NA  NA  NA      NA  33323.089     0
#> 136   10    5        <NA>          <NA>    NA  NA  NA      NA   7848.369     0
#> 137   10    6        <NA>          <NA>    NA  NA  NA      NA  26822.264     0
#> 138   10    7        <NA>          <NA>    NA  NA  NA      NA  21704.872     0
#> 139   10    8     overall            <=  68.5 140 141     142         NA   501
#> 140   10    9        <NA>          <NA>    NA  NA  NA      NA  33331.623     0
#> 141   10   10        <NA>          <NA>    NA  NA  NA      NA 315515.725   501
#> 142   10   11        <NA>          <NA>    NA  NA  NA      NA 213821.485     0
#> 143   10   12        <NA>          <NA>    NA  NA  NA      NA  89835.952     0
#> 144   11    0     overall            <=  72.5 145 155     156         NA   501
#> 145   11    1     overall            <=  60.5 146 147     154         NA     0
#> 146   11    2        <NA>          <NA>    NA  NA  NA      NA  12219.304     0
#> 147   11    3         age            <=  29.5 148 152     153         NA     0
#> 148   11    4   potential            <=  65.5 149 150     151         NA     0
#> 149   11    5        <NA>          <NA>    NA  NA  NA      NA  27239.223     0
#> 150   11    6        <NA>          <NA>    NA  NA  NA      NA  35179.051     0
#> 151   11    7        <NA>          <NA>    NA  NA  NA      NA  34392.107     0
#> 152   11    8        <NA>          <NA>    NA  NA  NA      NA  15312.252     0
#> 153   11    9        <NA>          <NA>    NA  NA  NA      NA  30076.066     0
#> 154   11   10        <NA>          <NA>    NA  NA  NA      NA  25751.121     0
#> 155   11   11        <NA>          <NA>    NA  NA  NA      NA 387464.153   501
#> 156   11   12        <NA>          <NA>    NA  NA  NA      NA  90542.043     0
#> 157   12    0   potential            <=  75.5 158 165     169         NA   501
#> 158   12    1     overall            <=  60.5 159 160     164         NA     0
#> 159   12    2        <NA>          <NA>    NA  NA  NA      NA  14497.374     0
#> 160   12    3   defending            <=  52.5 161 162     163         NA     0
#> 161   12    4        <NA>          <NA>    NA  NA  NA      NA  35871.409     0
#> 162   12    5        <NA>          <NA>    NA  NA  NA      NA  28656.938     0
#> 163   12    6        <NA>          <NA>    NA  NA  NA      NA  21306.107     0
#> 164   12    7        <NA>          <NA>    NA  NA  NA      NA  26479.395     0
#> 165   12    8     overall            <=  70.5 166 167     168         NA   501
#> 166   12    9        <NA>          <NA>    NA  NA  NA      NA  35850.561     0
#> 167   12   10        <NA>          <NA>    NA  NA  NA      NA 452217.737   501
#> 168   12   11        <NA>          <NA>    NA  NA  NA      NA 292570.561     0
#> 169   12   12        <NA>          <NA>    NA  NA  NA      NA  92543.610     0
#> 170   13    0   potential            <=  76.5 171 181     182         NA   501
#> 171   13    1     overall            <=  59.5 172 173     180         NA     0
#> 172   13    2        <NA>          <NA>    NA  NA  NA      NA  14547.636     0
#> 173   13    3   potential            <=  65.5 174 175     179         NA     0
#> 174   13    4        <NA>          <NA>    NA  NA  NA      NA  25385.999     0
#> 175   13    5         age            <=  31.5 176 177     178         NA     0
#> 176   13    6        <NA>          <NA>    NA  NA  NA      NA  34592.017     0
#> 177   13    7        <NA>          <NA>    NA  NA  NA      NA  -4554.681     0
#> 178   13    8        <NA>          <NA>    NA  NA  NA      NA  29558.191     0
#> 179   13    9        <NA>          <NA>    NA  NA  NA      NA  28896.062     0
#> 180   13   10        <NA>          <NA>    NA  NA  NA      NA  26049.463     0
#> 181   13   11        <NA>          <NA>    NA  NA  NA      NA 135724.474   501
#> 182   13   12        <NA>          <NA>    NA  NA  NA      NA  47710.848     0
#> 183   14    0     overall            <=  72.5 184 194     195         NA   501
#> 184   14    1     overall            <=  59.5 185 186     193         NA     0
#> 185   14    2        <NA>          <NA>    NA  NA  NA      NA  16592.873     0
#> 186   14    3   potential            <=  65.5 187 188     192         NA     0
#> 187   14    4        <NA>          <NA>    NA  NA  NA      NA  26228.060     0
#> 188   14    5   defending            <=  63.5 189 190     191         NA     0
#> 189   14    6        <NA>          <NA>    NA  NA  NA      NA  35529.378     0
#> 190   14    7        <NA>          <NA>    NA  NA  NA      NA  22955.341     0
#> 191   14    8        <NA>          <NA>    NA  NA  NA      NA  27715.903     0
#> 192   14    9        <NA>          <NA>    NA  NA  NA      NA  30566.140     0
#> 193   14   10        <NA>          <NA>    NA  NA  NA      NA  27853.279     0
#> 194   14   11        <NA>          <NA>    NA  NA  NA      NA 329423.516   501
#> 195   14   12        <NA>          <NA>    NA  NA  NA      NA  80254.367     0
#> 196   15    0     overall            <=  73.5 197 207     208         NA   501
#> 197   15    1     overall            <=  59.5 198 199     206         NA     0
#> 198   15    2        <NA>          <NA>    NA  NA  NA      NA  18433.585     0
#> 199   15    3         age            <=  30.5 200 204     205         NA     0
#> 200   15    4    shooting            <=  57.5 201 202     203         NA     0
#> 201   15    5        <NA>          <NA>    NA  NA  NA      NA  31612.069     0
#> 202   15    6        <NA>          <NA>    NA  NA  NA      NA  36700.480     0
#> 203   15    7        <NA>          <NA>    NA  NA  NA      NA  28444.312     0
#> 204   15    8        <NA>          <NA>    NA  NA  NA      NA  14479.927     0
#> 205   15    9        <NA>          <NA>    NA  NA  NA      NA  30243.309     0
#> 206   15   10        <NA>          <NA>    NA  NA  NA      NA  28018.859     0
#> 207   15   11        <NA>          <NA>    NA  NA  NA      NA 420981.165   501
#> 208   15   12        <NA>          <NA>    NA  NA  NA      NA  85636.705     0
#> 209   16    0     overall            <=  73.5 210 220     221         NA   501
#> 210   16    1     overall            <=  60.5 211 212     219         NA     0
#> 211   16    2        <NA>          <NA>    NA  NA  NA      NA  21450.227     0
#> 212   16    3   defending            <=  55.5 213 217     218         NA     0
#> 213   16    4   potential            <=  64.5 214 215     216         NA     0
#> 214   16    5        <NA>          <NA>    NA  NA  NA      NA  29085.628     0
#> 215   16    6        <NA>          <NA>    NA  NA  NA      NA  36893.150     0
#> 216   16    7        <NA>          <NA>    NA  NA  NA      NA  36142.038     0
#> 217   16    8        <NA>          <NA>    NA  NA  NA      NA  29702.371     0
#> 218   16    9        <NA>          <NA>    NA  NA  NA      NA  24411.631     0
#> 219   16   10        <NA>          <NA>    NA  NA  NA      NA  29257.620     0
#> 220   16   11        <NA>          <NA>    NA  NA  NA      NA 382383.048   501
#> 221   16   12        <NA>          <NA>    NA  NA  NA      NA  81691.279     0
#> 222   17    0     overall            <=  73.5 223 233     234         NA   501
#> 223   17    1     overall            <=  59.5 224 225     232         NA     0
#> 224   17    2        <NA>          <NA>    NA  NA  NA      NA  21445.204     0
#> 225   17    3   potential            <=  64.5 226 227     231         NA     0
#> 226   17    4        <NA>          <NA>    NA  NA  NA      NA  26796.584     0
#> 227   17    5   defending            <=  65.5 228 229     230         NA     0
#> 228   17    6        <NA>          <NA>    NA  NA  NA      NA  35269.669     0
#> 229   17    7        <NA>          <NA>    NA  NA  NA      NA  21764.596     0
#> 230   17    8        <NA>          <NA>    NA  NA  NA      NA  28277.771     0
#> 231   17    9        <NA>          <NA>    NA  NA  NA      NA  31318.241     0
#> 232   17   10        <NA>          <NA>    NA  NA  NA      NA  29454.092     0
#> 233   17   11        <NA>          <NA>    NA  NA  NA      NA 347644.743   501
#> 234   17   12        <NA>          <NA>    NA  NA  NA      NA  75655.974     0
#> 235   18    0     overall            <=  74.5 236 246     247         NA   501
#> 236   18    1     overall            <=  60.5 237 238     245         NA     0
#> 237   18    2        <NA>          <NA>    NA  NA  NA      NA  24160.684     0
#> 238   18    3         age            <=  31.5 239 243     244         NA     0
#> 239   18    4   potential            <=  64.5 240 241     242         NA     0
#> 240   18    5        <NA>          <NA>    NA  NA  NA      NA  29620.670     0
#> 241   18    6        <NA>          <NA>    NA  NA  NA      NA  34566.610     0
#> 242   18    7        <NA>          <NA>    NA  NA  NA      NA  34218.646     0
#> 243   18    8        <NA>          <NA>    NA  NA  NA      NA  11871.398     0
#> 244   18    9        <NA>          <NA>    NA  NA  NA      NA  31612.720     0
#> 245   18   10        <NA>          <NA>    NA  NA  NA      NA  29908.529     0
#> 246   18   11        <NA>          <NA>    NA  NA  NA      NA 416380.269   501
#> 247   18   12        <NA>          <NA>    NA  NA  NA      NA  74099.684     0
#> 248   19    0     overall            <=  74.5 249 259     260         NA   501
#> 249   19    1     overall            <=  58.5 250 251     258         NA     0
#> 250   19    2        <NA>          <NA>    NA  NA  NA      NA  22244.615     0
#> 251   19    3   potential            <=  65.5 252 253     257         NA     0
#> 252   19    4        <NA>          <NA>    NA  NA  NA      NA  29550.152     0
#> 253   19    5   defending            <=  65.5 254 255     256         NA     0
#> 254   19    6        <NA>          <NA>    NA  NA  NA      NA  35512.592     0
#> 255   19    7        <NA>          <NA>    NA  NA  NA      NA  23233.655     0
#> 256   19    8        <NA>          <NA>    NA  NA  NA      NA  29443.545     0
#> 257   19    9        <NA>          <NA>    NA  NA  NA      NA  31931.778     0
#> 258   19   10        <NA>          <NA>    NA  NA  NA      NA  30425.796     0
#> 259   19   11        <NA>          <NA>    NA  NA  NA      NA 378242.242   501
#> 260   19   12        <NA>          <NA>    NA  NA  NA      NA  71985.656     0
# }