Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node            Feature Decision.type Split Yes  No Missing
#> 1      0    0            overall            <=  65.5   2   9      13
#> 2      0    1          potential            <=  73.5   3   4       8
#> 3      0    2               <NA>          <NA>    NA  NA  NA      NA
#> 4      0    3            overall            <=  62.5   5   6       7
#> 5      0    4               <NA>          <NA>    NA  NA  NA      NA
#> 6      0    5               <NA>          <NA>    NA  NA  NA      NA
#> 7      0    6               <NA>          <NA>    NA  NA  NA      NA
#> 8      0    7               <NA>          <NA>    NA  NA  NA      NA
#> 9      0    8          potential            <=  68.5  10  11      12
#> 10     0    9               <NA>          <NA>    NA  NA  NA      NA
#> 11     0   10               <NA>          <NA>    NA  NA  NA      NA
#> 12     0   11               <NA>          <NA>    NA  NA  NA      NA
#> 13     0   12               <NA>          <NA>    NA  NA  NA      NA
#> 14     1    0            overall            <=  66.5  15  22      26
#> 15     1    1          potential            <=  72.5  16  17      21
#> 16     1    2               <NA>          <NA>    NA  NA  NA      NA
#> 17     1    3            overall            <=  63.5  18  19      20
#> 18     1    4               <NA>          <NA>    NA  NA  NA      NA
#> 19     1    5               <NA>          <NA>    NA  NA  NA      NA
#> 20     1    6               <NA>          <NA>    NA  NA  NA      NA
#> 21     1    7               <NA>          <NA>    NA  NA  NA      NA
#> 22     1    8          potential            <=  69.5  23  24      25
#> 23     1    9               <NA>          <NA>    NA  NA  NA      NA
#> 24     1   10               <NA>          <NA>    NA  NA  NA      NA
#> 25     1   11               <NA>          <NA>    NA  NA  NA      NA
#> 26     1   12               <NA>          <NA>    NA  NA  NA      NA
#> 27     2    0            overall            <=  67.5  28  35      39
#> 28     2    1            overall            <=  63.5  29  30      34
#> 29     2    2               <NA>          <NA>    NA  NA  NA      NA
#> 30     2    3                age            <=  23.5  31  32      33
#> 31     2    4               <NA>          <NA>    NA  NA  NA      NA
#> 32     2    5               <NA>          <NA>    NA  NA  NA      NA
#> 33     2    6               <NA>          <NA>    NA  NA  NA      NA
#> 34     2    7               <NA>          <NA>    NA  NA  NA      NA
#> 35     2    8          potential            <=  71.5  36  37      38
#> 36     2    9               <NA>          <NA>    NA  NA  NA      NA
#> 37     2   10               <NA>          <NA>    NA  NA  NA      NA
#> 38     2   11               <NA>          <NA>    NA  NA  NA      NA
#> 39     2   12               <NA>          <NA>    NA  NA  NA      NA
#> 40     3    0            overall            <=  69.5  41  51      52
#> 41     3    1            overall            <=  64.5  42  46      50
#> 42     3    2          potential            <=  75.5  43  44      45
#> 43     3    3               <NA>          <NA>    NA  NA  NA      NA
#> 44     3    4               <NA>          <NA>    NA  NA  NA      NA
#> 45     3    5               <NA>          <NA>    NA  NA  NA      NA
#> 46     3    6          defending            <=  59.5  47  48      49
#> 47     3    7               <NA>          <NA>    NA  NA  NA      NA
#> 48     3    8               <NA>          <NA>    NA  NA  NA      NA
#> 49     3    9               <NA>          <NA>    NA  NA  NA      NA
#> 50     3   10               <NA>          <NA>    NA  NA  NA      NA
#> 51     3   11               <NA>          <NA>    NA  NA  NA      NA
#> 52     3   12               <NA>          <NA>    NA  NA  NA      NA
#> 53     4    0            overall            <=  69.5  54  61      65
#> 54     4    1            overall            <=  63.5  55  56      60
#> 55     4    2               <NA>          <NA>    NA  NA  NA      NA
#> 56     4    3                age            <=  23.5  57  58      59
#> 57     4    4               <NA>          <NA>    NA  NA  NA      NA
#> 58     4    5               <NA>          <NA>    NA  NA  NA      NA
#> 59     4    6               <NA>          <NA>    NA  NA  NA      NA
#> 60     4    7               <NA>          <NA>    NA  NA  NA      NA
#> 61     4    8                age            <=  32.5  62  63      64
#> 62     4    9               <NA>          <NA>    NA  NA  NA      NA
#> 63     4   10               <NA>          <NA>    NA  NA  NA      NA
#> 64     4   11               <NA>          <NA>    NA  NA  NA      NA
#> 65     4   12               <NA>          <NA>    NA  NA  NA      NA
#> 66     5    0            overall            <=  70.5  67  77      78
#> 67     5    1            overall            <=  62.5  68  69      76
#> 68     5    2               <NA>          <NA>    NA  NA  NA      NA
#> 69     5    3          potential            <=  66.5  70  71      75
#> 70     5    4               <NA>          <NA>    NA  NA  NA      NA
#> 71     5    5          defending            <=  58.5  72  73      74
#> 72     5    6               <NA>          <NA>    NA  NA  NA      NA
#> 73     5    7               <NA>          <NA>    NA  NA  NA      NA
#> 74     5    8               <NA>          <NA>    NA  NA  NA      NA
#> 75     5    9               <NA>          <NA>    NA  NA  NA      NA
#> 76     5   10               <NA>          <NA>    NA  NA  NA      NA
#> 77     5   11               <NA>          <NA>    NA  NA  NA      NA
#> 78     5   12               <NA>          <NA>    NA  NA  NA      NA
#> 79     6    0            overall            <=  68.5  80  87      91
#> 80     6    1            overall            <=  62.5  81  82      86
#> 81     6    2               <NA>          <NA>    NA  NA  NA      NA
#> 82     6    3          potential            <=  66.5  83  84      85
#> 83     6    4               <NA>          <NA>    NA  NA  NA      NA
#> 84     6    5               <NA>          <NA>    NA  NA  NA      NA
#> 85     6    6               <NA>          <NA>    NA  NA  NA      NA
#> 86     6    7               <NA>          <NA>    NA  NA  NA      NA
#> 87     6    8          potential            <=  72.5  88  89      90
#> 88     6    9               <NA>          <NA>    NA  NA  NA      NA
#> 89     6   10               <NA>          <NA>    NA  NA  NA      NA
#> 90     6   11               <NA>          <NA>    NA  NA  NA      NA
#> 91     6   12               <NA>          <NA>    NA  NA  NA      NA
#> 92     7    0            overall            <=  70.5  93 103     104
#> 93     7    1            overall            <=  61.5  94  95     102
#> 94     7    2               <NA>          <NA>    NA  NA  NA      NA
#> 95     7    3                age            <=  21.5  96  97     101
#> 96     7    4               <NA>          <NA>    NA  NA  NA      NA
#> 97     7    5          dribbling            <=  64.5  98  99     100
#> 98     7    6               <NA>          <NA>    NA  NA  NA      NA
#> 99     7    7               <NA>          <NA>    NA  NA  NA      NA
#> 100    7    8               <NA>          <NA>    NA  NA  NA      NA
#> 101    7    9               <NA>          <NA>    NA  NA  NA      NA
#> 102    7   10               <NA>          <NA>    NA  NA  NA      NA
#> 103    7   11               <NA>          <NA>    NA  NA  NA      NA
#> 104    7   12               <NA>          <NA>    NA  NA  NA      NA
#> 105    8    0          potential            <=  74.5 106 113     117
#> 106    8    1           shooting            <=  59.5 107 111     112
#> 107    8    2            overall            <=  64.5 108 109     110
#> 108    8    3               <NA>          <NA>    NA  NA  NA      NA
#> 109    8    4               <NA>          <NA>    NA  NA  NA      NA
#> 110    8    5               <NA>          <NA>    NA  NA  NA      NA
#> 111    8    6               <NA>          <NA>    NA  NA  NA      NA
#> 112    8    7               <NA>          <NA>    NA  NA  NA      NA
#> 113    8    8            overall            <=  60.5 114 115     116
#> 114    8    9               <NA>          <NA>    NA  NA  NA      NA
#> 115    8   10               <NA>          <NA>    NA  NA  NA      NA
#> 116    8   11               <NA>          <NA>    NA  NA  NA      NA
#> 117    8   12               <NA>          <NA>    NA  NA  NA      NA
#> 118    9    0            overall            <=  71.5 119 129     130
#> 119    9    1            overall            <=  63.5 120 121     128
#> 120    9    2               <NA>          <NA>    NA  NA  NA      NA
#> 121    9    3                age            <=  30.5 122 126     127
#> 122    9    4          defending            <=  57.5 123 124     125
#> 123    9    5               <NA>          <NA>    NA  NA  NA      NA
#> 124    9    6               <NA>          <NA>    NA  NA  NA      NA
#> 125    9    7               <NA>          <NA>    NA  NA  NA      NA
#> 126    9    8               <NA>          <NA>    NA  NA  NA      NA
#> 127    9    9               <NA>          <NA>    NA  NA  NA      NA
#> 128    9   10               <NA>          <NA>    NA  NA  NA      NA
#> 129    9   11               <NA>          <NA>    NA  NA  NA      NA
#> 130    9   12               <NA>          <NA>    NA  NA  NA      NA
#> 131   10    0            overall            <=  72.5 132 142     143
#> 132   10    1 skill_ball_control            <=  61.5 133 134     141
#> 133   10    2               <NA>          <NA>    NA  NA  NA      NA
#> 134   10    3                age            <=  30.5 135 139     140
#> 135   10    4          potential            <=  65.5 136 137     138
#> 136   10    5               <NA>          <NA>    NA  NA  NA      NA
#> 137   10    6               <NA>          <NA>    NA  NA  NA      NA
#> 138   10    7               <NA>          <NA>    NA  NA  NA      NA
#> 139   10    8               <NA>          <NA>    NA  NA  NA      NA
#> 140   10    9               <NA>          <NA>    NA  NA  NA      NA
#> 141   10   10               <NA>          <NA>    NA  NA  NA      NA
#> 142   10   11               <NA>          <NA>    NA  NA  NA      NA
#> 143   10   12               <NA>          <NA>    NA  NA  NA      NA
#> 144   11    0            overall            <=  72.5 145 155     156
#> 145   11    1            overall            <=  61.5 146 147     154
#> 146   11    2               <NA>          <NA>    NA  NA  NA      NA
#> 147   11    3                age            <=  30.5 148 152     153
#> 148   11    4          defending            <=  53.5 149 150     151
#> 149   11    5               <NA>          <NA>    NA  NA  NA      NA
#> 150   11    6               <NA>          <NA>    NA  NA  NA      NA
#> 151   11    7               <NA>          <NA>    NA  NA  NA      NA
#> 152   11    8               <NA>          <NA>    NA  NA  NA      NA
#> 153   11    9               <NA>          <NA>    NA  NA  NA      NA
#> 154   11   10               <NA>          <NA>    NA  NA  NA      NA
#> 155   11   11               <NA>          <NA>    NA  NA  NA      NA
#> 156   11   12               <NA>          <NA>    NA  NA  NA      NA
#> 157   12    0            overall            <=  73.5 158 168     169
#> 158   12    1            overall            <=  59.5 159 160     167
#> 159   12    2               <NA>          <NA>    NA  NA  NA      NA
#> 160   12    3          potential            <=  65.5 161 162     166
#> 161   12    4               <NA>          <NA>    NA  NA  NA      NA
#> 162   12    5          defending            <=  54.5 163 164     165
#> 163   12    6               <NA>          <NA>    NA  NA  NA      NA
#> 164   12    7               <NA>          <NA>    NA  NA  NA      NA
#> 165   12    8               <NA>          <NA>    NA  NA  NA      NA
#> 166   12    9               <NA>          <NA>    NA  NA  NA      NA
#> 167   12   10               <NA>          <NA>    NA  NA  NA      NA
#> 168   12   11               <NA>          <NA>    NA  NA  NA      NA
#> 169   12   12               <NA>          <NA>    NA  NA  NA      NA
#> 170   13    0            overall            <=  71.5 171 178     182
#> 171   13    1            overall            <=  61.5 172 173     177
#> 172   13    2               <NA>          <NA>    NA  NA  NA      NA
#> 173   13    3          potential            <=  65.5 174 175     176
#> 174   13    4               <NA>          <NA>    NA  NA  NA      NA
#> 175   13    5               <NA>          <NA>    NA  NA  NA      NA
#> 176   13    6               <NA>          <NA>    NA  NA  NA      NA
#> 177   13    7               <NA>          <NA>    NA  NA  NA      NA
#> 178   13    8          potential            <=  73.5 179 180     181
#> 179   13    9               <NA>          <NA>    NA  NA  NA      NA
#> 180   13   10               <NA>          <NA>    NA  NA  NA      NA
#> 181   13   11               <NA>          <NA>    NA  NA  NA      NA
#> 182   13   12               <NA>          <NA>    NA  NA  NA      NA
#> 183   14    0          potential            <=  77.5 184 194     195
#> 184   14    1            overall            <=  59.5 185 186     193
#> 185   14    2               <NA>          <NA>    NA  NA  NA      NA
#> 186   14    3          potential            <=  65.5 187 188     192
#> 187   14    4               <NA>          <NA>    NA  NA  NA      NA
#> 188   14    5                age            <=  31.5 189 190     191
#> 189   14    6               <NA>          <NA>    NA  NA  NA      NA
#> 190   14    7               <NA>          <NA>    NA  NA  NA      NA
#> 191   14    8               <NA>          <NA>    NA  NA  NA      NA
#> 192   14    9               <NA>          <NA>    NA  NA  NA      NA
#> 193   14   10               <NA>          <NA>    NA  NA  NA      NA
#> 194   14   11               <NA>          <NA>    NA  NA  NA      NA
#> 195   14   12               <NA>          <NA>    NA  NA  NA      NA
#> 196   15    0            overall            <=  73.5 197 207     208
#> 197   15    1            overall            <=  59.5 198 199     206
#> 198   15    2               <NA>          <NA>    NA  NA  NA      NA
#> 199   15    3           shooting            <=  52.5 200 201     205
#> 200   15    4               <NA>          <NA>    NA  NA  NA      NA
#> 201   15    5                age            <=  31.5 202 203     204
#> 202   15    6               <NA>          <NA>    NA  NA  NA      NA
#> 203   15    7               <NA>          <NA>    NA  NA  NA      NA
#> 204   15    8               <NA>          <NA>    NA  NA  NA      NA
#> 205   15    9               <NA>          <NA>    NA  NA  NA      NA
#> 206   15   10               <NA>          <NA>    NA  NA  NA      NA
#> 207   15   11               <NA>          <NA>    NA  NA  NA      NA
#> 208   15   12               <NA>          <NA>    NA  NA  NA      NA
#> 209   16    0            overall            <=  73.5 210 220     221
#> 210   16    1            overall            <=  59.5 211 212     219
#> 211   16    2               <NA>          <NA>    NA  NA  NA      NA
#> 212   16    3          potential            <=  65.5 213 214     218
#> 213   16    4               <NA>          <NA>    NA  NA  NA      NA
#> 214   16    5          defending            <=  63.5 215 216     217
#> 215   16    6               <NA>          <NA>    NA  NA  NA      NA
#> 216   16    7               <NA>          <NA>    NA  NA  NA      NA
#> 217   16    8               <NA>          <NA>    NA  NA  NA      NA
#> 218   16    9               <NA>          <NA>    NA  NA  NA      NA
#> 219   16   10               <NA>          <NA>    NA  NA  NA      NA
#> 220   16   11               <NA>          <NA>    NA  NA  NA      NA
#> 221   16   12               <NA>          <NA>    NA  NA  NA      NA
#> 222   17    0            overall            <=  74.5 223 233     234
#> 223   17    1            overall            <=  59.5 224 225     232
#> 224   17    2               <NA>          <NA>    NA  NA  NA      NA
#> 225   17    3                age            <=  30.5 226 230     231
#> 226   17    4           shooting            <=  63.5 227 228     229
#> 227   17    5               <NA>          <NA>    NA  NA  NA      NA
#> 228   17    6               <NA>          <NA>    NA  NA  NA      NA
#> 229   17    7               <NA>          <NA>    NA  NA  NA      NA
#> 230   17    8               <NA>          <NA>    NA  NA  NA      NA
#> 231   17    9               <NA>          <NA>    NA  NA  NA      NA
#> 232   17   10               <NA>          <NA>    NA  NA  NA      NA
#> 233   17   11               <NA>          <NA>    NA  NA  NA      NA
#> 234   17   12               <NA>          <NA>    NA  NA  NA      NA
#> 235   18    0            overall            <=  60.5 236 237     247
#> 236   18    1               <NA>          <NA>    NA  NA  NA      NA
#> 237   18    2          dribbling            <=  71.5 238 245     246
#> 238   18    3                age            <=  31.5 239 243     244
#> 239   18    4          potential            <=  64.5 240 241     242
#> 240   18    5               <NA>          <NA>    NA  NA  NA      NA
#> 241   18    6               <NA>          <NA>    NA  NA  NA      NA
#> 242   18    7               <NA>          <NA>    NA  NA  NA      NA
#> 243   18    8               <NA>          <NA>    NA  NA  NA      NA
#> 244   18    9               <NA>          <NA>    NA  NA  NA      NA
#> 245   18   10               <NA>          <NA>    NA  NA  NA      NA
#> 246   18   11               <NA>          <NA>    NA  NA  NA      NA
#> 247   18   12               <NA>          <NA>    NA  NA  NA      NA
#> 248   19    0            overall            <=  58.5 249 250     260
#> 249   19    1               <NA>          <NA>    NA  NA  NA      NA
#> 250   19    2          potential            <=  65.5 251 252     259
#> 251   19    3               <NA>          <NA>    NA  NA  NA      NA
#> 252   19    4            overall            <=  74.5 253 257     258
#> 253   19    5                age            <=  32.5 254 255     256
#> 254   19    6               <NA>          <NA>    NA  NA  NA      NA
#> 255   19    7               <NA>          <NA>    NA  NA  NA      NA
#> 256   19    8               <NA>          <NA>    NA  NA  NA      NA
#> 257   19    9               <NA>          <NA>    NA  NA  NA      NA
#> 258   19   10               <NA>          <NA>    NA  NA  NA      NA
#> 259   19   11               <NA>          <NA>    NA  NA  NA      NA
#> 260   19   12               <NA>          <NA>    NA  NA  NA      NA
#>      Prediction Cover
#> 1            NA   501
#> 2            NA     0
#> 3    -6000.0000     0
#> 4            NA     0
#> 5     7500.0000     0
#> 6    42500.0000     0
#> 7    24425.0986     0
#> 8     -496.4345     0
#> 9            NA   501
#> 10   27500.0000     0
#> 11  205000.0000   501
#> 12  178691.7072     0
#> 13   96205.1647     0
#> 14           NA   501
#> 15           NA     0
#> 16    -900.0000     0
#> 17           NA     0
#> 18   12750.0000     0
#> 19   45500.0000     0
#> 20   26593.6681     0
#> 21    5748.4161     0
#> 22           NA   501
#> 23   35500.0000     0
#> 24  248000.0000   501
#> 25  213017.5976     0
#> 26  105629.5273     0
#> 27           NA   501
#> 28           NA     0
#> 29   -3310.0000     0
#> 30           NA     0
#> 31   40700.0000     0
#> 32   24200.0000     0
#> 33   29774.2240     0
#> 34   10530.0884     0
#> 35           NA   501
#> 36   43200.0000     0
#> 37  336700.0000   501
#> 38  257742.7451     0
#> 39  113997.4713     0
#> 40           NA   501
#> 41           NA     0
#> 42           NA     0
#> 43    1521.0000     0
#> 44   33806.0000     0
#> 45    4690.5727     0
#> 46           NA     0
#> 47   41270.0000     0
#> 48   28880.0000     0
#> 49   12380.0000     0
#> 50   16601.6777     0
#> 51  316530.0000   501
#> 52  109937.5075     0
#> 53           NA   501
#> 54           NA     0
#> 55    3868.9000     0
#> 56           NA     0
#> 57   40977.9000     0
#> 58   27117.9000     0
#> 59   31460.9870     0
#> 60   17976.0007     0
#> 61           NA   501
#> 62  338377.0000   435
#> 63   18377.0000    66
#> 64  303121.8082     0
#> 65  106618.0008     0
#> 66           NA   501
#> 67           NA     0
#> 68    2982.0100     0
#> 69           NA     0
#> 70   20406.1100     0
#> 71           NA     0
#> 72   42880.1100     0
#> 73   32982.0100     0
#> 74   19830.2100     0
#> 75   32601.4047     0
#> 76   21186.0532     0
#> 77  338039.3000   501
#> 78  103875.0775     0
#> 79           NA   501
#> 80           NA     0
#> 81    6183.8090     0
#> 82           NA     0
#> 83   21865.4990     0
#> 84   36197.1890     0
#> 85   32402.9775     0
#> 86   20620.2009     0
#> 87           NA   501
#> 88   36232.0090     0
#> 89  307735.3700   501
#> 90  229016.4200     0
#> 91   96987.3021     0
#> 92           NA   501
#> 93           NA     0
#> 94    7065.4281     0
#> 95           NA     0
#> 96   44247.4801     0
#> 97           NA     0
#> 98   27093.0491     0
#> 99   34204.0491     0
#> 100  16243.0491     0
#> 101  31546.5343     0
#> 102  23836.7160     0
#> 103 280461.8330   501
#> 104  89740.9888     0
#> 105          NA   501
#> 106          NA     0
#> 107          NA     0
#> 108  15856.1232     0
#> 109  28152.9852     0
#> 110  20813.6133     0
#> 111  37789.0752     0
#> 112   7858.8853     0
#> 113          NA   501
#> 114  17779.6353     0
#> 115 189053.0040   501
#> 116 178420.9790     0
#> 117  70995.1731     0
#> 118          NA   501
#> 119          NA     0
#> 120  16770.5109     0
#> 121          NA     0
#> 122          NA     0
#> 123  38966.3629     0
#> 124  30768.2496     0
#> 125  22768.7557     0
#> 126   8568.4457     0
#> 127  29154.0945     0
#> 128  23856.0805     0
#> 129 355636.7422   501
#> 130  98351.5305     0
#> 131          NA   501
#> 132          NA     0
#> 133  18596.2219     0
#> 134          NA     0
#> 135          NA     0
#> 136  27071.7808     0
#> 137  37365.7181     0
#> 138  36471.0269     0
#> 139  10836.6211     0
#> 140  32771.5080     0
#> 141  25438.8764     0
#> 142 408446.6751   501
#> 143  94421.3511     0
#> 144          NA   501
#> 145          NA     0
#> 146  15236.5997     0
#> 147          NA     0
#> 148          NA     0
#> 149  38371.5286     0
#> 150  32691.2395     0
#> 151  26208.5184     0
#> 152  10622.4710     0
#> 153  30413.0168     0
#> 154  25985.8002     0
#> 155 371102.0076   501
#> 156  88898.9856     0
#> 157          NA   501
#> 158          NA     0
#> 159  13212.9397     0
#> 160          NA     0
#> 161  23997.7095     0
#> 162          NA     0
#> 163  38075.9976     0
#> 164  31963.3756     0
#> 165  25012.6635     0
#> 166  32222.1127     0
#> 167  28522.9121     0
#> 168 487491.8068   501
#> 169  94965.1758     0
#> 170          NA   501
#> 171          NA     0
#> 172  18391.6458     0
#> 173          NA     0
#> 174  25754.3637     0
#> 175  34052.9330     0
#> 176  32860.7816     0
#> 177  28479.0695     0
#> 178          NA   501
#> 179   1984.9797     0
#> 180 340019.0477   501
#> 181 281678.1426     0
#> 182  83058.5830     0
#> 183          NA   501
#> 184          NA     0
#> 185  17052.4812     0
#> 186          NA     0
#> 187  25998.2532     0
#> 188          NA     0
#> 189  35650.9137     0
#> 190  -5033.5395     0
#> 191  30544.9253     0
#> 192  29834.4848     0
#> 193  27437.3353     0
#> 194 219402.0836   501
#> 195  59238.9141     0
#> 196          NA   501
#> 197          NA     0
#> 198  18711.3819     0
#> 199          NA     0
#> 200  30797.9914     0
#> 201          NA     0
#> 202  37233.5759     0
#> 203  10112.3539     0
#> 204  34171.5025     0
#> 205  21915.0092     0
#> 206  29110.4435     0
#> 207 393300.5130   501
#> 208  81154.6749     0
#> 209          NA   501
#> 210          NA     0
#> 211  20476.0949     0
#> 212          NA     0
#> 213  27982.4017     0
#> 214          NA     0
#> 215  36532.5471     0
#> 216  26285.5942     0
#> 217  25971.2026     0
#> 218  31883.8132     0
#> 219  29733.9195     0
#> 220 357470.4617   501
#> 221  77178.4370     0
#> 222          NA   501
#> 223          NA     0
#> 224  21728.2092     0
#> 225          NA     0
#> 226          NA     0
#> 227  33659.2859     0
#> 228  47667.3570     0
#> 229  29396.1729     0
#> 230  14402.3439     0
#> 231  32376.2740     0
#> 232  30422.3429     0
#> 233 475223.4155   501
#> 234  83473.3516     0
#> 235          NA   501
#> 236  24255.6645     0
#> 237          NA   501
#> 238          NA   109
#> 239          NA    83
#> 240  29487.3121     0
#> 241  36019.2814    83
#> 242  35515.5011     0
#> 243  13413.6766    26
#> 244  32985.4126     0
#> 245 103644.6363   336
#> 246  26750.0376    56
#> 247  41743.0206     0
#> 248          NA   501
#> 249  23129.8218     0
#> 250          NA   501
#> 251  29215.4820     0
#> 252          NA   501
#> 253          NA     0
#> 254  35286.5018     0
#> 255   -782.7467     0
#> 256  32992.9748     0
#> 257 381099.1458   501
#> 258  86567.0819     0
#> 259  79350.2244     0
#> 260  71918.9686     0
# }