Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node   Feature Decision.type Split Yes  No Missing Prediction Cover
#> 1      0    0   overall            <=  65.5   2   9      13         NA   501
#> 2      0    1 potential            <=  73.5   3   4       8         NA     0
#> 3      0    2      <NA>          <NA>    NA  NA  NA      NA  -7000.000     0
#> 4      0    3   overall            <=  62.5   5   6       7         NA     0
#> 5      0    4      <NA>          <NA>    NA  NA  NA      NA   7500.000     0
#> 6      0    5      <NA>          <NA>    NA  NA  NA      NA  45000.000     0
#> 7      0    6      <NA>          <NA>    NA  NA  NA      NA  24841.897     0
#> 8      0    7      <NA>          <NA>    NA  NA  NA      NA  -1245.714     0
#> 9      0    8 potential            <=  68.5  10  11      12         NA   501
#> 10     0    9      <NA>          <NA>    NA  NA  NA      NA  27500.000     0
#> 11     0   10      <NA>          <NA>    NA  NA  NA      NA 205000.000   501
#> 12     0   11      <NA>          <NA>    NA  NA  NA      NA 177938.348     0
#> 13     0   12      <NA>          <NA>    NA  NA  NA      NA  95590.929     0
#> 14     1    0   overall            <=  66.5  15  22      26         NA   501
#> 15     1    1 potential            <=  71.5  16  17      21         NA     0
#> 16     1    2      <NA>          <NA>    NA  NA  NA      NA  -1800.000     0
#> 17     1    3   overall            <=  63.5  18  19      20         NA     0
#> 18     1    4      <NA>          <NA>    NA  NA  NA      NA  12750.000     0
#> 19     1    5      <NA>          <NA>    NA  NA  NA      NA  44000.000     0
#> 20     1    6      <NA>          <NA>    NA  NA  NA      NA  25804.363     0
#> 21     1    7      <NA>          <NA>    NA  NA  NA      NA   6357.028     0
#> 22     1    8 potential            <=  69.5  23  24      25         NA   501
#> 23     1    9      <NA>          <NA>    NA  NA  NA      NA  35500.000     0
#> 24     1   10      <NA>          <NA>    NA  NA  NA      NA 238000.000   501
#> 25     1   11      <NA>          <NA>    NA  NA  NA      NA 203178.085     0
#> 26     1   12      <NA>          <NA>    NA  NA  NA      NA 101289.430     0
#> 27     2    0   overall            <=  67.5  28  35      39         NA   501
#> 28     2    1   overall            <=  63.5  29  30      34         NA     0
#> 29     2    2      <NA>          <NA>    NA  NA  NA      NA  -3120.000     0
#> 30     2    3       age            <=  24.5  31  32      33         NA     0
#> 31     2    4      <NA>          <NA>    NA  NA  NA      NA  38300.000     0
#> 32     2    5      <NA>          <NA>    NA  NA  NA      NA  22880.000     0
#> 33     2    6      <NA>          <NA>    NA  NA  NA      NA  29591.996     0
#> 34     2    7      <NA>          <NA>    NA  NA  NA      NA  10872.237     0
#> 35     2    8       age            <=  32.5  36  37      38         NA   501
#> 36     2    9      <NA>          <NA>    NA  NA  NA      NA 237700.000   435
#> 37     2   10      <NA>          <NA>    NA  NA  NA      NA  27700.000    66
#> 38     2   11      <NA>          <NA>    NA  NA  NA      NA 213785.017     0
#> 39     2   12      <NA>          <NA>    NA  NA  NA      NA  95487.688     0
#> 40     3    0   overall            <=  69.5  41  51      52         NA   501
#> 41     3    1 potential            <=  74.5  42  46      50         NA     0
#> 42     3    2   overall            <=  64.5  43  44      45         NA     0
#> 43     3    3      <NA>          <NA>    NA  NA  NA      NA   1692.000     0
#> 44     3    4      <NA>          <NA>    NA  NA  NA      NA  26592.000     0
#> 45     3    5      <NA>          <NA>    NA  NA  NA      NA  11737.453     0
#> 46     3    6   overall            <=  61.5  47  48      49         NA     0
#> 47     3    7      <NA>          <NA>    NA  NA  NA      NA  11287.000     0
#> 48     3    8      <NA>          <NA>    NA  NA  NA      NA  47370.000     0
#> 49     3    9      <NA>          <NA>    NA  NA  NA      NA  39436.611     0
#> 50     3   10      <NA>          <NA>    NA  NA  NA      NA  16309.886     0
#> 51     3   11      <NA>          <NA>    NA  NA  NA      NA 317430.000   501
#> 52     3   12      <NA>          <NA>    NA  NA  NA      NA 110444.930     0
#> 53     4    0   overall            <=  69.5  54  64      65         NA   501
#> 54     4    1 potential            <=  75.5  55  62      63         NA     0
#> 55     4    2   overall            <=  63.5  56  57      61         NA     0
#> 56     4    3      <NA>          <NA>    NA  NA  NA      NA   2022.800     0
#> 57     4    4 defending            <=  55.5  58  59      60         NA     0
#> 58     4    5      <NA>          <NA>    NA  NA  NA      NA  36270.800     0
#> 59     4    6      <NA>          <NA>    NA  NA  NA      NA  23982.800     0
#> 60     4    7      <NA>          <NA>    NA  NA  NA      NA  10971.800     0
#> 61     4    8      <NA>          <NA>    NA  NA  NA      NA  14455.788     0
#> 62     4    9      <NA>          <NA>    NA  NA  NA      NA  45533.000     0
#> 63     4   10      <NA>          <NA>    NA  NA  NA      NA  18226.724     0
#> 64     4   11      <NA>          <NA>    NA  NA  NA      NA 279187.000   501
#> 65     4   12      <NA>          <NA>    NA  NA  NA      NA  97751.220     0
#> 66     5    0 potential            <=  72.5  67  74      78         NA   501
#> 67     5    1  shooting            <=  60.5  68  72      73         NA     0
#> 68     5    2   overall            <=  64.5  69  70      71         NA     0
#> 69     5    3      <NA>          <NA>    NA  NA  NA      NA   6320.520     0
#> 70     5    4      <NA>          <NA>    NA  NA  NA      NA  24768.300     0
#> 71     5    5      <NA>          <NA>    NA  NA  NA      NA  13095.773     0
#> 72     5    6      <NA>          <NA>    NA  NA  NA      NA  33854.520     0
#> 73     5    7      <NA>          <NA>    NA  NA  NA      NA   1320.520     0
#> 74     5    8   overall            <=  62.5  75  76      77         NA   501
#> 75     5    9      <NA>          <NA>    NA  NA  NA      NA  17605.000     0
#> 76     5   10      <NA>          <NA>    NA  NA  NA      NA 194768.300   501
#> 77     5   11      <NA>          <NA>    NA  NA  NA      NA 171448.731     0
#> 78     5   12      <NA>          <NA>    NA  NA  NA      NA  80386.470     0
#> 79     6    0   overall            <=  70.5  80  87      91         NA   501
#> 80     6    1 dribbling            <=  65.5  81  82      86         NA     0
#> 81     6    2      <NA>          <NA>    NA  NA  NA      NA  18145.690     0
#> 82     6    3 potential            <=  77.5  83  84      85         NA     0
#> 83     6    4      <NA>          <NA>    NA  NA  NA      NA  29166.890     0
#> 84     6    5      <NA>          <NA>    NA  NA  NA      NA  59102.870     0
#> 85     6    6      <NA>          <NA>    NA  NA  NA      NA  33494.767     0
#> 86     6    7      <NA>          <NA>    NA  NA  NA      NA   5155.020     0
#> 87     6    8       age            <=  33.5  88  89      90         NA   501
#> 88     6    9      <NA>          <NA>    NA  NA  NA      NA 388791.470   462
#> 89     6   10      <NA>          <NA>    NA  NA  NA      NA   9791.470    39
#> 90     6   11      <NA>          <NA>    NA  NA  NA      NA 356089.094     0
#> 91     6   12      <NA>          <NA>    NA  NA  NA      NA 108514.618     0
#> 92     7    0   overall            <=  69.5  93 100     104         NA   501
#> 93     7    1   overall            <=  61.5  94  95      99         NA     0
#> 94     7    2      <NA>          <NA>    NA  NA  NA      NA   5672.966     0
#> 95     7    3 defending            <=  53.5  96  97      98         NA     0
#> 96     7    4      <NA>          <NA>    NA  NA  NA      NA  34349.099     0
#> 97     7    5      <NA>          <NA>    NA  NA  NA      NA  25373.899     0
#> 98     7    6      <NA>          <NA>    NA  NA  NA      NA  16280.627     0
#> 99     7    7      <NA>          <NA>    NA  NA  NA      NA  20203.564     0
#> 100    7    8 potential            <=  71.5 101 102     103         NA   501
#> 101    7    9      <NA>          <NA>    NA  NA  NA      NA   7476.901     0
#> 102    7   10      <NA>          <NA>    NA  NA  NA      NA 303412.323   501
#> 103    7   11      <NA>          <NA>    NA  NA  NA      NA 259707.882     0
#> 104    7   12      <NA>          <NA>    NA  NA  NA      NA  93137.202     0
#> 105    8    0   overall            <=  70.5 106 113     117         NA   501
#> 106    8    1   overall            <=  61.5 107 108     112         NA     0
#> 107    8    2      <NA>          <NA>    NA  NA  NA      NA   8605.669     0
#> 108    8    3 potential            <=  77.5 109 110     111         NA     0
#> 109    8    4      <NA>          <NA>    NA  NA  NA      NA  26781.611     0
#> 110    8    5      <NA>          <NA>    NA  NA  NA      NA  47150.911     0
#> 111    8    6      <NA>          <NA>    NA  NA  NA      NA  28784.062     0
#> 112    8    7      <NA>          <NA>    NA  NA  NA      NA  22378.745     0
#> 113    8    8 potential            <=  72.5 114 115     116         NA   501
#> 114    8    9      <NA>          <NA>    NA  NA  NA      NA  33164.633     0
#> 115    8   10      <NA>          <NA>    NA  NA  NA      NA 381021.091   501
#> 116    8   11      <NA>          <NA>    NA  NA  NA      NA 325958.385     0
#> 117    8   12      <NA>          <NA>    NA  NA  NA      NA 101703.417     0
#> 118    9    0   overall            <=  71.5 119 126     130         NA   501
#> 119    9    1 dribbling            <=  63.5 120 121     125         NA     0
#> 120    9    2      <NA>          <NA>    NA  NA  NA      NA  20060.199     0
#> 121    9    3 potential            <=  77.5 122 123     124         NA     0
#> 122    9    4      <NA>          <NA>    NA  NA  NA      NA  30725.450     0
#> 123    9    5      <NA>          <NA>    NA  NA  NA      NA  55602.582     0
#> 124    9    6      <NA>          <NA>    NA  NA  NA      NA  33859.725     0
#> 125    9    7      <NA>          <NA>    NA  NA  NA      NA  10381.842     0
#> 126    9    8       age            <=  32.5 127 128     129         NA   501
#> 127    9    9      <NA>          <NA>    NA  NA  NA      NA 391968.982   435
#> 128    9   10      <NA>          <NA>    NA  NA  NA      NA  11746.005    66
#> 129    9   11      <NA>          <NA>    NA  NA  NA      NA 346235.066     0
#> 130    9   12      <NA>          <NA>    NA  NA  NA      NA  94032.173     0
#> 131   10    0   overall            <=  71.5 132 139     143         NA   501
#> 132   10    1   overall            <=  60.5 133 134     138         NA     0
#> 133   10    2      <NA>          <NA>    NA  NA  NA      NA  10940.016     0
#> 134   10    3       age            <=  29.5 135 136     137         NA     0
#> 135   10    4      <NA>          <NA>    NA  NA  NA      NA  33571.550     0
#> 136   10    5      <NA>          <NA>    NA  NA  NA      NA  14828.558     0
#> 137   10    6      <NA>          <NA>    NA  NA  NA      NA  29512.944     0
#> 138   10    7      <NA>          <NA>    NA  NA  NA      NA  24713.257     0
#> 139   10    8 potential            <=  73.5 140 141     142         NA   501
#> 140   10    9      <NA>          <NA>    NA  NA  NA      NA  27149.107     0
#> 141   10   10      <NA>          <NA>    NA  NA  NA      NA 406272.083   501
#> 142   10   11      <NA>          <NA>    NA  NA  NA      NA 335672.937     0
#> 143   10   12      <NA>          <NA>    NA  NA  NA      NA  94329.572     0
#> 144   11    0   overall            <=  72.5 145 155     156         NA   501
#> 145   11    1   overall            <=  60.5 146 147     154         NA     0
#> 146   11    2      <NA>          <NA>    NA  NA  NA      NA  13346.014     0
#> 147   11    3       age            <=  30.5 148 152     153         NA     0
#> 148   11    4  shooting            <=  57.5 149 150     151         NA     0
#> 149   11    5      <NA>          <NA>    NA  NA  NA      NA  31300.863     0
#> 150   11    6      <NA>          <NA>    NA  NA  NA      NA  39704.208     0
#> 151   11    7      <NA>          <NA>    NA  NA  NA      NA  24971.403     0
#> 152   11    8      <NA>          <NA>    NA  NA  NA      NA   9961.121     0
#> 153   11    9      <NA>          <NA>    NA  NA  NA      NA  30105.935     0
#> 154   11   10      <NA>          <NA>    NA  NA  NA      NA  25956.732     0
#> 155   11   11      <NA>          <NA>    NA  NA  NA      NA 369144.875   501
#> 156   11   12      <NA>          <NA>    NA  NA  NA      NA  87466.982     0
#> 157   12    0   overall            <=  73.5 158 168     169         NA   501
#> 158   12    1   overall            <=  60.5 159 160     167         NA     0
#> 159   12    2      <NA>          <NA>    NA  NA  NA      NA  15511.413     0
#> 160   12    3       age            <=  29.5 161 165     166         NA     0
#> 161   12    4  shooting            <=  63.5 162 163     164         NA     0
#> 162   12    5      <NA>          <NA>    NA  NA  NA      NA  33731.396     0
#> 163   12    6      <NA>          <NA>    NA  NA  NA      NA  48297.568     0
#> 164   12    7      <NA>          <NA>    NA  NA  NA      NA  25974.263     0
#> 165   12    8      <NA>          <NA>    NA  NA  NA      NA  15988.200     0
#> 166   12    9      <NA>          <NA>    NA  NA  NA      NA  31399.788     0
#> 167   12   10      <NA>          <NA>    NA  NA  NA      NA  27751.241     0
#> 168   12   11      <NA>          <NA>    NA  NA  NA      NA 485730.388   501
#> 169   12   12      <NA>          <NA>    NA  NA  NA      NA  94451.125     0
#> 170   13    0 potential            <=  76.5 171 181     182         NA   501
#> 171   13    1   overall            <=  59.5 172 173     180         NA     0
#> 172   13    2      <NA>          <NA>    NA  NA  NA      NA  15460.271     0
#> 173   13    3 potential            <=  65.5 174 175     179         NA     0
#> 174   13    4      <NA>          <NA>    NA  NA  NA      NA  24291.947     0
#> 175   13    5 defending            <=  63.5 176 177     178         NA     0
#> 176   13    6      <NA>          <NA>    NA  NA  NA      NA  36546.187     0
#> 177   13    7      <NA>          <NA>    NA  NA  NA      NA  24372.206     0
#> 178   13    8      <NA>          <NA>    NA  NA  NA      NA  23246.108     0
#> 179   13    9      <NA>          <NA>    NA  NA  NA      NA  30361.665     0
#> 180   13   10      <NA>          <NA>    NA  NA  NA      NA  27481.563     0
#> 181   13   11      <NA>          <NA>    NA  NA  NA      NA 190657.349   501
#> 182   13   12      <NA>          <NA>    NA  NA  NA      NA  61048.745     0
#> 183   14    0   overall            <=  73.5 184 194     195         NA   501
#> 184   14    1   overall            <=  59.5 185 186     193         NA     0
#> 185   14    2      <NA>          <NA>    NA  NA  NA      NA  17414.244     0
#> 186   14    3 potential            <=  65.5 187 188     192         NA     0
#> 187   14    4      <NA>          <NA>    NA  NA  NA      NA  25362.752     0
#> 188   14    5       age            <=  31.5 189 190     191         NA     0
#> 189   14    6      <NA>          <NA>    NA  NA  NA      NA  33894.754     0
#> 190   14    7      <NA>          <NA>    NA  NA  NA      NA   2261.259     0
#> 191   14    8      <NA>          <NA>    NA  NA  NA      NA  31014.220     0
#> 192   14    9      <NA>          <NA>    NA  NA  NA      NA  30206.868     0
#> 193   14   10      <NA>          <NA>    NA  NA  NA      NA  27789.029     0
#> 194   14   11      <NA>          <NA>    NA  NA  NA      NA 425091.614   501
#> 195   14   12      <NA>          <NA>    NA  NA  NA      NA  85130.326     0
#> 196   15    0   overall            <=  71.5 197 204     208         NA   501
#> 197   15    1   overall            <=  59.5 198 199     203         NA     0
#> 198   15    2      <NA>          <NA>    NA  NA  NA      NA  19172.820     0
#> 199   15    3 potential            <=  65.5 200 201     202         NA     0
#> 200   15    4      <NA>          <NA>    NA  NA  NA      NA  26530.477     0
#> 201   15    5      <NA>          <NA>    NA  NA  NA      NA  31747.853     0
#> 202   15    6      <NA>          <NA>    NA  NA  NA      NA  30912.925     0
#> 203   15    7      <NA>          <NA>    NA  NA  NA      NA  28484.791     0
#> 204   15    8 potential            <=  74.5 205 206     207         NA   501
#> 205   15    9      <NA>          <NA>    NA  NA  NA      NA  18727.251     0
#> 206   15   10      <NA>          <NA>    NA  NA  NA      NA 336082.453   501
#> 207   15   11      <NA>          <NA>    NA  NA  NA      NA 244353.111     0
#> 208   15   12      <NA>          <NA>    NA  NA  NA      NA  76127.465     0
#> 209   16    0   overall            <=  72.5 210 217     221         NA   501
#> 210   16    1   overall            <=  61.5 211 212     216         NA     0
#> 211   16    2      <NA>          <NA>    NA  NA  NA      NA  22759.064     0
#> 212   16    3       age            <=  30.5 213 214     215         NA     0
#> 213   16    4      <NA>          <NA>    NA  NA  NA      NA  33747.686     0
#> 214   16    5      <NA>          <NA>    NA  NA  NA      NA  17654.494     0
#> 215   16    6      <NA>          <NA>    NA  NA  NA      NA  31004.836     0
#> 216   16    7      <NA>          <NA>    NA  NA  NA      NA  28650.789     0
#> 217   16    8 potential            <=  74.5 218 219     220         NA   501
#> 218   16    9      <NA>          <NA>    NA  NA  NA      NA  14338.242     0
#> 219   16   10      <NA>          <NA>    NA  NA  NA      NA 355974.207   501
#> 220   16   11      <NA>          <NA>    NA  NA  NA      NA 287647.014     0
#> 221   16   12      <NA>          <NA>    NA  NA  NA      NA  75411.241     0
#> 222   17    0   overall            <=  73.5 223 233     234         NA   501
#> 223   17    1   overall            <=  59.5 224 225     232         NA     0
#> 224   17    2      <NA>          <NA>    NA  NA  NA      NA  21979.631     0
#> 225   17    3 potential            <=  64.5 226 227     231         NA     0
#> 226   17    4      <NA>          <NA>    NA  NA  NA      NA  26190.573     0
#> 227   17    5 defending            <=  57.5 228 229     230         NA     0
#> 228   17    6      <NA>          <NA>    NA  NA  NA      NA  37707.045     0
#> 229   17    7      <NA>          <NA>    NA  NA  NA      NA  29736.876     0
#> 230   17    8      <NA>          <NA>    NA  NA  NA      NA  22870.508     0
#> 231   17    9      <NA>          <NA>    NA  NA  NA      NA  31762.333     0
#> 232   17   10      <NA>          <NA>    NA  NA  NA      NA  29903.132     0
#> 233   17   11      <NA>          <NA>    NA  NA  NA      NA 323876.787   501
#> 234   17   12      <NA>          <NA>    NA  NA  NA      NA  72363.491     0
#> 235   18    0 dribbling            <=  71.5 236 246     247         NA   501
#> 236   18    1   overall            <=  58.5 237 238     245         NA   109
#> 237   18    2      <NA>          <NA>    NA  NA  NA      NA  22281.668     0
#> 238   18    3 potential            <=  64.5 239 240     244         NA   109
#> 239   18    4      <NA>          <NA>    NA  NA  NA      NA  27281.668     0
#> 240   18    5       age            <=  31.5 241 242     243         NA   109
#> 241   18    6      <NA>          <NA>    NA  NA  NA      NA  34711.545    83
#> 242   18    7      <NA>          <NA>    NA  NA  NA      NA  10137.136    26
#> 243   18    8      <NA>          <NA>    NA  NA  NA      NA  32432.775     0
#> 244   18    9      <NA>          <NA>    NA  NA  NA      NA  31890.695     0
#> 245   18   10      <NA>          <NA>    NA  NA  NA      NA  30514.499     0
#> 246   18   11      <NA>          <NA>    NA  NA  NA      NA 106469.050   336
#> 247   18   12      <NA>          <NA>    NA  NA  NA      NA  23048.571    56
#> 248   19    0 potential            <=  78.5 249 256     260         NA   501
#> 249   19    1   overall            <=  58.5 250 251     255         NA     0
#> 250   19    2      <NA>          <NA>    NA  NA  NA      NA  23553.501     0
#> 251   19    3 potential            <=  64.5 252 253     254         NA     0
#> 252   19    4      <NA>          <NA>    NA  NA  NA      NA  27324.053     0
#> 253   19    5      <NA>          <NA>    NA  NA  NA      NA  32723.627     0
#> 254   19    6      <NA>          <NA>    NA  NA  NA      NA  32148.229     0
#> 255   19    7      <NA>          <NA>    NA  NA  NA      NA  30835.718     0
#> 256   19    8   overall            <=  64.5 257 258     259         NA   501
#> 257   19    9      <NA>          <NA>    NA  NA  NA      NA  17884.923     0
#> 258   19   10      <NA>          <NA>    NA  NA  NA      NA 337842.203   501
#> 259   19   11      <NA>          <NA>    NA  NA  NA      NA 304232.133     0
#> 260   19   12      <NA>          <NA>    NA  NA  NA      NA  67003.380     0
# }