Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node   Feature Decision.type Split Yes  No Missing Prediction Cover
#> 1      0    0   overall            <=  65.5   2   9      13         NA   501
#> 2      0    1 potential            <=  73.5   3   4       8         NA     0
#> 3      0    2      <NA>          <NA>    NA  NA  NA      NA  -7000.000     0
#> 4      0    3   overall            <=  62.5   5   6       7         NA     0
#> 5      0    4      <NA>          <NA>    NA  NA  NA      NA   7500.000     0
#> 6      0    5      <NA>          <NA>    NA  NA  NA      NA  42500.000     0
#> 7      0    6      <NA>          <NA>    NA  NA  NA      NA  24765.416     0
#> 8      0    7      <NA>          <NA>    NA  NA  NA      NA  -1365.906     0
#> 9      0    8 potential            <=  68.5  10  11      12         NA   501
#> 10     0    9      <NA>          <NA>    NA  NA  NA      NA  27500.000     0
#> 11     0   10      <NA>          <NA>    NA  NA  NA      NA 205000.000   501
#> 12     0   11      <NA>          <NA>    NA  NA  NA      NA 179056.862     0
#> 13     0   12      <NA>          <NA>    NA  NA  NA      NA  96021.720     0
#> 14     1    0   overall            <=  66.5  15  22      26         NA   501
#> 15     1    1 potential            <=  71.5  16  17      21         NA     0
#> 16     1    2      <NA>          <NA>    NA  NA  NA      NA  -1800.000     0
#> 17     1    3   overall            <=  63.5  18  19      20         NA     0
#> 18     1    4      <NA>          <NA>    NA  NA  NA      NA  12750.000     0
#> 19     1    5      <NA>          <NA>    NA  NA  NA      NA  43000.000     0
#> 20     1    6      <NA>          <NA>    NA  NA  NA      NA  25226.233     0
#> 21     1    7      <NA>          <NA>    NA  NA  NA      NA   6096.763     0
#> 22     1    8 potential            <=  69.5  23  24      25         NA   501
#> 23     1    9      <NA>          <NA>    NA  NA  NA      NA  35500.000     0
#> 24     1   10      <NA>          <NA>    NA  NA  NA      NA 248000.000   501
#> 25     1   11      <NA>          <NA>    NA  NA  NA      NA 212103.080     0
#> 26     1   12      <NA>          <NA>    NA  NA  NA      NA 104174.614     0
#> 27     2    0   overall            <=  67.5  28  35      39         NA   501
#> 28     2    1   overall            <=  63.5  29  30      34         NA     0
#> 29     2    2      <NA>          <NA>    NA  NA  NA      NA  -3120.000     0
#> 30     2    3       age            <=  23.5  31  32      33         NA     0
#> 31     2    4      <NA>          <NA>    NA  NA  NA      NA  40950.000     0
#> 32     2    5      <NA>          <NA>    NA  NA  NA      NA  24430.000     0
#> 33     2    6      <NA>          <NA>    NA  NA  NA      NA  30125.779     0
#> 34     2    7      <NA>          <NA>    NA  NA  NA      NA  10795.814     0
#> 35     2    8       age            <=  33.5  36  37      38         NA   501
#> 36     2    9      <NA>          <NA>    NA  NA  NA      NA 226700.000   462
#> 37     2   10      <NA>          <NA>    NA  NA  NA      NA  15450.000    39
#> 38     2   11      <NA>          <NA>    NA  NA  NA      NA 211783.074     0
#> 39     2   12      <NA>          <NA>    NA  NA  NA      NA  95509.998     0
#> 40     3    0   overall            <=  69.5  41  51      52         NA   501
#> 41     3    1 potential            <=  74.5  42  46      50         NA     0
#> 42     3    2   overall            <=  64.5  43  44      45         NA     0
#> 43     3    3      <NA>          <NA>    NA  NA  NA      NA   1692.000     0
#> 44     3    4      <NA>          <NA>    NA  NA  NA      NA  26437.000     0
#> 45     3    5      <NA>          <NA>    NA  NA  NA      NA  11617.452     0
#> 46     3    6   overall            <=  61.5  47  48      49         NA     0
#> 47     3    7      <NA>          <NA>    NA  NA  NA      NA  12537.000     0
#> 48     3    8      <NA>          <NA>    NA  NA  NA      NA  46605.000     0
#> 49     3    9      <NA>          <NA>    NA  NA  NA      NA  39161.571     0
#> 50     3   10      <NA>          <NA>    NA  NA  NA      NA  16300.696     0
#> 51     3   11      <NA>          <NA>    NA  NA  NA      NA 317530.000   501
#> 52     3   12      <NA>          <NA>    NA  NA  NA      NA 109909.540     0
#> 53     4    0   overall            <=  69.5  54  64      65         NA   501
#> 54     4    1 potential            <=  75.5  55  59      63         NA     0
#> 55     4    2   overall            <=  63.5  56  57      58         NA     0
#> 56     4    3      <NA>          <NA>    NA  NA  NA      NA   1022.800     0
#> 57     4    4      <NA>          <NA>    NA  NA  NA      NA  27293.300     0
#> 58     4    5      <NA>          <NA>    NA  NA  NA      NA  14225.559     0
#> 59     4    6   overall            <=  61.5  60  61      62         NA     0
#> 60     4    7      <NA>          <NA>    NA  NA  NA      NA  18533.300     0
#> 61     4    8      <NA>          <NA>    NA  NA  NA      NA  49126.500     0
#> 62     4    9      <NA>          <NA>    NA  NA  NA      NA  43340.641     0
#> 63     4   10      <NA>          <NA>    NA  NA  NA      NA  17989.999     0
#> 64     4   11      <NA>          <NA>    NA  NA  NA      NA 289277.000   501
#> 65     4   12      <NA>          <NA>    NA  NA  NA      NA 103540.840     0
#> 66     5    0   overall            <=  70.5  67  77      78         NA   501
#> 67     5    1   overall            <=  62.5  68  69      76         NA     0
#> 68     5    2      <NA>          <NA>    NA  NA  NA      NA   3420.520     0
#> 69     5    3 potential            <=  73.5  70  74      75         NA     0
#> 70     5    4 defending            <=  54.5  71  72      73         NA     0
#> 71     5    5      <NA>          <NA>    NA  NA  NA      NA  35538.470     0
#> 72     5    6      <NA>          <NA>    NA  NA  NA      NA  23156.970     0
#> 73     5    7      <NA>          <NA>    NA  NA  NA      NA   9156.970     0
#> 74     5    8      <NA>          <NA>    NA  NA  NA      NA  44956.850     0
#> 75     5    9      <NA>          <NA>    NA  NA  NA      NA  30819.935     0
#> 76     5   10      <NA>          <NA>    NA  NA  NA      NA  20303.562     0
#> 77     5   11      <NA>          <NA>    NA  NA  NA      NA 353849.300   501
#> 78     5   12      <NA>          <NA>    NA  NA  NA      NA 107567.799     0
#> 79     6    0 potential            <=  71.5  80  87      91         NA   501
#> 80     6    1   overall            <=  63.5  81  82      86         NA     0
#> 81     6    2      <NA>          <NA>    NA  NA  NA      NA   5578.468     0
#> 82     6    3       age            <=  28.5  83  84      85         NA     0
#> 83     6    4      <NA>          <NA>    NA  NA  NA      NA  33795.453     0
#> 84     6    5      <NA>          <NA>    NA  NA  NA      NA  15033.603     0
#> 85     6    6      <NA>          <NA>    NA  NA  NA      NA  24624.661     0
#> 86     6    7      <NA>          <NA>    NA  NA  NA      NA  15628.169     0
#> 87     6    8   overall            <=  61.5  88  89      90         NA   501
#> 88     6    9      <NA>          <NA>    NA  NA  NA      NA  14123.468     0
#> 89     6   10      <NA>          <NA>    NA  NA  NA      NA 101964.370   501
#> 90     6   11      <NA>          <NA>    NA  NA  NA      NA  91685.328     0
#> 91     6   12      <NA>          <NA>    NA  NA  NA      NA  52254.447     0
#> 92     7    0   overall            <=  69.5  93 100     104         NA   501
#> 93     7    1 dribbling            <=  63.5  94  95      99         NA     0
#> 94     7    2      <NA>          <NA>    NA  NA  NA      NA  18188.728     0
#> 95     7    3 potential            <=  76.5  96  97      98         NA     0
#> 96     7    4      <NA>          <NA>    NA  NA  NA      NA  30668.728     0
#> 97     7    5      <NA>          <NA>    NA  NA  NA      NA  47264.728     0
#> 98     7    6      <NA>          <NA>    NA  NA  NA      NA  33309.916     0
#> 99     7    7      <NA>          <NA>    NA  NA  NA      NA   8723.096     0
#> 100    7    8 potential            <=  71.5 101 102     103         NA   501
#> 101    7    9      <NA>          <NA>    NA  NA  NA      NA   7030.243     0
#> 102    7   10      <NA>          <NA>    NA  NA  NA      NA 305267.933   501
#> 103    7   11      <NA>          <NA>    NA  NA  NA      NA 265181.094     0
#> 104    7   12      <NA>          <NA>    NA  NA  NA      NA  97603.515     0
#> 105    8    0   overall            <=  70.5 106 116     117         NA   501
#> 106    8    1   overall            <=  62.5 107 108     115         NA     0
#> 107    8    2      <NA>          <NA>    NA  NA  NA      NA  11148.312     0
#> 108    8    3 potential            <=  76.5 109 113     114         NA     0
#> 109    8    4 defending            <=  57.5 110 111     112         NA     0
#> 110    8    5      <NA>          <NA>    NA  NA  NA      NA  34801.855     0
#> 111    8    6      <NA>          <NA>    NA  NA  NA      NA  26049.855     0
#> 112    8    7      <NA>          <NA>    NA  NA  NA      NA  17196.418     0
#> 113    8    8      <NA>          <NA>    NA  NA  NA      NA  51363.255     0
#> 114    8    9      <NA>          <NA>    NA  NA  NA      NA  31533.340     0
#> 115    8   10      <NA>          <NA>    NA  NA  NA      NA  23641.585     0
#> 116    8   11      <NA>          <NA>    NA  NA  NA      NA 288241.140   501
#> 117    8   12      <NA>          <NA>    NA  NA  NA      NA  91999.124     0
#> 118    9    0   overall            <=  71.5 119 129     130         NA   501
#> 119    9    1   overall            <=  62.5 120 121     128         NA     0
#> 120    9    2      <NA>          <NA>    NA  NA  NA      NA  13840.674     0
#> 121    9    3       age            <=  29.5 122 126     127         NA     0
#> 122    9    4 defending            <=  54.5 123 124     125         NA     0
#> 123    9    5      <NA>          <NA>    NA  NA  NA      NA  39261.768     0
#> 124    9    6      <NA>          <NA>    NA  NA  NA      NA  32236.900     0
#> 125    9    7      <NA>          <NA>    NA  NA  NA      NA  22651.276     0
#> 126    9    8      <NA>          <NA>    NA  NA  NA      NA  12795.554     0
#> 127    9    9      <NA>          <NA>    NA  NA  NA      NA  29358.526     0
#> 128    9   10      <NA>          <NA>    NA  NA  NA      NA  23661.573     0
#> 129    9   11      <NA>          <NA>    NA  NA  NA      NA 362917.026   501
#> 130    9   12      <NA>          <NA>    NA  NA  NA      NA  99612.624     0
#> 131   10    0 potential            <=  74.5 132 139     143         NA   501
#> 132   10    1  shooting            <=  59.5 133 134     138         NA     0
#> 133   10    2      <NA>          <NA>    NA  NA  NA      NA  22066.179     0
#> 134   10    3       age            <=  30.5 135 136     137         NA     0
#> 135   10    4      <NA>          <NA>    NA  NA  NA      NA  38192.778     0
#> 136   10    5      <NA>          <NA>    NA  NA  NA      NA  10125.323     0
#> 137   10    6      <NA>          <NA>    NA  NA  NA      NA  30774.218     0
#> 138   10    7      <NA>          <NA>    NA  NA  NA      NA  10649.413     0
#> 139   10    8   overall            <=  68.5 140 141     142         NA   501
#> 140   10    9      <NA>          <NA>    NA  NA  NA      NA  34107.800     0
#> 141   10   10      <NA>          <NA>    NA  NA  NA      NA 380125.323   501
#> 142   10   11      <NA>          <NA>    NA  NA  NA      NA 265910.799     0
#> 143   10   12      <NA>          <NA>    NA  NA  NA      NA  96538.143     0
#> 144   11    0   overall            <=  71.5 145 155     156         NA   501
#> 145   11    1   overall            <=  60.5 146 147     154         NA     0
#> 146   11    2      <NA>          <NA>    NA  NA  NA      NA  12996.232     0
#> 147   11    3 defending            <=  54.5 148 152     153         NA     0
#> 148   11    4 potential            <=  65.5 149 150     151         NA     0
#> 149   11    5      <NA>          <NA>    NA  NA  NA      NA  26083.371     0
#> 150   11    6      <NA>          <NA>    NA  NA  NA      NA  38204.642     0
#> 151   11    7      <NA>          <NA>    NA  NA  NA      NA  36406.198     0
#> 152   11    8      <NA>          <NA>    NA  NA  NA      NA  29192.810     0
#> 153   11    9      <NA>          <NA>    NA  NA  NA      NA  20084.472     0
#> 154   11   10      <NA>          <NA>    NA  NA  NA      NA  26331.980     0
#> 155   11   11      <NA>          <NA>    NA  NA  NA      NA 295612.791   501
#> 156   11   12      <NA>          <NA>    NA  NA  NA      NA  85969.179     0
#> 157   12    0   overall            <=  72.5 158 168     169         NA   501
#> 158   12    1   overall            <=  60.5 159 160     167         NA     0
#> 159   12    2      <NA>          <NA>    NA  NA  NA      NA  15196.609     0
#> 160   12    3       age            <=  28.5 161 165     166         NA     0
#> 161   12    4 defending            <=  53.5 162 163     164         NA     0
#> 162   12    5      <NA>          <NA>    NA  NA  NA      NA  37563.108     0
#> 163   12    6      <NA>          <NA>    NA  NA  NA      NA  33105.179     0
#> 164   12    7      <NA>          <NA>    NA  NA  NA      NA  27576.025     0
#> 165   12    8      <NA>          <NA>    NA  NA  NA      NA  18608.423     0
#> 166   12    9      <NA>          <NA>    NA  NA  NA      NA  29953.026     0
#> 167   12   10      <NA>          <NA>    NA  NA  NA      NA  26287.130     0
#> 168   12   11      <NA>          <NA>    NA  NA  NA      NA 369551.512   501
#> 169   12   12      <NA>          <NA>    NA  NA  NA      NA  88937.856     0
#> 170   13    0 potential            <=  75.5 171 181     182         NA   501
#> 171   13    1   overall            <=  60.5 172 173     180         NA     0
#> 172   13    2      <NA>          <NA>    NA  NA  NA      NA  17176.948     0
#> 173   13    3       age            <=  30.5 174 178     179         NA     0
#> 174   13    4 potential            <=  65.5 175 176     177         NA     0
#> 175   13    5      <NA>          <NA>    NA  NA  NA      NA  28036.763     0
#> 176   13    6      <NA>          <NA>    NA  NA  NA      NA  35266.433     0
#> 177   13    7      <NA>          <NA>    NA  NA  NA      NA  34338.901     0
#> 178   13    8      <NA>          <NA>    NA  NA  NA      NA  10529.258     0
#> 179   13    9      <NA>          <NA>    NA  NA  NA      NA  28968.653     0
#> 180   13   10      <NA>          <NA>    NA  NA  NA      NA  26018.577     0
#> 181   13   11      <NA>          <NA>    NA  NA  NA      NA 119451.005   501
#> 182   13   12      <NA>          <NA>    NA  NA  NA      NA  49328.123     0
#> 183   14    0   overall            <=  72.5 184 194     195         NA   501
#> 184   14    1   overall            <=  59.5 185 186     193         NA     0
#> 185   14    2      <NA>          <NA>    NA  NA  NA      NA  16959.253     0
#> 186   14    3 potential            <=  65.5 187 188     192         NA     0
#> 187   14    4      <NA>          <NA>    NA  NA  NA      NA  26715.248     0
#> 188   14    5 defending            <=  65.5 189 190     191         NA     0
#> 189   14    6      <NA>          <NA>    NA  NA  NA      NA  35196.200     0
#> 190   14    7      <NA>          <NA>    NA  NA  NA      NA  18676.980     0
#> 191   14    8      <NA>          <NA>    NA  NA  NA      NA  23379.199     0
#> 192   14    9      <NA>          <NA>    NA  NA  NA      NA  30248.069     0
#> 193   14   10      <NA>          <NA>    NA  NA  NA      NA  27660.835     0
#> 194   14   11      <NA>          <NA>    NA  NA  NA      NA 336069.717   501
#> 195   14   12      <NA>          <NA>    NA  NA  NA      NA  83005.026     0
#> 196   15    0   overall            <=  73.5 197 207     208         NA   501
#> 197   15    1   overall            <=  59.5 198 199     206         NA     0
#> 198   15    2      <NA>          <NA>    NA  NA  NA      NA  18453.828     0
#> 199   15    3       age            <=  30.5 200 204     205         NA     0
#> 200   15    4 potential            <=  65.5 201 202     203         NA     0
#> 201   15    5      <NA>          <NA>    NA  NA  NA      NA  28987.810     0
#> 202   15    6      <NA>          <NA>    NA  NA  NA      NA  34240.635     0
#> 203   15    7      <NA>          <NA>    NA  NA  NA      NA  33648.516     0
#> 204   15    8      <NA>          <NA>    NA  NA  NA      NA  13477.134     0
#> 205   15    9      <NA>          <NA>    NA  NA  NA      NA  30422.373     0
#> 206   15   10      <NA>          <NA>    NA  NA  NA      NA  28141.262     0
#> 207   15   11      <NA>          <NA>    NA  NA  NA      NA 447544.288   501
#> 208   15   12      <NA>          <NA>    NA  NA  NA      NA  89498.286     0
#> 209   16    0   overall            <=  73.5 210 220     221         NA   501
#> 210   16    1   overall            <=  59.5 211 212     219         NA     0
#> 211   16    2      <NA>          <NA>    NA  NA  NA      NA  20417.945     0
#> 212   16    3       age            <=  29.5 213 217     218         NA     0
#> 213   16    4  shooting            <=  52.5 214 215     216         NA     0
#> 214   16    5      <NA>          <NA>    NA  NA  NA      NA  32015.570     0
#> 215   16    6      <NA>          <NA>    NA  NA  NA      NA  36240.154     0
#> 216   16    7      <NA>          <NA>    NA  NA  NA      NA  27420.239     0
#> 217   16    8      <NA>          <NA>    NA  NA  NA      NA  20643.018     0
#> 218   16    9      <NA>          <NA>    NA  NA  NA      NA  30968.001     0
#> 219   16   10      <NA>          <NA>    NA  NA  NA      NA  28959.489     0
#> 220   16   11      <NA>          <NA>    NA  NA  NA      NA 406289.860   501
#> 221   16   12      <NA>          <NA>    NA  NA  NA      NA  81849.325     0
#> 222   17    0   overall            <=  73.5 223 233     234         NA   501
#> 223   17    1   overall            <=  59.5 224 225     232         NA     0
#> 224   17    2      <NA>          <NA>    NA  NA  NA      NA  21876.150     0
#> 225   17    3       age            <=  31.5 226 230     231         NA     0
#> 226   17    4 potential            <=  64.5 227 228     229         NA     0
#> 227   17    5      <NA>          <NA>    NA  NA  NA      NA  28190.894     0
#> 228   17    6      <NA>          <NA>    NA  NA  NA      NA  34587.379     0
#> 229   17    7      <NA>          <NA>    NA  NA  NA      NA  34099.599     0
#> 230   17    8      <NA>          <NA>    NA  NA  NA      NA  11322.158     0
#> 231   17    9      <NA>          <NA>    NA  NA  NA      NA  31482.756     0
#> 232   17   10      <NA>          <NA>    NA  NA  NA      NA  29620.503     0
#> 233   17   11      <NA>          <NA>    NA  NA  NA      NA 369160.874   501
#> 234   17   12      <NA>          <NA>    NA  NA  NA      NA  80334.214     0
#> 235   18    0 potential            <=  77.5 236 243     247         NA   501
#> 236   18    1   overall            <=  59.5 237 238     242         NA     0
#> 237   18    2      <NA>          <NA>    NA  NA  NA      NA  23188.535     0
#> 238   18    3       age            <=  31.5 239 240     241         NA     0
#> 239   18    4      <NA>          <NA>    NA  NA  NA      NA  33468.528     0
#> 240   18    5      <NA>          <NA>    NA  NA  NA      NA  10596.296     0
#> 241   18    6      <NA>          <NA>    NA  NA  NA      NA  30192.596     0
#> 242   18    7      <NA>          <NA>    NA  NA  NA      NA  28872.830     0
#> 243   18    8   overall            <=  63.5 244 245     246         NA   501
#> 244   18    9      <NA>          <NA>    NA  NA  NA      NA  26667.932     0
#> 245   18   10      <NA>          <NA>    NA  NA  NA      NA 285744.786   501
#> 246   18   11      <NA>          <NA>    NA  NA  NA      NA 258722.429     0
#> 247   18   12      <NA>          <NA>    NA  NA  NA      NA  67453.559     0
#> 248   19    0   overall            <=  73.5 249 259     260         NA   501
#> 249   19    1   overall            <=  58.5 250 251     258         NA     0
#> 250   19    2      <NA>          <NA>    NA  NA  NA      NA  22369.682     0
#> 251   19    3 potential            <=  64.5 252 253     257         NA     0
#> 252   19    4      <NA>          <NA>    NA  NA  NA      NA  28369.682     0
#> 253   19    5 defending            <=  63.5 254 255     256         NA     0
#> 254   19    6      <NA>          <NA>    NA  NA  NA      NA  35039.159     0
#> 255   19    7      <NA>          <NA>    NA  NA  NA      NA  24693.031     0
#> 256   19    8      <NA>          <NA>    NA  NA  NA      NA  26201.279     0
#> 257   19    9      <NA>          <NA>    NA  NA  NA      NA  31014.152     0
#> 258   19   10      <NA>          <NA>    NA  NA  NA      NA  29632.762     0
#> 259   19   11      <NA>          <NA>    NA  NA  NA      NA 310670.308   501
#> 260   19   12      <NA>          <NA>    NA  NA  NA      NA  70316.937     0
# }