Set reference dataset

Change a dataset used as reference for calculating SHAP values. Reference dataset is initially set with data argument in unifying function. Usually reference dataset is dataset used to train the model. Important property of reference dataset is that SHAP values for each observation add up to its deviation from mean prediction for a reference dataset.

set_reference_dataset(unified_model, x)

Arguments

unified_model

Unified model representation of the model created with a (model).unify function. (model_unified.object).

x

Reference dataset. A data.frame or matrix with the same columns as in the training set of the model.

Value

model_unified.object. Unified representation of the model as created with a (model).unify function, but with changed reference dataset (Cover column containing updated values).

See also

Examples

# \donttest{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
formula = value_eur ~ .,
  data = data,
  distribution = "laplace",
  n.trees = 20,
  interaction.depth = 4,
  n.cores = 1)
unified <- gbm.unify(gbm_model, data)
set_reference_dataset(unified, data[200:700, ])
#>     Tree Node   Feature Decision.type Split Yes  No Missing  Prediction Cover
#> 1      0    0   overall            <=  65.5   2   9      13          NA   501
#> 2      0    1 potential            <=  73.5   3   4       8          NA     0
#> 3      0    2      <NA>          <NA>    NA  NA  NA      NA  -7000.0000     0
#> 4      0    3   overall            <=  62.5   5   6       7          NA     0
#> 5      0    4      <NA>          <NA>    NA  NA  NA      NA   7500.0000     0
#> 6      0    5      <NA>          <NA>    NA  NA  NA      NA  45000.0000     0
#> 7      0    6      <NA>          <NA>    NA  NA  NA      NA  24729.7297     0
#> 8      0    7      <NA>          <NA>    NA  NA  NA      NA  -1320.2709     0
#> 9      0    8 potential            <=  68.5  10  11      12          NA   501
#> 10     0    9      <NA>          <NA>    NA  NA  NA      NA  30000.0000     0
#> 11     0   10      <NA>          <NA>    NA  NA  NA      NA 215000.0000   501
#> 12     0   11      <NA>          <NA>    NA  NA  NA      NA 186723.2767     0
#> 13     0   12      <NA>          <NA>    NA  NA  NA      NA 101662.3263     0
#> 14     1    0   overall            <=  66.5  15  22      26          NA   501
#> 15     1    1 potential            <=  74.5  16  17      21          NA     0
#> 16     1    2      <NA>          <NA>    NA  NA  NA      NA   1700.0000     0
#> 17     1    3   overall            <=  62.5  18  19      20          NA     0
#> 18     1    4      <NA>          <NA>    NA  NA  NA      NA  12750.0000     0
#> 19     1    5      <NA>          <NA>    NA  NA  NA      NA  47000.0000     0
#> 20     1    6      <NA>          <NA>    NA  NA  NA      NA  32058.0737     0
#> 21     1    7      <NA>          <NA>    NA  NA  NA      NA   6181.9741     0
#> 22     1    8 potential            <=  69.5  23  24      25          NA   501
#> 23     1    9      <NA>          <NA>    NA  NA  NA      NA  33000.0000     0
#> 24     1   10      <NA>          <NA>    NA  NA  NA      NA 247000.0000   501
#> 25     1   11      <NA>          <NA>    NA  NA  NA      NA 211341.5194     0
#> 26     1   12      <NA>          <NA>    NA  NA  NA      NA 103991.3776     0
#> 27     2    0   overall            <=  67.5  28  35      39          NA   501
#> 28     2    1   overall            <=  63.5  29  30      34          NA     0
#> 29     2    2      <NA>          <NA>    NA  NA  NA      NA  -3025.0000     0
#> 30     2    3       age            <=  24.5  31  32      33          NA     0
#> 31     2    4      <NA>          <NA>    NA  NA  NA      NA  40300.0000     0
#> 32     2    5      <NA>          <NA>    NA  NA  NA      NA  22530.0000     0
#> 33     2    6      <NA>          <NA>    NA  NA  NA      NA  30278.6818     0
#> 34     2    7      <NA>          <NA>    NA  NA  NA      NA  11433.7332     0
#> 35     2    8 potential            <=  70.5  36  37      38          NA   501
#> 36     2    9      <NA>          <NA>    NA  NA  NA      NA  38200.0000     0
#> 37     2   10      <NA>          <NA>    NA  NA  NA      NA 305800.0000   501
#> 38     2   11      <NA>          <NA>    NA  NA  NA      NA 251753.3920     0
#> 39     2   12      <NA>          <NA>    NA  NA  NA      NA 112989.1027     0
#> 40     3    0   overall            <=  69.5  41  51      52          NA   501
#> 41     3    1   overall            <=  64.5  42  46      50          NA     0
#> 42     3    2 potential            <=  75.5  43  44      45          NA     0
#> 43     3    3      <NA>          <NA>    NA  NA  NA      NA   2332.5000     0
#> 44     3    4      <NA>          <NA>    NA  NA  NA      NA  34102.5000     0
#> 45     3    5      <NA>          <NA>    NA  NA  NA      NA   5702.0455     0
#> 46     3    6 defending            <=  59.5  47  48      49          NA     0
#> 47     3    7      <NA>          <NA>    NA  NA  NA      NA  41880.0000     0
#> 48     3    8      <NA>          <NA>    NA  NA  NA      NA  28720.0000     0
#> 49     3    9      <NA>          <NA>    NA  NA  NA      NA  11577.0000     0
#> 50     3   10      <NA>          <NA>    NA  NA  NA      NA  17273.0147     0
#> 51     3   11      <NA>          <NA>    NA  NA  NA      NA 308720.0000   501
#> 52     3   12      <NA>          <NA>    NA  NA  NA      NA 107044.6816     0
#> 53     4    0   overall            <=  69.5  54  61      65          NA   501
#> 54     4    1   overall            <=  63.5  55  56      60          NA     0
#> 55     4    2      <NA>          <NA>    NA  NA  NA      NA   3599.2500     0
#> 56     4    3       age            <=  23.5  57  58      59          NA     0
#> 57     4    4      <NA>          <NA>    NA  NA  NA      NA  42628.0000     0
#> 58     4    5      <NA>          <NA>    NA  NA  NA      NA  25889.0000     0
#> 59     4    6      <NA>          <NA>    NA  NA  NA      NA  31164.9763     0
#> 60     4    7      <NA>          <NA>    NA  NA  NA      NA  17919.3366     0
#> 61     4    8       age            <=  32.5  62  63      64          NA   501
#> 62     4    9      <NA>          <NA>    NA  NA  NA      NA 331348.0000   435
#> 63     4   10      <NA>          <NA>    NA  NA  NA      NA  21348.0000    66
#> 64     4   11      <NA>          <NA>    NA  NA  NA      NA 295599.4762     0
#> 65     4   12      <NA>          <NA>    NA  NA  NA      NA 105395.1131     0
#> 66     5    0 potential            <=  72.5  67  74      78          NA   501
#> 67     5    1   overall            <=  64.5  68  69      73          NA     0
#> 68     5    2      <NA>          <NA>    NA  NA  NA      NA   5739.3250     0
#> 69     5    3 defending            <=  60.5  70  71      72          NA     0
#> 70     5    4      <NA>          <NA>    NA  NA  NA      NA  40161.6000     0
#> 71     5    5      <NA>          <NA>    NA  NA  NA      NA  26616.1000     0
#> 72     5    6      <NA>          <NA>    NA  NA  NA      NA   9081.9000     0
#> 73     5    7      <NA>          <NA>    NA  NA  NA      NA  17363.5914     0
#> 74     5    8   overall            <=  61.5  75  76      77          NA   501
#> 75     5    9      <NA>          <NA>    NA  NA  NA      NA  14507.3250     0
#> 76     5   10      <NA>          <NA>    NA  NA  NA      NA 161713.2000   501
#> 77     5   11      <NA>          <NA>    NA  NA  NA      NA 146313.7975     0
#> 78     5   12      <NA>          <NA>    NA  NA  NA      NA  70910.5913     0
#> 79     6    0   overall            <=  70.5  80  90      91          NA   501
#> 80     6    1   overall            <=  63.5  81  82      89          NA     0
#> 81     6    2      <NA>          <NA>    NA  NA  NA      NA   9665.3925     0
#> 82     6    3 potential            <=  77.5  83  87      88          NA     0
#> 83     6    4  shooting            <=  52.5  84  85      86          NA     0
#> 84     6    5      <NA>          <NA>    NA  NA  NA      NA  25880.9175     0
#> 85     6    6      <NA>          <NA>    NA  NA  NA      NA  32757.4900     0
#> 86     6    7      <NA>          <NA>    NA  NA  NA      NA  12792.2100     0
#> 87     6    8      <NA>          <NA>    NA  NA  NA      NA  58147.8800     0
#> 88     6    9      <NA>          <NA>    NA  NA  NA      NA  31086.1226     0
#> 89     6   10      <NA>          <NA>    NA  NA  NA      NA  21405.9051     0
#> 90     6   11      <NA>          <NA>    NA  NA  NA      NA 329041.8800   501
#> 91     6   12      <NA>          <NA>    NA  NA  NA      NA 101554.8554     0
#> 92     7    0   overall            <=  70.5  93 100     104          NA   501
#> 93     7    1   overall            <=  62.5  94  95      99          NA     0
#> 94     7    2      <NA>          <NA>    NA  NA  NA      NA   8198.8532     0
#> 95     7    3 defending            <=  56.5  96  97      98          NA     0
#> 96     7    4      <NA>          <NA>    NA  NA  NA      NA  35872.1310     0
#> 97     7    5      <NA>          <NA>    NA  NA  NA      NA  28089.3983     0
#> 98     7    6      <NA>          <NA>    NA  NA  NA      NA  18101.6965     0
#> 99     7    7      <NA>          <NA>    NA  NA  NA      NA  21772.9264     0
#> 100    7    8 potential            <=  71.5 101 102     103          NA   501
#> 101    7    9      <NA>          <NA>    NA  NA  NA      NA   3147.4020     0
#> 102    7   10      <NA>          <NA>    NA  NA  NA      NA 349637.6920   501
#> 103    7   11      <NA>          <NA>    NA  NA  NA      NA 324103.7032     0
#> 104    7   12      <NA>          <NA>    NA  NA  NA      NA  98984.8788     0
#> 105    8    0   overall            <=  71.5 106 116     117          NA   501
#> 106    8    1 dribbling            <=  63.5 107 111     115          NA     0
#> 107    8    2   overall            <=  62.5 108 109     110          NA     0
#> 108    8    3      <NA>          <NA>    NA  NA  NA      NA   9878.9679     0
#> 109    8    4      <NA>          <NA>    NA  NA  NA      NA  28620.9779     0
#> 110    8    5      <NA>          <NA>    NA  NA  NA      NA  18723.1621     0
#> 111    8    6 potential            <=  77.5 112 113     114          NA     0
#> 112    8    7      <NA>          <NA>    NA  NA  NA      NA  31502.5752     0
#> 113    8    8      <NA>          <NA>    NA  NA  NA      NA  58745.8789     0
#> 114    8    9      <NA>          <NA>    NA  NA  NA      NA  35215.2530     0
#> 115    8   10      <NA>          <NA>    NA  NA  NA      NA   8878.9679     0
#> 116    8   11      <NA>          <NA>    NA  NA  NA      NA 368173.9228   501
#> 117    8   12      <NA>          <NA>    NA  NA  NA      NA 100981.7877     0
#> 118    9    0   overall            <=  71.5 119 129     130          NA   501
#> 119    9    1   overall            <=  60.5 120 121     128          NA     0
#> 120    9    2      <NA>          <NA>    NA  NA  NA      NA   9391.0711     0
#> 121    9    3 potential            <=  66.5 122 123     127          NA     0
#> 122    9    4      <NA>          <NA>    NA  NA  NA      NA  22622.4284     0
#> 123    9    5  shooting            <=  52.5 124 125     126          NA     0
#> 124    9    6      <NA>          <NA>    NA  NA  NA      NA  31344.7881     0
#> 125    9    7      <NA>          <NA>    NA  NA  NA      NA  38022.5437     0
#> 126    9    8      <NA>          <NA>    NA  NA  NA      NA  23491.0711     0
#> 127    9    9      <NA>          <NA>    NA  NA  NA      NA  31583.6482     0
#> 128    9   10      <NA>          <NA>    NA  NA  NA      NA  25637.1476     0
#> 129    9   11      <NA>          <NA>    NA  NA  NA      NA 334856.5305   501
#> 130    9   12      <NA>          <NA>    NA  NA  NA      NA  94051.8092     0
#> 131   10    0   overall            <=  72.5 132 142     143          NA   501
#> 132   10    1   overall            <=  60.5 133 134     141          NA     0
#> 133   10    2      <NA>          <NA>    NA  NA  NA      NA  11951.9640     0
#> 134   10    3 potential            <=  65.5 135 136     140          NA     0
#> 135   10    4      <NA>          <NA>    NA  NA  NA      NA  22133.1178     0
#> 136   10    5       age            <=  30.5 137 138     139          NA     0
#> 137   10    6      <NA>          <NA>    NA  NA  NA      NA  35562.4060     0
#> 138   10    7      <NA>          <NA>    NA  NA  NA      NA   1575.7221     0
#> 139   10    8      <NA>          <NA>    NA  NA  NA      NA  31254.6288     0
#> 140   10    9      <NA>          <NA>    NA  NA  NA      NA  29959.8380     0
#> 141   10   10      <NA>          <NA>    NA  NA  NA      NA  25569.6674     0
#> 142   10   11      <NA>          <NA>    NA  NA  NA      NA 404870.8775   501
#> 143   10   12      <NA>          <NA>    NA  NA  NA      NA  94009.0695     0
#> 144   11    0 potential            <=  75.5 145 152     156          NA   501
#> 145   11    1  shooting            <=  58.5 146 150     151          NA     0
#> 146   11    2   overall            <=  61.5 147 148     149          NA     0
#> 147   11    3      <NA>          <NA>    NA  NA  NA      NA  16256.7676     0
#> 148   11    4      <NA>          <NA>    NA  NA  NA      NA  29373.7014     0
#> 149   11    5      <NA>          <NA>    NA  NA  NA      NA  24260.9887     0
#> 150   11    6      <NA>          <NA>    NA  NA  NA      NA  35532.5762     0
#> 151   11    7      <NA>          <NA>    NA  NA  NA      NA  12356.7676     0
#> 152   11    8   overall            <=  69.5 153 154     155          NA   501
#> 153   11    9      <NA>          <NA>    NA  NA  NA      NA  35237.3350     0
#> 154   11   10      <NA>          <NA>    NA  NA  NA      NA 467883.7897   501
#> 155   11   11      <NA>          <NA>    NA  NA  NA      NA 317078.8244     0
#> 156   11   12      <NA>          <NA>    NA  NA  NA      NA  99207.6884     0
#> 157   12    0   overall            <=  72.5 158 168     169          NA   501
#> 158   12    1   overall            <=  62.5 159 160     167          NA     0
#> 159   12    2      <NA>          <NA>    NA  NA  NA      NA  19621.0909     0
#> 160   12    3       age            <=  30.5 161 165     166          NA     0
#> 161   12    4 potential            <=  77.5 162 163     164          NA     0
#> 162   12    5      <NA>          <NA>    NA  NA  NA      NA  32576.6611     0
#> 163   12    6      <NA>          <NA>    NA  NA  NA      NA  51526.2579     0
#> 164   12    7      <NA>          <NA>    NA  NA  NA      NA  34928.3450     0
#> 165   12    8      <NA>          <NA>    NA  NA  NA      NA  10792.6783     0
#> 166   12    9      <NA>          <NA>    NA  NA  NA      NA  30544.4899     0
#> 167   12   10      <NA>          <NA>    NA  NA  NA      NA  26790.4798     0
#> 168   12   11      <NA>          <NA>    NA  NA  NA      NA 367830.5321   501
#> 169   12   12      <NA>          <NA>    NA  NA  NA      NA  87355.9689     0
#> 170   13    0 potential            <=  76.5 171 181     182          NA   501
#> 171   13    1   overall            <=  59.5 172 173     180          NA     0
#> 172   13    2      <NA>          <NA>    NA  NA  NA      NA  15668.9818     0
#> 173   13    3 potential            <=  66.5 174 175     179          NA     0
#> 174   13    4      <NA>          <NA>    NA  NA  NA      NA  26869.3793     0
#> 175   13    5       age            <=  31.5 176 177     178          NA     0
#> 176   13    6      <NA>          <NA>    NA  NA  NA      NA  35353.9003     0
#> 177   13    7      <NA>          <NA>    NA  NA  NA      NA -11307.0982     0
#> 178   13    8      <NA>          <NA>    NA  NA  NA      NA  29494.8493     0
#> 179   13    9      <NA>          <NA>    NA  NA  NA      NA  28942.1188     0
#> 180   13   10      <NA>          <NA>    NA  NA  NA      NA  26323.4236     0
#> 181   13   11      <NA>          <NA>    NA  NA  NA      NA 141312.3575   501
#> 182   13   12      <NA>          <NA>    NA  NA  NA      NA  49285.9801     0
#> 183   14    0   overall            <=  73.5 184 194     195          NA   501
#> 184   14    1   overall            <=  59.5 185 186     193          NA     0
#> 185   14    2      <NA>          <NA>    NA  NA  NA      NA  17602.0836     0
#> 186   14    3 potential            <=  65.5 187 188     192          NA     0
#> 187   14    4      <NA>          <NA>    NA  NA  NA      NA  26170.2868     0
#> 188   14    5 defending            <=  54.5 189 190     191          NA     0
#> 189   14    6      <NA>          <NA>    NA  NA  NA      NA  37959.8517     0
#> 190   14    7      <NA>          <NA>    NA  NA  NA      NA  31177.2150     0
#> 191   14    8      <NA>          <NA>    NA  NA  NA      NA  25262.2394     0
#> 192   14    9      <NA>          <NA>    NA  NA  NA      NA  32157.1660     0
#> 193   14   10      <NA>          <NA>    NA  NA  NA      NA  29463.1614     0
#> 194   14   11      <NA>          <NA>    NA  NA  NA      NA 391276.9675   501
#> 195   14   12      <NA>          <NA>    NA  NA  NA      NA  82197.1574     0
#> 196   15    0   overall            <=  73.5 197 207     208          NA   501
#> 197   15    1   overall            <=  59.5 198 199     206          NA     0
#> 198   15    2      <NA>          <NA>    NA  NA  NA      NA  18831.8752     0
#> 199   15    3 potential            <=  65.5 200 201     205          NA     0
#> 200   15    4      <NA>          <NA>    NA  NA  NA      NA  27365.0152     0
#> 201   15    5       age            <=  31.5 202 203     204          NA     0
#> 202   15    6      <NA>          <NA>    NA  NA  NA      NA  34443.3255     0
#> 203   15    7      <NA>          <NA>    NA  NA  NA      NA   2234.1845     0
#> 204   15    8      <NA>          <NA>    NA  NA  NA      NA  31397.7113     0
#> 205   15    9      <NA>          <NA>    NA  NA  NA      NA  30825.1445     0
#> 206   15   10      <NA>          <NA>    NA  NA  NA      NA  28551.1913     0
#> 207   15   11      <NA>          <NA>    NA  NA  NA      NA 395053.4250   501
#> 208   15   12      <NA>          <NA>    NA  NA  NA      NA  80324.2938     0
#> 209   16    0   overall            <=  73.5 210 220     221          NA   501
#> 210   16    1   overall            <=  60.5 211 212     219          NA     0
#> 211   16    2      <NA>          <NA>    NA  NA  NA      NA  21958.6877     0
#> 212   16    3       age            <=  31.5 213 217     218          NA     0
#> 213   16    4  shooting            <=  62.5 214 215     216          NA     0
#> 214   16    5      <NA>          <NA>    NA  NA  NA      NA  32703.8687     0
#> 215   16    6      <NA>          <NA>    NA  NA  NA      NA  40725.6374     0
#> 216   16    7      <NA>          <NA>    NA  NA  NA      NA  26298.7622     0
#> 217   16    8      <NA>          <NA>    NA  NA  NA      NA   9546.2415     0
#> 218   16    9      <NA>          <NA>    NA  NA  NA      NA  31016.6281     0
#> 219   16   10      <NA>          <NA>    NA  NA  NA      NA  28874.7581     0
#> 220   16   11      <NA>          <NA>    NA  NA  NA      NA 359048.0825   501
#> 221   16   12      <NA>          <NA>    NA  NA  NA      NA  75479.8121     0
#> 222   17    0   overall            <=  74.5 223 233     234          NA   501
#> 223   17    1   overall            <=  59.5 224 225     232          NA     0
#> 224   17    2      <NA>          <NA>    NA  NA  NA      NA  21936.0189     0
#> 225   17    3 potential            <=  65.5 226 227     231          NA     0
#> 226   17    4      <NA>          <NA>    NA  NA  NA      NA  28016.8757     0
#> 227   17    5       age            <=  32.5 228 229     230          NA     0
#> 228   17    6      <NA>          <NA>    NA  NA  NA      NA  34471.1148     0
#> 229   17    7      <NA>          <NA>    NA  NA  NA      NA   -734.8626     0
#> 230   17    8      <NA>          <NA>    NA  NA  NA      NA  32033.3020     0
#> 231   17    9      <NA>          <NA>    NA  NA  NA      NA  31466.9762     0
#> 232   17   10      <NA>          <NA>    NA  NA  NA      NA  29774.7402     0
#> 233   17   11      <NA>          <NA>    NA  NA  NA      NA 426643.2743   501
#> 234   17   12      <NA>          <NA>    NA  NA  NA      NA  76891.7507     0
#> 235   18    0   overall            <=  74.5 236 246     247          NA   501
#> 236   18    1   overall            <=  58.5 237 238     245          NA     0
#> 237   18    2      <NA>          <NA>    NA  NA  NA      NA  22406.8563     0
#> 238   18    3 potential            <=  65.5 239 240     244          NA     0
#> 239   18    4      <NA>          <NA>    NA  NA  NA      NA  28241.1340     0
#> 240   18    5       age            <=  32.5 241 242     243          NA     0
#> 241   18    6      <NA>          <NA>    NA  NA  NA      NA  34700.1875     0
#> 242   18    7      <NA>          <NA>    NA  NA  NA      NA   3681.8702     0
#> 243   18    8      <NA>          <NA>    NA  NA  NA      NA  32575.8613     0
#> 244   18    9      <NA>          <NA>    NA  NA  NA      NA  31964.2785     0
#> 245   18   10      <NA>          <NA>    NA  NA  NA      NA  30519.6206     0
#> 246   18   11      <NA>          <NA>    NA  NA  NA      NA 387478.9468   501
#> 247   18   12      <NA>          <NA>    NA  NA  NA      NA  71453.3523     0
#> 248   19    0 potential            <=  78.5 249 256     260          NA   501
#> 249   19    1   overall            <=  58.5 250 251     255          NA     0
#> 250   19    2      <NA>          <NA>    NA  NA  NA      NA  23318.5314     0
#> 251   19    3 potential            <=  64.5 252 253     254          NA     0
#> 252   19    4      <NA>          <NA>    NA  NA  NA      NA  28014.5413     0
#> 253   19    5      <NA>          <NA>    NA  NA  NA      NA  33532.0168     0
#> 254   19    6      <NA>          <NA>    NA  NA  NA      NA  32959.4641     0
#> 255   19    7      <NA>          <NA>    NA  NA  NA      NA  31507.8533     0
#> 256   19    8   overall            <=  64.5 257 258     259          NA   501
#> 257   19    9      <NA>          <NA>    NA  NA  NA      NA  24355.8362     0
#> 258   19   10      <NA>          <NA>    NA  NA  NA      NA 302231.0522   501
#> 259   19   11      <NA>          <NA>    NA  NA  NA      NA 273041.5134     0
#> 260   19   12      <NA>          <NA>    NA  NA  NA      NA  63460.3858     0
# }